Raw coefficients for bootstrap computations of PLSGLR models

A function passed to boot to perform bootstrap.

Usage

coefs.plsRglm.raw(
  dataset,
  ind,
  nt,
  modele,
  family = NULL,
  maxcoefvalues,
  ifbootfail,
  verbose
)

Arguments

dataset

dataset to resample

ind

indices for resampling

nt

number of components to use

modele

type of modele to use, see plsRglm

family

glm family to use, see plsRglm

maxcoefvalues

maximum values allowed for the estimates of the coefficients to discard those coming from singular bootstrap samples

ifbootfail

value to return if the estimation fails on a bootstrap sample

verbose

should info messages be displayed ?

Value

estimates on a bootstrap sample or ifbootfail value if the bootstrap computation fails.

See also

Author

Frédéric Bertrand
frederic.bertrand@lecnam.net
https://fbertran.github.io/homepage/

Examples

data(Cornell)

# (Y,X) bootstrap of a PLSGLR model
set.seed(250)
modplsglm <- coefs.plsRglm.raw(Cornell[,-8],1:nrow(Cornell),nt=3,
modele="pls-glm-family",family=gaussian,maxcoefvalues=1e5,
ifbootfail=rep(0,3),verbose=FALSE)