Extracts and computes information criteria and fits statistics for k-fold cross validated partial least squares models

This function extracts and computes information criteria and fits statistics for k-fold cross validated partial least squares models for both formula or classic specifications of the model.

Usage

kfolds2CVinfos_lm(pls_kfolds, MClassed = FALSE, verbose = TRUE)

Arguments

pls_kfolds

an object computed using PLS_lm_kfoldcv

MClassed

should number of miss classed be computed

verbose

should infos be displayed ?

Value

list

table of fit statistics for first group partition

list()

...

list

table of fit statistics for last group partition

Details

The Mclassed option should only set to TRUE if the response is binary.

Note

Use summary and cv.plsR instead.

References

Nicolas Meyer, Myriam Maumy-Bertrand et Frédéric Bertrand (2010). Comparing the linear and the logistic PLS regression with qualitative predictors: application to allelotyping data. Journal de la Societe Francaise de Statistique, 151(2), pages 1-18. https://www.numdam.org/item/JSFS_2010__151_2_1_0/

See also

Author

Frédéric Bertrand
frederic.bertrand@lecnam.net
https://fbertran.github.io/homepage/

Examples

# \donttest{
data(Cornell)
summary(cv.plsR(Y~.,data=Cornell,nt=10,K=6,verbose=FALSE))
#> ____************************************************____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ____Predicting X without NA neither in X nor in Y____
#> ****________________________________________________****
#> 
#> 
#> NK: 1
#> [[1]]
#>                AIC     Q2cum_Y LimQ2_Y      Q2_Y  PRESS_Y      RSS_Y      R2_Y
#> Nb_Comp_0 82.01205          NA      NA        NA       NA 467.796667        NA
#> Nb_Comp_1 53.15173   0.8519520  0.0975  0.851952 69.25638  35.742486 0.9235940
#> Nb_Comp_2 41.08283   0.8116966  0.0975 -0.271907 45.46112  11.066606 0.9763431
#> Nb_Comp_3 32.06411   0.5620698  0.0975 -1.325664 25.73720   4.418081 0.9905556
#> Nb_Comp_4 33.76477  -2.5555677  0.0975 -7.119027 35.87052   4.309235 0.9907882
#> Nb_Comp_5 33.34373 -27.5253750  0.0975 -7.022734 34.57185   3.521924 0.9924713
#> Nb_Comp_6 35.25533          NA  0.0975        NA       NA   3.496074 0.9925265
#>              AIC.std  DoF.dof sigmahat.dof    AIC.dof    BIC.dof GMDL.dof
#> Nb_Comp_0  37.010388 1.000000    6.5212706 46.0708838 47.7893514 27.59461
#> Nb_Comp_1   8.150064 2.740749    1.8665281  4.5699686  4.9558156 21.34020
#> Nb_Comp_2  -3.918831 5.085967    1.1825195  2.1075461  2.3949331 27.40202
#> Nb_Comp_3 -12.937550 5.121086    0.7488308  0.8467795  0.9628191 24.40842
#> Nb_Comp_4 -11.236891 5.103312    0.7387162  0.8232505  0.9357846 24.23105
#> Nb_Comp_5 -11.657929 6.006316    0.7096382  0.7976101  0.9198348 28.21184
#> Nb_Comp_6  -9.746328 7.000001    0.7633342  0.9711319  1.1359499 33.18347
#>           DoF.naive sigmahat.naive  AIC.naive  BIC.naive GMDL.naive
#> Nb_Comp_0         1      6.5212706 46.0708838 47.7893514   27.59461
#> Nb_Comp_1         2      1.8905683  4.1699567  4.4588195   18.37545
#> Nb_Comp_2         3      1.1088836  1.5370286  1.6860917   17.71117
#> Nb_Comp_3         4      0.7431421  0.7363469  0.8256118   19.01033
#> Nb_Comp_4         5      0.7846050  0.8721072  0.9964867   24.16510
#> Nb_Comp_5         6      0.7661509  0.8804809  1.0227979   28.64206
#> Nb_Comp_6         7      0.8361907  1.1070902  1.3048716   33.63927
#> attr(,"computed_nt")
#> [1] 6
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"

data(pine)
summary(cv.plsR(x11~.,data=pine,nt=10,NK=3,verbose=FALSE),verbose=FALSE)
#> [[1]]
#>                 AIC       Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0  82.41888            NA      NA         NA       NA 20.800152
#> Nb_Comp_1  63.61896    0.40418742  0.0975  0.4041874 12.39299 11.074659
#> Nb_Comp_2  58.47638    0.33193392  0.0975 -0.1212688 12.41767  8.919303
#> Nb_Comp_3  56.55421    0.03276264  0.0975 -0.4478169 12.91352  7.919786
#> Nb_Comp_4  54.35053   -0.49824723  0.0975 -0.5489964 12.26772  6.972542
#> Nb_Comp_5  55.99834   -1.71552715  0.0975 -0.8124693 12.63752  6.898523
#> Nb_Comp_6  57.69592   -4.14934105  0.0975 -0.8962584 13.08138  6.835594
#> Nb_Comp_7  59.37953   -9.62943751  0.0975 -1.0642326 14.11026  6.770369
#> Nb_Comp_8  61.21213  -21.05361351  0.0975 -1.0747677 14.04694  6.736112
#> Nb_Comp_9  63.18426  -46.06101653  0.0975 -1.1339368 14.37444  6.730426
#> Nb_Comp_10 65.15982 -102.26235038  0.0975 -1.1942227 14.76805  6.725443
#>                 R2_Y  AIC.std   DoF.dof sigmahat.dof   AIC.dof   BIC.dof
#> Nb_Comp_0         NA 96.63448  1.000000    0.8062287 0.6697018 0.6991787
#> Nb_Comp_1  0.4675684 77.83455  3.176360    0.5994089 0.4047616 0.4565153
#> Nb_Comp_2  0.5711905 72.69198  7.133559    0.5761829 0.4138120 0.5212090
#> Nb_Comp_3  0.6192438 70.76981  8.778329    0.5603634 0.4070516 0.5320535
#> Nb_Comp_4  0.6647841 68.56612  8.427874    0.5221703 0.3505594 0.4547689
#> Nb_Comp_5  0.6683426 70.21393  9.308247    0.5285695 0.3666578 0.4845912
#> Nb_Comp_6  0.6713681 71.91152  9.291931    0.5259794 0.3629363 0.4795121
#> Nb_Comp_7  0.6745039 73.59512  9.756305    0.5284535 0.3702885 0.4938445
#> Nb_Comp_8  0.6761508 75.42772 10.363948    0.5338475 0.3831339 0.5170783
#> Nb_Comp_9  0.6764242 77.39986 10.732146    0.5378276 0.3920957 0.5328746
#> Nb_Comp_10 0.6766638 79.37542 11.000000    0.5407500 0.3987417 0.5446065
#>             GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0  -3.605128         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1  -9.875081         2      0.5977015 0.3788984 0.4112998 -11.451340
#> Nb_Comp_2  -6.985517         3      0.5452615 0.3243383 0.3647862 -12.822703
#> Nb_Comp_3  -6.260610         4      0.5225859 0.3061986 0.3557368 -12.756838
#> Nb_Comp_4  -8.152986         5      0.4990184 0.2867496 0.3432131 -12.811575
#> Nb_Comp_5  -7.111583         6      0.5054709 0.3019556 0.3714754 -11.329638
#> Nb_Comp_6  -7.233043         7      0.5127450 0.3186757 0.4021333  -9.918688
#> Nb_Comp_7  -6.742195         8      0.5203986 0.3364668 0.4347156  -8.592770
#> Nb_Comp_8  -6.038372         9      0.5297842 0.3572181 0.4717708  -7.287834
#> Nb_Comp_9  -5.600237        10      0.5409503 0.3813021 0.5140048  -6.008747
#> Nb_Comp_10 -5.288422        11      0.5529032 0.4076026 0.5600977  -4.799453
#> attr(,"computed_nt")
#> [1] 10
#> 
#> [[2]]
#>                 AIC     Q2cum_Y LimQ2_Y        Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0  82.41888          NA      NA          NA       NA 20.800152
#> Nb_Comp_1  63.61896   0.3530191  0.0975  0.35301905 13.45730 11.074659
#> Nb_Comp_2  58.47638   0.3264710  0.0975 -0.04103368 11.52909  8.919303
#> Nb_Comp_3  56.55421   0.1686448  0.0975 -0.23432733 11.00934  7.919786
#> Nb_Comp_4  54.35053  -0.1370542  0.0975 -0.36771171 10.83198  6.972542
#> Nb_Comp_5  55.99834  -0.7283234  0.0975 -0.52000082 10.59827  6.898523
#> Nb_Comp_6  57.69592  -1.9563661  0.0975 -0.71053990 11.80020  6.835594
#> Nb_Comp_7  59.37953  -4.4934253  0.0975 -0.85816814 12.70168  6.770369
#> Nb_Comp_8  61.21213  -9.1146498  0.0975 -0.84122824 12.46579  6.736112
#> Nb_Comp_9  63.18426 -21.3247736  0.0975 -1.20717218 14.86776  6.730426
#> Nb_Comp_10 65.15982 -50.9708963  0.0975 -1.32794730 15.66808  6.725443
#>                 R2_Y  AIC.std   DoF.dof sigmahat.dof   AIC.dof   BIC.dof
#> Nb_Comp_0         NA 96.63448  1.000000    0.8062287 0.6697018 0.6991787
#> Nb_Comp_1  0.4675684 77.83455  3.176360    0.5994089 0.4047616 0.4565153
#> Nb_Comp_2  0.5711905 72.69198  7.133559    0.5761829 0.4138120 0.5212090
#> Nb_Comp_3  0.6192438 70.76981  8.778329    0.5603634 0.4070516 0.5320535
#> Nb_Comp_4  0.6647841 68.56612  8.427874    0.5221703 0.3505594 0.4547689
#> Nb_Comp_5  0.6683426 70.21393  9.308247    0.5285695 0.3666578 0.4845912
#> Nb_Comp_6  0.6713681 71.91152  9.291931    0.5259794 0.3629363 0.4795121
#> Nb_Comp_7  0.6745039 73.59512  9.756305    0.5284535 0.3702885 0.4938445
#> Nb_Comp_8  0.6761508 75.42772 10.363948    0.5338475 0.3831339 0.5170783
#> Nb_Comp_9  0.6764242 77.39986 10.732146    0.5378276 0.3920957 0.5328746
#> Nb_Comp_10 0.6766638 79.37542 11.000000    0.5407500 0.3987417 0.5446065
#>             GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0  -3.605128         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1  -9.875081         2      0.5977015 0.3788984 0.4112998 -11.451340
#> Nb_Comp_2  -6.985517         3      0.5452615 0.3243383 0.3647862 -12.822703
#> Nb_Comp_3  -6.260610         4      0.5225859 0.3061986 0.3557368 -12.756838
#> Nb_Comp_4  -8.152986         5      0.4990184 0.2867496 0.3432131 -12.811575
#> Nb_Comp_5  -7.111583         6      0.5054709 0.3019556 0.3714754 -11.329638
#> Nb_Comp_6  -7.233043         7      0.5127450 0.3186757 0.4021333  -9.918688
#> Nb_Comp_7  -6.742195         8      0.5203986 0.3364668 0.4347156  -8.592770
#> Nb_Comp_8  -6.038372         9      0.5297842 0.3572181 0.4717708  -7.287834
#> Nb_Comp_9  -5.600237        10      0.5409503 0.3813021 0.5140048  -6.008747
#> Nb_Comp_10 -5.288422        11      0.5529032 0.4076026 0.5600977  -4.799453
#> attr(,"computed_nt")
#> [1] 10
#> 
#> [[3]]
#>                 AIC       Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0  82.41888            NA      NA         NA       NA 20.800152
#> Nb_Comp_1  63.61896    0.27325368  0.0975  0.2732537 15.11643 11.074659
#> Nb_Comp_2  58.47638    0.05720826  0.0975 -0.2972776 14.36691  8.919303
#> Nb_Comp_3  56.55421   -0.64711661  0.0975 -0.7470631 15.58258  7.919786
#> Nb_Comp_4  54.35053   -2.12708354  0.0975 -0.8985198 15.03587  6.972542
#> Nb_Comp_5  55.99834   -5.78392162  0.0975 -1.1694085 15.12629  6.898523
#> Nb_Comp_6  57.69592  -13.77626131  0.0975 -1.1781297 15.02588  6.835594
#> Nb_Comp_7  59.37953  -33.32015187  0.0975 -1.3226546 15.87672  6.770369
#> Nb_Comp_8  61.21213  -80.02191131  0.0975 -1.3607679 15.98327  6.736112
#> Nb_Comp_9  63.18426 -193.57847082  0.0975 -1.4015537 16.17713  6.730426
#> Nb_Comp_10 65.15982 -481.05433452  0.0975 -1.4774289 16.67415  6.725443
#>                 R2_Y  AIC.std   DoF.dof sigmahat.dof   AIC.dof   BIC.dof
#> Nb_Comp_0         NA 96.63448  1.000000    0.8062287 0.6697018 0.6991787
#> Nb_Comp_1  0.4675684 77.83455  3.176360    0.5994089 0.4047616 0.4565153
#> Nb_Comp_2  0.5711905 72.69198  7.133559    0.5761829 0.4138120 0.5212090
#> Nb_Comp_3  0.6192438 70.76981  8.778329    0.5603634 0.4070516 0.5320535
#> Nb_Comp_4  0.6647841 68.56612  8.427874    0.5221703 0.3505594 0.4547689
#> Nb_Comp_5  0.6683426 70.21393  9.308247    0.5285695 0.3666578 0.4845912
#> Nb_Comp_6  0.6713681 71.91152  9.291931    0.5259794 0.3629363 0.4795121
#> Nb_Comp_7  0.6745039 73.59512  9.756305    0.5284535 0.3702885 0.4938445
#> Nb_Comp_8  0.6761508 75.42772 10.363948    0.5338475 0.3831339 0.5170783
#> Nb_Comp_9  0.6764242 77.39986 10.732146    0.5378276 0.3920957 0.5328746
#> Nb_Comp_10 0.6766638 79.37542 11.000000    0.5407500 0.3987417 0.5446065
#>             GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0  -3.605128         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1  -9.875081         2      0.5977015 0.3788984 0.4112998 -11.451340
#> Nb_Comp_2  -6.985517         3      0.5452615 0.3243383 0.3647862 -12.822703
#> Nb_Comp_3  -6.260610         4      0.5225859 0.3061986 0.3557368 -12.756838
#> Nb_Comp_4  -8.152986         5      0.4990184 0.2867496 0.3432131 -12.811575
#> Nb_Comp_5  -7.111583         6      0.5054709 0.3019556 0.3714754 -11.329638
#> Nb_Comp_6  -7.233043         7      0.5127450 0.3186757 0.4021333  -9.918688
#> Nb_Comp_7  -6.742195         8      0.5203986 0.3364668 0.4347156  -8.592770
#> Nb_Comp_8  -6.038372         9      0.5297842 0.3572181 0.4717708  -7.287834
#> Nb_Comp_9  -5.600237        10      0.5409503 0.3813021 0.5140048  -6.008747
#> Nb_Comp_10 -5.288422        11      0.5529032 0.4076026 0.5600977  -4.799453
#> attr(,"computed_nt")
#> [1] 10
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"
data(pineNAX21)
summary(cv.plsR(x11~.,data=pineNAX21,nt=10,NK=3,
verbose=FALSE),verbose=FALSE)
#> [[1]]
#>                AIC      Q2cum_Y LimQ2_Y        Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0 82.41888           NA      NA          NA       NA 20.800152
#> Nb_Comp_1 63.69250    0.2942436  0.0975   0.2942436 14.67984 11.099368
#> Nb_Comp_2 58.35228    0.0716092  0.0975  -0.3154551 14.60072  8.885823
#> Nb_Comp_3 56.36553   -0.5380093  0.0975  -0.6566399 14.72061  7.874634
#> Nb_Comp_4 54.02416   -1.7052155  0.0975  -0.7589072 13.85075  6.903925
#> Nb_Comp_5 55.80450   -3.8193757  0.0975  -0.7815127 12.29943  6.858120
#> Nb_Comp_6 57.45753   -7.8982861  0.0975  -0.8463566 12.66254  6.786392
#> Nb_Comp_7 58.73951 -102.7525320  0.0975 -10.6598332 79.12819  6.640327
#> Nb_Comp_8 60.61227 -177.7431389  0.0975  -0.7227834 11.43985  6.614773
#> Nb_Comp_9 62.25948 -505.2225209  0.0975  -1.8321228 18.73385  6.544432
#>                R2_Y  AIC.std DoF.dof sigmahat.dof AIC.dof BIC.dof GMDL.dof
#> Nb_Comp_0        NA 96.63448      NA           NA      NA      NA       NA
#> Nb_Comp_1 0.4663804 77.90810      NA           NA      NA      NA       NA
#> Nb_Comp_2 0.5728001 72.56787      NA           NA      NA      NA       NA
#> Nb_Comp_3 0.6214146 70.58113      NA           NA      NA      NA       NA
#> Nb_Comp_4 0.6680830 68.23976      NA           NA      NA      NA       NA
#> Nb_Comp_5 0.6702851 70.02009      NA           NA      NA      NA       NA
#> Nb_Comp_6 0.6737336 71.67313      NA           NA      NA      NA       NA
#> Nb_Comp_7 0.6807558 72.95511      NA           NA      NA      NA       NA
#> Nb_Comp_8 0.6819844 74.82787      NA           NA      NA      NA       NA
#> Nb_Comp_9 0.6853661 76.47507      NA           NA      NA      NA       NA
#>           DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1         2      0.5983679 0.3797438 0.4122175 -11.413749
#> Nb_Comp_2         3      0.5442372 0.3231208 0.3634169 -12.847656
#> Nb_Comp_3         4      0.5210941 0.3044529 0.3537087 -12.776843
#> Nb_Comp_4         5      0.4965569 0.2839276 0.3398355 -12.891035
#> Nb_Comp_5         6      0.5039885 0.3001871 0.3692997 -11.349498
#> Nb_Comp_6         7      0.5108963 0.3163819 0.3992388  -9.922119
#> Nb_Comp_7         8      0.5153766 0.3300041 0.4263658  -8.696873
#> Nb_Comp_8         9      0.5249910 0.3507834 0.4632727  -7.337679
#> Nb_Comp_9        10      0.5334234 0.3707649 0.4998004  -6.033403
#> attr(,"computed_nt")
#> [1] 9
#> 
#> [[2]]
#>                AIC       Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0 82.41888            NA      NA         NA       NA 20.800152
#> Nb_Comp_1 63.69250  2.655287e-01  0.0975  0.2655287 15.27711 11.099368
#> Nb_Comp_2 58.35228  2.373341e-02  0.0975 -0.3292100 14.75339  8.885823
#> Nb_Comp_3 56.36553 -6.672436e-01  0.0975 -0.7077750 15.17499  7.874634
#> Nb_Comp_4 54.02416 -2.354612e+00  0.0975 -1.0120706 15.84432  6.903925
#> Nb_Comp_5 55.80450 -5.192725e+00  0.0975 -0.8460333 12.74487  6.858120
#> Nb_Comp_6 57.45753 -1.329328e+01  0.0975 -1.3080766 15.82907  6.786392
#> Nb_Comp_7 58.73951 -9.615451e+01  0.0975 -5.7972139 46.12856  6.640327
#> Nb_Comp_8 60.61227 -3.484028e+02  0.0975 -2.5963624 23.88102  6.614773
#> Nb_Comp_9 62.25948 -1.209162e+03  0.0975 -2.4635163 22.91037  6.544432
#>                R2_Y  AIC.std DoF.dof sigmahat.dof AIC.dof BIC.dof GMDL.dof
#> Nb_Comp_0        NA 96.63448      NA           NA      NA      NA       NA
#> Nb_Comp_1 0.4663804 77.90810      NA           NA      NA      NA       NA
#> Nb_Comp_2 0.5728001 72.56787      NA           NA      NA      NA       NA
#> Nb_Comp_3 0.6214146 70.58113      NA           NA      NA      NA       NA
#> Nb_Comp_4 0.6680830 68.23976      NA           NA      NA      NA       NA
#> Nb_Comp_5 0.6702851 70.02009      NA           NA      NA      NA       NA
#> Nb_Comp_6 0.6737336 71.67313      NA           NA      NA      NA       NA
#> Nb_Comp_7 0.6807558 72.95511      NA           NA      NA      NA       NA
#> Nb_Comp_8 0.6819844 74.82787      NA           NA      NA      NA       NA
#> Nb_Comp_9 0.6853661 76.47507      NA           NA      NA      NA       NA
#>           DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1         2      0.5983679 0.3797438 0.4122175 -11.413749
#> Nb_Comp_2         3      0.5442372 0.3231208 0.3634169 -12.847656
#> Nb_Comp_3         4      0.5210941 0.3044529 0.3537087 -12.776843
#> Nb_Comp_4         5      0.4965569 0.2839276 0.3398355 -12.891035
#> Nb_Comp_5         6      0.5039885 0.3001871 0.3692997 -11.349498
#> Nb_Comp_6         7      0.5108963 0.3163819 0.3992388  -9.922119
#> Nb_Comp_7         8      0.5153766 0.3300041 0.4263658  -8.696873
#> Nb_Comp_8         9      0.5249910 0.3507834 0.4632727  -7.337679
#> Nb_Comp_9        10      0.5334234 0.3707649 0.4998004  -6.033403
#> attr(,"computed_nt")
#> [1] 9
#> 
#> [[3]]
#>                AIC       Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0 82.41888            NA      NA         NA       NA 20.800152
#> Nb_Comp_1 63.69250  3.331660e-01  0.0975  0.3331660 13.87025 11.099368
#> Nb_Comp_2 58.35228  6.088959e-02  0.0975 -0.4083122 15.63138  8.885823
#> Nb_Comp_3 56.36553 -9.486122e-01  0.0975 -1.0749554 18.43769  7.874634
#> Nb_Comp_4 54.02416 -3.096530e+00  0.0975 -1.1022805 16.55469  6.903925
#> Nb_Comp_5 55.80450 -1.552594e+01  0.0975 -3.0341307 27.85133  6.858120
#> Nb_Comp_6 57.45753 -5.855316e+01  0.0975 -2.6036179 24.71404  6.786392
#> Nb_Comp_7 58.73951 -2.331510e+02  0.0975 -2.9317991 26.68273  6.640327
#> Nb_Comp_8 60.61227 -5.147180e+02  0.0975 -1.2025015 14.62533  6.614773
#> Nb_Comp_9 62.25948 -1.271335e+03  0.0975 -1.4671131 16.31939  6.544432
#>                R2_Y  AIC.std DoF.dof sigmahat.dof AIC.dof BIC.dof GMDL.dof
#> Nb_Comp_0        NA 96.63448      NA           NA      NA      NA       NA
#> Nb_Comp_1 0.4663804 77.90810      NA           NA      NA      NA       NA
#> Nb_Comp_2 0.5728001 72.56787      NA           NA      NA      NA       NA
#> Nb_Comp_3 0.6214146 70.58113      NA           NA      NA      NA       NA
#> Nb_Comp_4 0.6680830 68.23976      NA           NA      NA      NA       NA
#> Nb_Comp_5 0.6702851 70.02009      NA           NA      NA      NA       NA
#> Nb_Comp_6 0.6737336 71.67313      NA           NA      NA      NA       NA
#> Nb_Comp_7 0.6807558 72.95511      NA           NA      NA      NA       NA
#> Nb_Comp_8 0.6819844 74.82787      NA           NA      NA      NA       NA
#> Nb_Comp_9 0.6853661 76.47507      NA           NA      NA      NA       NA
#>           DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1         2      0.5983679 0.3797438 0.4122175 -11.413749
#> Nb_Comp_2         3      0.5442372 0.3231208 0.3634169 -12.847656
#> Nb_Comp_3         4      0.5210941 0.3044529 0.3537087 -12.776843
#> Nb_Comp_4         5      0.4965569 0.2839276 0.3398355 -12.891035
#> Nb_Comp_5         6      0.5039885 0.3001871 0.3692997 -11.349498
#> Nb_Comp_6         7      0.5108963 0.3163819 0.3992388  -9.922119
#> Nb_Comp_7         8      0.5153766 0.3300041 0.4263658  -8.696873
#> Nb_Comp_8         9      0.5249910 0.3507834 0.4632727  -7.337679
#> Nb_Comp_9        10      0.5334234 0.3707649 0.4998004  -6.033403
#> attr(,"computed_nt")
#> [1] 9
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"

data(aze_compl)
summary(cv.plsR(y~.,data=aze_compl,nt=10,K=8,NK=3,
verbose=FALSE),MClassed=TRUE,verbose=FALSE)
#> [[1]]
#>                 AIC MissClassed CV_MissClassed       Q2cum_Y LimQ2_Y       Q2_Y
#> Nb_Comp_0  154.6179          49             NA            NA      NA         NA
#> Nb_Comp_1  126.4083          27             47    -0.1647434  0.0975 -0.1647434
#> Nb_Comp_2  119.3375          25             47    -0.8822439  0.0975 -0.6160160
#> Nb_Comp_3  114.2313          27             47    -2.3852200  0.0975 -0.7985023
#> Nb_Comp_4  112.3463          23             50    -6.3321174  0.0975 -1.1659205
#> Nb_Comp_5  113.2362          22             47   -15.4256691  0.0975 -1.2402354
#> Nb_Comp_6  114.7620          21             49   -36.5043152  0.0975 -1.2832747
#> Nb_Comp_7  116.5264          20             48   -84.7618135  0.0975 -1.2867186
#> Nb_Comp_8  118.4601          20             47  -195.7083439  0.0975 -1.2936589
#> Nb_Comp_9  120.4452          19             47  -448.6419012  0.0975 -1.2858303
#> Nb_Comp_10 122.4395          19             47 -1026.6364377  0.0975 -1.2854552
#>             PRESS_Y    RSS_Y      R2_Y  AIC.std  DoF.dof sigmahat.dof   AIC.dof
#> Nb_Comp_0        NA 25.91346        NA 298.1344  1.00000    0.5015845 0.2540061
#> Nb_Comp_1  30.18253 19.38086 0.2520929 269.9248 22.55372    0.4848429 0.2883114
#> Nb_Comp_2  31.31978 17.76209 0.3145613 262.8540 27.31542    0.4781670 0.2908950
#> Nb_Comp_3  31.94516 16.58896 0.3598323 257.7478 30.52370    0.4719550 0.2902572
#> Nb_Comp_4  35.93037 15.98071 0.3833049 255.8628 34.00000    0.4744263 0.3008285
#> Nb_Comp_5  35.80054 15.81104 0.3898523 256.7527 34.00000    0.4719012 0.2976347
#> Nb_Comp_6  36.10095 15.73910 0.3926285 258.2785 34.00000    0.4708264 0.2962804
#> Nb_Comp_7  35.99089 15.70350 0.3940024 260.0429 33.71066    0.4693382 0.2937976
#> Nb_Comp_8  36.01846 15.69348 0.3943888 261.9766 34.00000    0.4701436 0.2954217
#> Nb_Comp_9  35.87264 15.69123 0.3944758 263.9617 33.87284    0.4696894 0.2945815
#> Nb_Comp_10 35.86160 15.69037 0.3945088 265.9560 34.00000    0.4700970 0.2953632
#>              BIC.dof  GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive
#> Nb_Comp_0  0.2604032 -67.17645         1      0.5015845 0.2540061 0.2604032
#> Nb_Comp_1  0.4231184 -53.56607         2      0.4358996 0.1936625 0.2033251
#> Nb_Comp_2  0.4496983 -52.42272         3      0.4193593 0.1809352 0.1943501
#> Nb_Comp_3  0.4631316 -51.93343         4      0.4072955 0.1722700 0.1891422
#> Nb_Comp_4  0.4954133 -50.37079         5      0.4017727 0.1691819 0.1897041
#> Nb_Comp_5  0.4901536 -50.65724         6      0.4016679 0.1706451 0.1952588
#> Nb_Comp_6  0.4879234 -50.78005         7      0.4028135 0.1731800 0.2020601
#> Nb_Comp_7  0.4826103 -51.05525         8      0.4044479 0.1761610 0.2094352
#> Nb_Comp_8  0.4865092 -50.85833         9      0.4064413 0.1794902 0.2172936
#> Nb_Comp_9  0.4845867 -50.95616        10      0.4085682 0.1829787 0.2254232
#> Nb_Comp_10 0.4864128 -50.86368        11      0.4107477 0.1865584 0.2337468
#>            GMDL.naive
#> Nb_Comp_0   -67.17645
#> Nb_Comp_1   -79.67755
#> Nb_Comp_2   -81.93501
#> Nb_Comp_3   -83.31503
#> Nb_Comp_4   -83.23369
#> Nb_Comp_5   -81.93513
#> Nb_Comp_6   -80.42345
#> Nb_Comp_7   -78.87607
#> Nb_Comp_8   -77.31942
#> Nb_Comp_9   -75.80069
#> Nb_Comp_10  -74.33325
#> attr(,"computed_nt")
#> [1] 10
#> 
#> [[2]]
#>                 AIC MissClassed CV_MissClassed      Q2cum_Y LimQ2_Y       Q2_Y
#> Nb_Comp_0  154.6179          49             NA           NA      NA         NA
#> Nb_Comp_1  126.4083          27             42   -0.1171446  0.0975 -0.1171446
#> Nb_Comp_2  119.3375          25             45   -0.7467203  0.0975 -0.5635579
#> Nb_Comp_3  114.2313          27             43   -2.1919948  0.0975 -0.8274218
#> Nb_Comp_4  112.3463          23             47   -5.4824416  0.0975 -1.0308434
#> Nb_Comp_5  113.2362          22             46  -12.6136137  0.0975 -1.1000750
#> Nb_Comp_6  114.7620          21             48  -27.7014965  0.0975 -1.1082937
#> Nb_Comp_7  116.5264          20             47  -59.6778836  0.0975 -1.1141017
#> Nb_Comp_8  118.4601          20             47 -128.5070391  0.0975 -1.1343368
#> Nb_Comp_9  120.4452          19             47 -276.3780478  0.0975 -1.1417990
#> Nb_Comp_10 122.4395          19             48 -593.6802226  0.0975 -1.1439340
#>             PRESS_Y    RSS_Y      R2_Y  AIC.std  DoF.dof sigmahat.dof   AIC.dof
#> Nb_Comp_0        NA 25.91346        NA 298.1344  1.00000    0.5015845 0.2540061
#> Nb_Comp_1  28.94908 19.38086 0.2520929 269.9248 22.55372    0.4848429 0.2883114
#> Nb_Comp_2  30.30310 17.76209 0.3145613 262.8540 27.31542    0.4781670 0.2908950
#> Nb_Comp_3  32.45883 16.58896 0.3598323 257.7478 30.52370    0.4719550 0.2902572
#> Nb_Comp_4  33.68958 15.98071 0.3833049 255.8628 34.00000    0.4744263 0.3008285
#> Nb_Comp_5  33.56068 15.81104 0.3898523 256.7527 34.00000    0.4719012 0.2976347
#> Nb_Comp_6  33.33432 15.73910 0.3926285 258.2785 34.00000    0.4708264 0.2962804
#> Nb_Comp_7  33.27406 15.70350 0.3940024 260.0429 33.71066    0.4693382 0.2937976
#> Nb_Comp_8  33.51655 15.69348 0.3943888 261.9766 34.00000    0.4701436 0.2954217
#> Nb_Comp_9  33.61228 15.69123 0.3944758 263.9617 33.87284    0.4696894 0.2945815
#> Nb_Comp_10 33.64096 15.69037 0.3945088 265.9560 34.00000    0.4700970 0.2953632
#>              BIC.dof  GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive
#> Nb_Comp_0  0.2604032 -67.17645         1      0.5015845 0.2540061 0.2604032
#> Nb_Comp_1  0.4231184 -53.56607         2      0.4358996 0.1936625 0.2033251
#> Nb_Comp_2  0.4496983 -52.42272         3      0.4193593 0.1809352 0.1943501
#> Nb_Comp_3  0.4631316 -51.93343         4      0.4072955 0.1722700 0.1891422
#> Nb_Comp_4  0.4954133 -50.37079         5      0.4017727 0.1691819 0.1897041
#> Nb_Comp_5  0.4901536 -50.65724         6      0.4016679 0.1706451 0.1952588
#> Nb_Comp_6  0.4879234 -50.78005         7      0.4028135 0.1731800 0.2020601
#> Nb_Comp_7  0.4826103 -51.05525         8      0.4044479 0.1761610 0.2094352
#> Nb_Comp_8  0.4865092 -50.85833         9      0.4064413 0.1794902 0.2172936
#> Nb_Comp_9  0.4845867 -50.95616        10      0.4085682 0.1829787 0.2254232
#> Nb_Comp_10 0.4864128 -50.86368        11      0.4107477 0.1865584 0.2337468
#>            GMDL.naive
#> Nb_Comp_0   -67.17645
#> Nb_Comp_1   -79.67755
#> Nb_Comp_2   -81.93501
#> Nb_Comp_3   -83.31503
#> Nb_Comp_4   -83.23369
#> Nb_Comp_5   -81.93513
#> Nb_Comp_6   -80.42345
#> Nb_Comp_7   -78.87607
#> Nb_Comp_8   -77.31942
#> Nb_Comp_9   -75.80069
#> Nb_Comp_10  -74.33325
#> attr(,"computed_nt")
#> [1] 10
#> 
#> [[3]]
#>                 AIC MissClassed CV_MissClassed       Q2cum_Y LimQ2_Y       Q2_Y
#> Nb_Comp_0  154.6179          49             NA            NA      NA         NA
#> Nb_Comp_1  126.4083          27             46    -0.1622540  0.0975 -0.1622540
#> Nb_Comp_2  119.3375          25             44    -0.8255452  0.0975 -0.5706938
#> Nb_Comp_3  114.2313          27             45    -2.4072694  0.0975 -0.8664393
#> Nb_Comp_4  112.3463          23             51    -6.4582755  0.0975 -1.1889304
#> Nb_Comp_5  113.2362          22             55   -16.1937772  0.0975 -1.3053288
#> Nb_Comp_6  114.7620          21             52   -39.7491244  0.0975 -1.3699926
#> Nb_Comp_7  116.5264          20             51   -97.1711222  0.0975 -1.4091591
#> Nb_Comp_8  118.4601          20             51  -237.2500136  0.0975 -1.4268849
#> Nb_Comp_9  120.4452          19             51  -579.3076122  0.0975 -1.4357086
#> Nb_Comp_10 122.4395          19             51 -1416.9673756  0.0975 -1.4434754
#>             PRESS_Y    RSS_Y      R2_Y  AIC.std  DoF.dof sigmahat.dof   AIC.dof
#> Nb_Comp_0        NA 25.91346        NA 298.1344  1.00000    0.5015845 0.2540061
#> Nb_Comp_1  30.11803 19.38086 0.2520929 269.9248 22.55372    0.4848429 0.2883114
#> Nb_Comp_2  30.44140 17.76209 0.3145613 262.8540 27.31542    0.4781670 0.2908950
#> Nb_Comp_3  33.15186 16.58896 0.3598323 257.7478 30.52370    0.4719550 0.2902572
#> Nb_Comp_4  36.31208 15.98071 0.3833049 255.8628 34.00000    0.4744263 0.3008285
#> Nb_Comp_5  36.84078 15.81104 0.3898523 256.7527 34.00000    0.4719012 0.2976347
#> Nb_Comp_6  37.47205 15.73910 0.3926285 258.2785 34.00000    0.4708264 0.2962804
#> Nb_Comp_7  37.91799 15.70350 0.3940024 260.0429 33.71066    0.4693382 0.2937976
#> Nb_Comp_8  38.11058 15.69348 0.3943888 261.9766 34.00000    0.4701436 0.2954217
#> Nb_Comp_9  38.22475 15.69123 0.3944758 263.9617 33.87284    0.4696894 0.2945815
#> Nb_Comp_10 38.34113 15.69037 0.3945088 265.9560 34.00000    0.4700970 0.2953632
#>              BIC.dof  GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive
#> Nb_Comp_0  0.2604032 -67.17645         1      0.5015845 0.2540061 0.2604032
#> Nb_Comp_1  0.4231184 -53.56607         2      0.4358996 0.1936625 0.2033251
#> Nb_Comp_2  0.4496983 -52.42272         3      0.4193593 0.1809352 0.1943501
#> Nb_Comp_3  0.4631316 -51.93343         4      0.4072955 0.1722700 0.1891422
#> Nb_Comp_4  0.4954133 -50.37079         5      0.4017727 0.1691819 0.1897041
#> Nb_Comp_5  0.4901536 -50.65724         6      0.4016679 0.1706451 0.1952588
#> Nb_Comp_6  0.4879234 -50.78005         7      0.4028135 0.1731800 0.2020601
#> Nb_Comp_7  0.4826103 -51.05525         8      0.4044479 0.1761610 0.2094352
#> Nb_Comp_8  0.4865092 -50.85833         9      0.4064413 0.1794902 0.2172936
#> Nb_Comp_9  0.4845867 -50.95616        10      0.4085682 0.1829787 0.2254232
#> Nb_Comp_10 0.4864128 -50.86368        11      0.4107477 0.1865584 0.2337468
#>            GMDL.naive
#> Nb_Comp_0   -67.17645
#> Nb_Comp_1   -79.67755
#> Nb_Comp_2   -81.93501
#> Nb_Comp_3   -83.31503
#> Nb_Comp_4   -83.23369
#> Nb_Comp_5   -81.93513
#> Nb_Comp_6   -80.42345
#> Nb_Comp_7   -78.87607
#> Nb_Comp_8   -77.31942
#> Nb_Comp_9   -75.80069
#> Nb_Comp_10  -74.33325
#> attr(,"computed_nt")
#> [1] 10
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"
# }