Data generating function for univariate plsR models

This function generates a single univariate response value \(Y\) and a vector of explanatory variables \((X_1,\ldots,X_{totdim})\) drawn from a model with a given number of latent components.

simul_data_UniYX(totdim, ncomp)

Arguments

totdim
Number of columns of the X vector (from ncomp to hardware limits)

ncomp
Number of latent components in the model (from 2 to 6)

Details

This function should be combined with the replicate function to give rise to a larger dataset. The algorithm used is a R port of the one described in the article of Li which is a multivariate generalization of the algorithm of Naes and Martens.

Value

vector

\((Y,X_1,\ldots,X_{totdim})\)

References

T. Naes, H. Martens, Comparison of prediction methods for multicollinear data, Commun. Stat., Simul. 14 (1985) 545-576.
Baibing Li, Julian Morris, Elaine B. Martin, Model selection for partial least squares regression, Chemometrics and Intelligent Laboratory Systems 64 (2002) 79-89.
http://dx.doi.org/10.1016/S0169-7439(02)00051-5

See also

Examples

simul_data_UniYX(20,6)

#>          Y         X1         X2         X3         X4         X5         X6 
#> -4.9621927 -2.3329311 -2.8124086  2.8179302 -3.1531052 -0.4919760 -1.9678874 
#>         X7         X8         X9        X10        X11        X12        X13 
#>  4.6752300 -2.3025227 -1.2038769 -4.1418744 -0.8984016  0.3910133 -1.5313235 
#>        X14        X15        X16        X17        X18        X19        X20 
#>  3.2070599 -1.2004015 -4.1392078 -0.8764455  0.3822851 -1.5373799  3.2230673 

dimX <- 6
Astar <- 2
simul_data_UniYX(dimX,Astar)

#>          Y         X1         X2         X3         X4         X5         X6 
#> -12.088176  -8.960577  -8.979948  -8.361891  -8.977458  -8.976908  -8.376012 

(dataAstar2 <- data.frame(t(replicate(50,simul_data_UniYX(dimX,Astar)))))

#>              Y         X1         X2           X3         X4         X5
#> 1    9.7051381  3.0744448  3.0823647   7.71682447  3.0876750  3.0897737
#> 2   -3.3539600  0.5234967  0.5219664  -3.07347402  0.5491451  0.5231788
#> 3  -10.2237913 -2.7205506 -2.7176771  -8.10366616 -2.7346575 -2.7416298
#> 4   -1.4571866  2.3170977  2.3428754  -1.95924597  2.3067684  2.3346929
#> 5   -3.7208297  1.8898771  1.9021424  -3.83587119  1.9070902  1.9041304
#> 6    8.5235788  1.1845944  1.1945158   7.06844366  1.2168181  1.2055997
#> 7   -5.8399385  4.0561791  4.0499628  -6.09281865  4.0542860  4.0574137
#> 8   10.2178225 -1.5764643 -1.5748160   9.42223908 -1.5775424 -1.5889152
#> 9   -9.3905301 -1.9246655 -1.9104927  -7.76417846 -1.9133505 -1.9309751
#> 10  15.8334110  9.2700784  9.2517779  11.59897804  9.2559991  9.2595424
#> 11  -0.9299327 -3.1372805 -3.1305553  -0.15894433 -3.1382926 -3.1439896
#> 12   5.7335136  4.1408458  4.1195731   4.23713946  4.1382392  4.1346998
#> 13   8.4010984  2.3447257  2.3562921   6.96826162  2.3453956  2.3435881
#> 14   9.3594655 -4.0441230 -4.0731723   9.12109678 -4.0670187 -4.0468874
#> 15   6.0439890 -2.0390072 -2.0202547   5.81924504 -2.0223381 -2.0487484
#> 16  -4.7766122 -6.5909710 -6.6001560  -2.47086189 -6.5864802 -6.6057301
#> 17   8.1333877  4.3975598  4.3980298   6.27263004  4.4069417  4.3996800
#> 18  12.9631260 -1.7955666 -1.7711113  12.01118378 -1.7718280 -1.7685012
#> 19   3.0500987 -0.1377115 -0.1351427   2.60540294 -0.1314509 -0.1295387
#> 20   9.9100545 -6.1463928 -6.1345467  10.12158604 -6.1280719 -6.1395008
#> 21 -11.4632761  1.1670180  1.1507101 -10.34631734  1.1599721  1.1743924
#> 22 -12.7611154 -2.6712727 -2.6765390 -10.32129483 -2.6951728 -2.6924172
#> 23   2.2570450  6.7194472  6.7167541   0.09787755  6.7112326  6.7262618
#> 24 -14.9528520  0.1604965  0.1718592 -13.21602524  0.1469414  0.1618849
#> 25   0.2086731 -1.7862281 -1.7706744   0.67927861 -1.7741933 -1.7701187
#> 26  -9.3111643 -2.5921944 -2.5950622  -7.48682362 -2.5878596 -2.5912413
#> 27  -3.7035994 -3.2980956 -3.3005212  -2.63175103 -3.2878482 -3.2912878
#> 28   5.6858865 -2.6047891 -2.6018232   5.66213783 -2.5838411 -2.6031045
#> 29 -17.0800437 -5.1886612 -5.1870338 -13.89327416 -5.1914030 -5.1969199
#> 30  -1.0395227  1.7895019  1.7711890  -1.40393388  1.7643551  1.7726942
#> 31   5.4575060  7.2681505  7.2712727   3.15399249  7.2839536  7.2651630
#> 32  -4.7049402 -3.2628058 -3.2379026  -3.59502409 -3.2680513 -3.2479541
#> 33  -8.7892965 -0.8337623 -0.8495717  -7.44000180 -0.8254203 -0.8158928
#> 34  -2.1591178 -0.8481164 -0.8434961  -1.80265646 -0.8207269 -0.8308481
#> 35   9.2966703  3.0183710  3.0299480   7.45682359  3.0305849  3.0205319
#> 36  10.5207046 -3.0282407 -3.0302081   9.96658949 -3.0095733 -3.0168071
#> 37   0.6451022  1.0352329  1.0558159   0.26730883  1.0041943  1.0438538
#> 38   1.1579346  2.5365267  2.5059079   0.37732536  2.5226571  2.5100652
#> 39  -3.3992341  9.7051131  9.7042880  -5.37331789  9.7237207  9.7139544
#> 40  -7.8383363 -3.3779091 -3.3829224  -5.94229862 -3.3659934 -3.3777016
#> 41   0.4191362 -0.1637186 -0.1594651   0.34701916 -0.1824409 -0.1573322
#> 42  -1.4898387  6.1247564  6.1166666  -2.90415682  6.1195073  6.1122943
#> 43   4.0672766  6.8022399  6.8096388   1.79069462  6.7891900  6.8035978
#> 44  -5.9318495 -5.2762514 -5.2820730  -3.93749774 -5.2646544 -5.2669288
#> 45   9.6804485 -3.2149632 -3.2137196   9.18282119 -3.2083586 -3.2106410
#> 46   1.1672461 -3.5046078 -3.4920273   1.84278077 -3.4948853 -3.5018542
#> 47  -3.3920787  5.2510265  5.2362713  -4.11521282  5.2608203  5.2779395
#> 48 -10.0334279 -0.2731161 -0.2787071  -8.75362808 -0.2847004 -0.2638168
#> 49  -2.0019224 -6.4093444 -6.4192146  -0.24279803 -6.4249104 -6.4021046
#> 50   9.5877734  0.3986771  0.4057520   8.34748121  0.4076436  0.3814497
#>              X6
#> 1    7.69184197
#> 2   -3.05457299
#> 3   -8.11665746
#> 4   -1.94095460
#> 5   -3.84606481
#> 6    7.07163918
#> 7   -6.11090074
#> 8    9.42871207
#> 9   -7.74364899
#> 10  11.59737361
#> 11  -0.17232575
#> 12   4.23501313
#> 13   6.95630485
#> 14   9.13668879
#> 15   5.82190599
#> 16  -2.47617381
#> 17   6.28597735
#> 18  11.97891199
#> 19   2.61375046
#> 20  10.10901086
#> 21 -10.34377351
#> 22 -10.31720288
#> 23   0.06794467
#> 24 -13.23041017
#> 25   0.66801002
#> 26  -7.48936784
#> 27  -2.64078254
#> 28   5.66002484
#> 29 -13.87689041
#> 30  -1.42375030
#> 31   3.15641036
#> 32  -3.58542765
#> 33  -7.41152667
#> 34  -1.80187281
#> 35   7.46617413
#> 36   9.96199656
#> 37   0.27107014
#> 38   0.37332456
#> 39  -5.36959906
#> 40  -5.95801110
#> 41   0.35476240
#> 42  -2.88385793
#> 43   1.79124457
#> 44  -3.93164118
#> 45   9.14356466
#> 46   1.83626167
#> 47  -4.10011162
#> 48  -8.73300529
#> 49  -0.26519363
#> 50   8.36836764

cvtable(summary(cv.plsR(Y~.,data=dataAstar2,5,NK=100, verbose=FALSE)))

#> ____************************************************____

#> Error in is.data.frame(data): object 'dataAstar2' not found

dimX <- 6
Astar <- 3
simul_data_UniYX(dimX,Astar)

#>            Y           X1           X2           X3           X4           X5 
#> -10.77810061  -1.51920347   0.05240594  -9.94226364  -1.51366148   0.06004913 
#>           X6 
#>  -9.91911328 

(dataAstar3 <- data.frame(t(replicate(50,simul_data_UniYX(dimX,Astar)))))

#>             Y          X1          X2          X3          X4          X5
#> 1  13.5624009  -3.4032160   5.4319923   7.0686717  -3.4006451   5.4237726
#> 2  -3.9058453  -0.6777562  -3.3491685  -1.5629716  -0.6561557  -3.3519911
#> 3  -1.7737095  -5.2559980  -6.6798238   0.5266034  -5.2527070  -6.6875592
#> 4  -1.8578467  -7.4350761  -2.5954319  -2.8359573  -7.4260181  -2.5801070
#> 5  -6.1558908  -3.7712913  -7.3205670  -2.1656845  -3.7727895  -7.3020670
#> 6  -9.7256489   5.5600727  -5.7306476  -3.0755835   5.5538143  -5.7540670
#> 7   3.2171523   0.8031348   1.0019977   2.2451047   0.7861168   1.0142705
#> 8  -0.1470448   6.5340719  -1.2605345   3.1571212   6.5105560  -1.2438242
#> 9  -5.8165654  -3.6330842  -0.7654251  -6.1938357  -3.6571109  -0.7387401
#> 10  0.7889181  -7.0354064   1.9980639  -3.0557540  -7.0414552   1.9926039
#> 11 -3.0999240  -7.7332770  -8.9302148  -0.2266418  -7.7346143  -8.9136322
#> 12 -2.6527575  -6.9999202  -2.1092626  -3.4991308  -6.9922302  -2.1131293
#> 13 -0.2273686   0.8991743   0.5853912  -0.2098982   0.9091902   0.5842820
#> 14  2.5003005   4.7226682   8.1955769  -0.9983622   4.7293561   8.2123281
#> 15 -4.1776649  -6.2278873  -5.4065216  -2.6658220  -6.2162342  -5.4200793
#> 16 -3.1504660  -2.6172075  -2.3796848  -2.3603750  -2.5954842  -2.4014755
#> 17 -6.9219474   2.7840320  -3.9512438  -2.4911974   2.7832919  -3.9337724
#> 18  5.2449185 -10.7309856  -2.9816801   2.3531400 -10.7101971  -2.9767484
#> 19  3.2664646   1.5764947   1.8093088   2.5083556   1.5760452   1.8168125
#> 20 -9.1080604  14.8310629  11.9355554  -9.7552874  14.8445733  11.9092841
#> 21  1.1250806  -3.9780479   0.6645192  -0.9279217  -3.9946076   0.6662758
#> 22 -2.6191707   2.4059436  -8.8370167   4.1547020   2.4006100  -8.8208405
#> 23 -4.8177175   4.2554185   0.1007634  -2.9107540   4.2624944   0.1116745
#> 24 -3.1692248   5.5278763  -1.4890608   0.3393101   5.5254609  -1.4962202
#> 25 -5.1199083  -2.3649326  -5.1731804  -2.0911590  -2.3719328  -5.1722637
#> 26 -2.8018117   4.8647596  -6.7465951   3.6582067   4.8590820  -6.7447793
#> 27 11.7607774  -4.5876939   1.7725027   7.4743901  -4.5754621   1.7608623
#> 28 -4.6112595  -1.1108051 -10.2341794   1.8970719  -1.0906935 -10.2306188
#> 29  5.4659816   5.8941502   7.4095104   2.3601630   5.8899130   7.4181778
#> 30 -6.5031636   5.5056802   5.1682288  -7.0176510   5.5099042   5.1724530
#> 31  5.8116119   4.2349245  10.6119065   0.3240308   4.2367188  10.5996275
#> 32  9.9460676  -9.0601973   3.4597340   3.2252262  -9.0647765   3.4540188
#> 33  8.6697083  -1.1842448   0.6650803   6.7977492  -1.1947321   0.6502210
#> 34 -1.7486882   2.0401189  -2.3587462   0.6509778   2.0515882  -2.3630276
#> 35  5.1437345  -0.5191375   2.3512528   3.2441842  -0.5039723   2.3494899
#> 36 17.7893663  -3.5321520  12.8270367   6.0347113  -3.5405553  12.8042306
#> 37  1.6098197  -5.0569358  -7.8196608   4.3150615  -5.0557004  -7.8343312
#> 38 -4.6134799   7.5912666   3.0175635  -3.1564042   7.6135055   3.0198611
#> 39 -9.5283588   0.1803370   3.9418348 -10.5797144   0.2221012   3.9507698
#> 40 -5.5145172  -1.5209476  -5.9055671  -1.9033426  -1.5154657  -5.8952534
#> 41 -2.5293787   4.0394267   7.9618795  -5.6078215   4.0462038   7.9566530
#> 42  3.3190079  -1.9812218   4.7472114  -1.0886730  -1.9700341   4.7629379
#> 43 -2.4465910  -5.0259850   5.4260163  -7.6370223  -5.0379558   5.4082411
#> 44  0.6281354   1.1270520   3.9879378  -1.5402667   1.1259538   3.9998097
#> 45  1.0510882   5.0286650   2.1481979   1.4019631   5.0171404   2.1242041
#> 46  6.8767027   0.9569209   7.0548605   1.9075775   0.9540479   7.0319538
#> 47  4.8272469  -0.2588829   3.8730948   1.4073039  -0.2770279   3.8589375
#> 48 -8.5724562  -2.2345447  -2.8232356  -6.7144273  -2.2230256  -2.8119039
#> 49 -2.9475943  -4.8926370  -6.1909277  -0.7781096  -4.9061285  -6.1902049
#> 50 -2.4620743   7.8370564   2.0828816  -0.2931437   7.8401758   2.0667762
#>             X6
#> 1    7.0556158
#> 2   -1.5977286
#> 3    0.5295466
#> 4   -2.8198460
#> 5   -2.1823580
#> 6   -3.0560219
#> 7    2.2534308
#> 8    3.1780129
#> 9   -6.1801532
#> 10  -3.0607988
#> 11  -0.2452582
#> 12  -3.4949962
#> 13  -0.1904027
#> 14  -0.9912375
#> 15  -2.6760536
#> 16  -2.3615335
#> 17  -2.5019638
#> 18   2.3567144
#> 19   2.4964999
#> 20  -9.7527388
#> 21  -0.9421372
#> 22   4.1482887
#> 23  -2.9089414
#> 24   0.3596660
#> 25  -2.1101153
#> 26   3.6570708
#> 27   7.4652494
#> 28   1.8781354
#> 29   2.3655071
#> 30  -7.0117873
#> 31   0.3255542
#> 32   3.2161454
#> 33   6.7819415
#> 34   0.6365060
#> 35   3.2288021
#> 36   6.0423344
#> 37   4.3072639
#> 38  -3.1337472
#> 39 -10.5897780
#> 40  -1.9041799
#> 41  -5.5859728
#> 42  -1.0826979
#> 43  -7.6284155
#> 44  -1.5553101
#> 45   1.4026965
#> 46   1.9071408
#> 47   1.4128462
#> 48  -6.7199935
#> 49  -0.7446489
#> 50  -0.2942979

cvtable(summary(cv.plsR(Y~.,data=dataAstar3,5,NK=100, verbose=FALSE)))

#> ____************************************************____

#> Error in is.data.frame(data): object 'dataAstar3' not found

dimX <- 6
Astar <- 4
simul_data_UniYX(dimX,Astar)

#>         Y        X1        X2        X3        X4        X5        X6 
#> -9.070558 -2.536314 -8.040252 -4.857946  3.801125 -1.696731  1.467276 

(dataAstar4 <- data.frame(t(replicate(50,simul_data_UniYX(dimX,Astar)))))

#>              Y           X1          X2          X3           X4            X5
#> 1   -2.4565783  -1.20650869  -3.9718487  -0.3640601  -1.50677699  -4.230686216
#> 2   -2.3613044   6.56756460   9.9946235  -6.4738863   6.54991468   9.960202708
#> 3    9.5557571   5.73624449  -3.1796840  13.7028174   0.73720024  -8.167294366
#> 4   -1.6937627   8.46092919   0.4898175   1.2596673  12.64112387   4.693392678
#> 5    0.7694968   5.23991932   2.1456532   1.4979151   7.85371994   4.801421329
#> 6    0.5209114   1.70298995  -0.7364991   1.3157237  -4.64266001  -7.095291568
#> 7    0.2775724 -10.95683197  -1.6541087  -2.8792530 -10.27117232  -0.985837650
#> 8  -11.1090119   4.78085796  -5.5864841  -5.6312041   6.54459173  -3.791696644
#> 9   -4.7592285  -7.73241851  -8.1735708  -2.8872960  -0.86591567  -1.296814518
#> 10  11.0218170 -10.52643569   4.6707163   4.1456769 -10.71549148   4.503018858
#> 11   7.8326177  -3.83084922  -4.6302412   9.2294478  -4.17876046  -4.948503553
#> 12  -2.2145302   0.49011812   0.2806683  -2.3522424   2.84404629   2.648778808
#> 13   6.4637226   6.21760998   5.7702597   5.2122211  -0.02506007  -0.462858992
#> 14  -9.5743378  -1.53908766  -1.5162603  -9.4381296  -1.38958065  -1.368555530
#> 15  -2.6169403   0.43454604  -1.1628363  -1.6852919  -3.09273576  -4.698695469
#> 16 -14.8127371   8.08955620   7.7300699 -16.7735613  11.08564177  10.746420764
#> 17   4.5725028  -8.04864363  -1.2919497   2.5469716  -9.25318261  -2.480055058
#> 18   5.5443317   6.27425571   9.1430312   2.3469449   8.84572290  11.742066569
#> 19   0.4064244   0.77935156   5.6283527  -2.8238844   1.89175947   6.713794826
#> 20   3.8564088  -1.29443315  -1.1091808   3.8418468  -3.58088554  -3.405293646
#> 21  -0.2538374  -3.61159839  -7.7100003   3.2550211  -5.53557237  -9.612343840
#> 22  -0.6317777  -1.90439402  -1.6000073  -0.1286038  -2.22760179  -1.934162130
#> 23   3.2062936  -3.78224432   4.9920949  -1.4539928  -6.58613857   2.192703053
#> 24   8.5328160  -2.84472549   2.2057062   6.0633600  -4.83099359   0.201337603
#> 25  -1.2781508  -8.25804295   1.1216402  -5.1364501  -9.34208104   0.044642841
#> 26  15.6537831   5.08009475   8.4802640  12.5562517   1.38406761   4.803896519
#> 27  -4.0952604   5.54592603  -4.0198613   0.1265071   9.55436901   0.007104527
#> 28  -3.6910529   0.71233137  -5.7248216   0.1718035   2.56391113  -3.922212895
#> 29  -6.4111423  -4.23316228  -7.0543324  -3.4820191   4.15916746   1.331234734
#> 30  -0.8481313  -1.42448618  -1.6672377  -0.2928312   0.22242112  -0.045415618
#> 31  12.7558696  -0.69186906   6.1847710   8.4511132  -4.25821614   2.646709994
#> 32   8.7057923   8.61918792  10.8612097   5.2167651   8.29218857  10.538699809
#> 33   0.8088476  -1.40264064   0.9039005  -0.4756979  -4.61571213  -2.300538210
#> 34  -4.0369714  -4.71853598  -0.2779843  -5.9569947  -3.62930425   0.813181773
#> 35   4.4729387   1.71827345   9.1130064  -0.5192941  -4.83767900   2.547594780
#> 36   0.1553155  -2.89814012 -12.9304916   6.9749988  -3.15787557 -13.187899553
#> 37 -17.1389523  -2.84808555 -13.8232629  -9.9679297   0.02112013 -10.924398791
#> 38  -1.5836340   8.87892679   3.2062493  -0.1991505   9.33000565   3.676466086
#> 39   6.7073010   3.06692143   5.1974466   4.7153279  -2.11625169   0.019791999
#> 40   7.4020140  -3.26282732  -3.8622895   8.6293526  -2.61416937  -3.214909788
#> 41  -3.1183632  -9.02343147 -13.2570871   1.5013350 -10.89357696 -15.073031357
#> 42  -2.5986712  -4.04161617  -6.0157633  -0.2305151  -6.16290467  -8.153974485
#> 43  -3.4656031 -20.15996560 -13.0729753  -3.3984204 -17.72058003 -10.622248008
#> 44  -2.5234593   0.84612213   1.5296099  -3.0895013  -0.58414627   0.119785059
#> 45 -10.0661633  -0.99255992 -11.3024341  -3.4470199   2.47946682  -7.792226940
#> 46   5.0534928  -6.97786009  -2.0019856   3.8041474  -6.30809510  -1.294972972
#> 47   2.2721017   0.05725126   1.8153012   1.3026439  -2.05133226  -0.276708572
#> 48  -4.0903371  -6.67522471  -5.6870103  -3.0711909  -7.05585696  -6.053843608
#> 49 -11.3182926   2.34276845  -3.0862643  -8.5541829   3.63070098  -1.771767240
#> 50   7.5832686  -6.94978370  -0.3238126   5.1145066  -1.67226098   4.930192686
#>              X6
#> 1   -0.63812620
#> 2   -6.50714301
#> 3    8.72801869
#> 4    5.46064493
#> 5    4.13607533
#> 6   -5.03757118
#> 7   -2.23825050
#> 8   -3.84743484
#> 9    3.98635064
#> 10   3.93717339
#> 11   8.89041964
#> 12  -0.02122645
#> 13  -1.00931633
#> 14  -9.28113629
#> 15  -5.21097097
#> 16 -13.75465564
#> 17   1.38119530
#> 18   4.92899315
#> 19  -1.72318942
#> 20   1.56732448
#> 21   1.36976872
#> 22  -0.46684173
#> 23  -4.25413973
#> 24   4.06374000
#> 25  -6.23032231
#> 26   8.89348509
#> 27   4.14082783
#> 28   2.00108633
#> 29   4.91829050
#> 30   1.33666290
#> 31   4.92137314
#> 32   4.90785264
#> 33  -3.69063202
#> 34  -4.83322626
#> 35  -7.09685408
#> 36   6.70230100
#> 37  -7.10877437
#> 38   0.28664645
#> 39  -0.48577302
#> 40   9.26032973
#> 41  -0.33667881
#> 42  -2.33035246
#> 43  -0.92667129
#> 44  -4.54180370
#> 45   0.07665750
#> 46   4.45695790
#> 47  -0.82110948
#> 48  -3.43430697
#> 49  -7.24153072
#> 50  10.37111882

cvtable(summary(cv.plsR(Y~.,data=dataAstar4,5,NK=100, verbose=FALSE)))

#> ____************************************************____

#> Error in is.data.frame(data): object 'dataAstar4' not found

rm(list=c("dimX","Astar","dataAstar2","dataAstar3","dataAstar4"))