Extracts and computes information criteria and fits statistics for k-fold cross validated partial least squares models

This function extracts and computes information criteria and fits statistics for k-fold cross validated partial least squares models for both formula or classic specifications of the model.

Usage

kfolds2CVinfos_lm(pls_kfolds, MClassed = FALSE, verbose = TRUE)

Arguments

pls_kfolds

an object computed using PLS_lm_kfoldcv

MClassed

should number of miss classed be computed

verbose

should infos be displayed ?

Value

list

table of fit statistics for first group partition

list()

...

list

table of fit statistics for last group partition

Details

The Mclassed option should only set to TRUE if the response is binary.

Note

Use summary and cv.plsR instead.

References

Nicolas Meyer, Myriam Maumy-Bertrand et Frédéric Bertrand (2010). Comparing the linear and the logistic PLS regression with qualitative predictors: application to allelotyping data. Journal de la Societe Francaise de Statistique, 151(2), pages 1-18. https://www.numdam.org/item/JSFS_2010__151_2_1_0/

See also

Author

Frédéric Bertrand
frederic.bertrand@lecnam.net
https://fbertran.github.io/homepage/

Examples

# \donttest{
data(Cornell)
summary(cv.plsR(Y~.,data=Cornell,nt=10,K=6,verbose=FALSE))
#> ____************************************************____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> Warning : 1 2 3 4 5 6 7 < 10^{-12}
#> Warning only 6 components could thus be extracted
#> ____Predicting X without NA neither in X nor in Y____
#> ****________________________________________________****
#> 
#> 
#> NK: 1
#> [[1]]
#>                AIC    Q2cum_Y LimQ2_Y        Q2_Y  PRESS_Y      RSS_Y      R2_Y
#> Nb_Comp_0 82.01205         NA      NA          NA       NA 467.796667        NA
#> Nb_Comp_1 53.15173  0.8833692  0.0975  0.88336915 54.55952  35.742486 0.9235940
#> Nb_Comp_2 41.08283  0.8805715  0.0975 -0.02398696 36.59984  11.066606 0.9763431
#> Nb_Comp_3 32.06411  0.7922562  0.0975 -0.73948309 19.25017   4.418081 0.9905556
#> Nb_Comp_4 33.76477  0.2690413  0.0975 -2.51855868 15.54528   4.309235 0.9907882
#> Nb_Comp_5 33.34373 -3.6031367  0.0975 -5.29739609 27.13696   3.521924 0.9924713
#> Nb_Comp_6 35.25533         NA  0.0975          NA       NA   3.496074 0.9925265
#>              AIC.std  DoF.dof sigmahat.dof    AIC.dof    BIC.dof GMDL.dof
#> Nb_Comp_0  37.010388 1.000000    6.5212706 46.0708838 47.7893514 27.59461
#> Nb_Comp_1   8.150064 2.740749    1.8665281  4.5699686  4.9558156 21.34020
#> Nb_Comp_2  -3.918831 5.085967    1.1825195  2.1075461  2.3949331 27.40202
#> Nb_Comp_3 -12.937550 5.121086    0.7488308  0.8467795  0.9628191 24.40842
#> Nb_Comp_4 -11.236891 5.103312    0.7387162  0.8232505  0.9357846 24.23105
#> Nb_Comp_5 -11.657929 6.006316    0.7096382  0.7976101  0.9198348 28.21184
#> Nb_Comp_6  -9.746328 7.000002    0.7633343  0.9711322  1.1359502 33.18348
#>           DoF.naive sigmahat.naive  AIC.naive  BIC.naive GMDL.naive
#> Nb_Comp_0         1      6.5212706 46.0708838 47.7893514   27.59461
#> Nb_Comp_1         2      1.8905683  4.1699567  4.4588195   18.37545
#> Nb_Comp_2         3      1.1088836  1.5370286  1.6860917   17.71117
#> Nb_Comp_3         4      0.7431421  0.7363469  0.8256118   19.01033
#> Nb_Comp_4         5      0.7846050  0.8721072  0.9964867   24.16510
#> Nb_Comp_5         6      0.7661509  0.8804809  1.0227979   28.64206
#> Nb_Comp_6         7      0.8361907  1.1070902  1.3048716   33.63927
#> attr(,"computed_nt")
#> [1] 6
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"

data(pine)
summary(cv.plsR(x11~.,data=pine,nt=10,NK=3,verbose=FALSE),verbose=FALSE)
#> [[1]]
#>                 AIC       Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0  82.41888            NA      NA         NA       NA 20.800152
#> Nb_Comp_1  63.61896    0.33236241  0.0975  0.3323624 13.88696 11.074659
#> Nb_Comp_2  58.47638    0.08602135  0.0975 -0.3689742 15.16092  8.919303
#> Nb_Comp_3  56.55421   -0.64686696  0.0975 -0.8018659 16.07139  7.919786
#> Nb_Comp_4  54.35053   -2.07842135  0.0975 -0.8692593 14.80413  6.972542
#> Nb_Comp_5  55.99834   -5.26758156  0.0975 -1.0359726 14.19590  6.898523
#> Nb_Comp_6  57.69592  -11.24110783  0.0975 -0.9530831 13.47339  6.835594
#> Nb_Comp_7  59.37953  -23.44871994  0.0975 -0.9972637 13.65248  6.770369
#> Nb_Comp_8  61.21213  -51.30545992  0.0975 -1.1393946 14.48449  6.736112
#> Nb_Comp_9  63.18426 -112.65581728  0.0975 -1.1729245 14.63706  6.730426
#> Nb_Comp_10 65.15982 -248.80085723  0.0975 -1.1978713 14.79261  6.725443
#>                 R2_Y  AIC.std   DoF.dof sigmahat.dof   AIC.dof   BIC.dof
#> Nb_Comp_0         NA 96.63448  1.000000    0.8062287 0.6697018 0.6991787
#> Nb_Comp_1  0.4675684 77.83455  3.176360    0.5994089 0.4047616 0.4565153
#> Nb_Comp_2  0.5711905 72.69198  7.133559    0.5761829 0.4138120 0.5212090
#> Nb_Comp_3  0.6192438 70.76981  8.778329    0.5603634 0.4070516 0.5320535
#> Nb_Comp_4  0.6647841 68.56612  8.427874    0.5221703 0.3505594 0.4547689
#> Nb_Comp_5  0.6683426 70.21393  9.308247    0.5285695 0.3666578 0.4845912
#> Nb_Comp_6  0.6713681 71.91152  9.291931    0.5259794 0.3629363 0.4795121
#> Nb_Comp_7  0.6745039 73.59512  9.756305    0.5284535 0.3702885 0.4938445
#> Nb_Comp_8  0.6761508 75.42772 10.363948    0.5338475 0.3831339 0.5170783
#> Nb_Comp_9  0.6764242 77.39986 10.732148    0.5378277 0.3920957 0.5328747
#> Nb_Comp_10 0.6766638 79.37542 11.000000    0.5407500 0.3987417 0.5446065
#>             GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0  -3.605128         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1  -9.875081         2      0.5977015 0.3788984 0.4112998 -11.451340
#> Nb_Comp_2  -6.985517         3      0.5452615 0.3243383 0.3647862 -12.822703
#> Nb_Comp_3  -6.260610         4      0.5225859 0.3061986 0.3557368 -12.756838
#> Nb_Comp_4  -8.152986         5      0.4990184 0.2867496 0.3432131 -12.811575
#> Nb_Comp_5  -7.111583         6      0.5054709 0.3019556 0.3714754 -11.329638
#> Nb_Comp_6  -7.233043         7      0.5127450 0.3186757 0.4021333  -9.918688
#> Nb_Comp_7  -6.742195         8      0.5203986 0.3364668 0.4347156  -8.592770
#> Nb_Comp_8  -6.038372         9      0.5297842 0.3572181 0.4717708  -7.287834
#> Nb_Comp_9  -5.600235        10      0.5409503 0.3813021 0.5140048  -6.008747
#> Nb_Comp_10 -5.288422        11      0.5529032 0.4076026 0.5600977  -4.799453
#> attr(,"computed_nt")
#> [1] 10
#> 
#> [[2]]
#>                 AIC     Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y      R2_Y
#> Nb_Comp_0  82.41888          NA      NA         NA       NA 20.800152        NA
#> Nb_Comp_1  63.61896   0.2883611  0.0975  0.2883611 14.80220 11.074659 0.4675684
#> Nb_Comp_2  58.47638   0.1495497  0.0975 -0.1950588 13.23487  8.919303 0.5711905
#> Nb_Comp_3  56.55421  -0.2339077  0.0975 -0.4508876 12.94091  7.919786 0.6192438
#> Nb_Comp_4  54.35053  -0.8654831  0.0975 -0.5118498 11.97353  6.972542 0.6647841
#> Nb_Comp_5  55.99834  -2.0764393  0.0975 -0.6491381 11.49868  6.898523 0.6683426
#> Nb_Comp_6  57.69592  -3.9268641  0.0975 -0.6014826 11.04787  6.835594 0.6713681
#> Nb_Comp_7  59.37953  -7.4958507  0.0975 -0.7243932 11.78725  6.770369 0.6745039
#> Nb_Comp_8  61.21213 -14.0681481  0.0975 -0.7735891 12.00785  6.736112 0.6761508
#> Nb_Comp_9  63.18426 -28.2071979  0.0975 -0.9383402 13.05688  6.730426 0.6764242
#> Nb_Comp_10 65.15982 -55.9051868  0.0975 -0.9483275 13.11307  6.725443 0.6766638
#>             AIC.std   DoF.dof sigmahat.dof   AIC.dof   BIC.dof  GMDL.dof
#> Nb_Comp_0  96.63448  1.000000    0.8062287 0.6697018 0.6991787 -3.605128
#> Nb_Comp_1  77.83455  3.176360    0.5994089 0.4047616 0.4565153 -9.875081
#> Nb_Comp_2  72.69198  7.133559    0.5761829 0.4138120 0.5212090 -6.985517
#> Nb_Comp_3  70.76981  8.778329    0.5603634 0.4070516 0.5320535 -6.260610
#> Nb_Comp_4  68.56612  8.427874    0.5221703 0.3505594 0.4547689 -8.152986
#> Nb_Comp_5  70.21393  9.308247    0.5285695 0.3666578 0.4845912 -7.111583
#> Nb_Comp_6  71.91152  9.291931    0.5259794 0.3629363 0.4795121 -7.233043
#> Nb_Comp_7  73.59512  9.756305    0.5284535 0.3702885 0.4938445 -6.742195
#> Nb_Comp_8  75.42772 10.363948    0.5338475 0.3831339 0.5170783 -6.038372
#> Nb_Comp_9  77.39986 10.732148    0.5378277 0.3920957 0.5328747 -5.600235
#> Nb_Comp_10 79.37542 11.000000    0.5407500 0.3987417 0.5446065 -5.288422
#>            DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0          1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1          2      0.5977015 0.3788984 0.4112998 -11.451340
#> Nb_Comp_2          3      0.5452615 0.3243383 0.3647862 -12.822703
#> Nb_Comp_3          4      0.5225859 0.3061986 0.3557368 -12.756838
#> Nb_Comp_4          5      0.4990184 0.2867496 0.3432131 -12.811575
#> Nb_Comp_5          6      0.5054709 0.3019556 0.3714754 -11.329638
#> Nb_Comp_6          7      0.5127450 0.3186757 0.4021333  -9.918688
#> Nb_Comp_7          8      0.5203986 0.3364668 0.4347156  -8.592770
#> Nb_Comp_8          9      0.5297842 0.3572181 0.4717708  -7.287834
#> Nb_Comp_9         10      0.5409503 0.3813021 0.5140048  -6.008747
#> Nb_Comp_10        11      0.5529032 0.4076026 0.5600977  -4.799453
#> attr(,"computed_nt")
#> [1] 10
#> 
#> [[3]]
#>                 AIC      Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0  82.41888           NA      NA         NA       NA 20.800152
#> Nb_Comp_1  63.61896    0.3706202  0.0975  0.3706202 13.09120 11.074659
#> Nb_Comp_2  58.47638    0.2227707  0.0975 -0.2349129 13.67624  8.919303
#> Nb_Comp_3  56.55421   -0.2745327  0.0975 -0.6398414 14.62624  7.919786
#> Nb_Comp_4  54.35053   -0.9662890  0.0975 -0.5427529 12.21827  6.972542
#> Nb_Comp_5  55.99834   -2.2579216  0.0975 -0.6568885 11.55272  6.898523
#> Nb_Comp_6  57.69592   -4.5773450  0.0975 -0.7119335 11.80981  6.835594
#> Nb_Comp_7  59.37953   -9.2588465  0.0975 -0.8393781 12.57324  6.770369
#> Nb_Comp_8  61.21213  -18.3325334  0.0975 -0.8844744 12.75859  6.736112
#> Nb_Comp_9  63.18426  -43.1148562  0.0975 -1.2818973 15.37112  6.730426
#> Nb_Comp_10 65.15982 -108.3254239  0.0975 -1.4781997 16.67934  6.725443
#>                 R2_Y  AIC.std   DoF.dof sigmahat.dof   AIC.dof   BIC.dof
#> Nb_Comp_0         NA 96.63448  1.000000    0.8062287 0.6697018 0.6991787
#> Nb_Comp_1  0.4675684 77.83455  3.176360    0.5994089 0.4047616 0.4565153
#> Nb_Comp_2  0.5711905 72.69198  7.133559    0.5761829 0.4138120 0.5212090
#> Nb_Comp_3  0.6192438 70.76981  8.778329    0.5603634 0.4070516 0.5320535
#> Nb_Comp_4  0.6647841 68.56612  8.427874    0.5221703 0.3505594 0.4547689
#> Nb_Comp_5  0.6683426 70.21393  9.308247    0.5285695 0.3666578 0.4845912
#> Nb_Comp_6  0.6713681 71.91152  9.291931    0.5259794 0.3629363 0.4795121
#> Nb_Comp_7  0.6745039 73.59512  9.756305    0.5284535 0.3702885 0.4938445
#> Nb_Comp_8  0.6761508 75.42772 10.363948    0.5338475 0.3831339 0.5170783
#> Nb_Comp_9  0.6764242 77.39986 10.732148    0.5378277 0.3920957 0.5328747
#> Nb_Comp_10 0.6766638 79.37542 11.000000    0.5407500 0.3987417 0.5446065
#>             GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0  -3.605128         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1  -9.875081         2      0.5977015 0.3788984 0.4112998 -11.451340
#> Nb_Comp_2  -6.985517         3      0.5452615 0.3243383 0.3647862 -12.822703
#> Nb_Comp_3  -6.260610         4      0.5225859 0.3061986 0.3557368 -12.756838
#> Nb_Comp_4  -8.152986         5      0.4990184 0.2867496 0.3432131 -12.811575
#> Nb_Comp_5  -7.111583         6      0.5054709 0.3019556 0.3714754 -11.329638
#> Nb_Comp_6  -7.233043         7      0.5127450 0.3186757 0.4021333  -9.918688
#> Nb_Comp_7  -6.742195         8      0.5203986 0.3364668 0.4347156  -8.592770
#> Nb_Comp_8  -6.038372         9      0.5297842 0.3572181 0.4717708  -7.287834
#> Nb_Comp_9  -5.600235        10      0.5409503 0.3813021 0.5140048  -6.008747
#> Nb_Comp_10 -5.288422        11      0.5529032 0.4076026 0.5600977  -4.799453
#> attr(,"computed_nt")
#> [1] 10
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"
data(pineNAX21)
summary(cv.plsR(x11~.,data=pineNAX21,nt=10,NK=3,
verbose=FALSE),verbose=FALSE)
#> [[1]]
#>                AIC      Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y      R2_Y
#> Nb_Comp_0 82.41888           NA      NA         NA       NA 20.800152        NA
#> Nb_Comp_1 63.69250    0.1644558  0.0975  0.1644558 17.37945 11.099368 0.4663804
#> Nb_Comp_2 58.35228   -0.1643470  0.0975 -0.3935193 15.46718  8.885823 0.5728001
#> Nb_Comp_3 56.36553   -1.0741068  0.0975 -0.7813476 15.82874  7.874634 0.6214146
#> Nb_Comp_4 54.02416   -2.2496103  0.0975 -0.5667516 12.33760  6.903925 0.6680830
#> Nb_Comp_5 55.80450   -5.6345125  0.0975 -1.0416333 14.09528  6.858120 0.6702851
#> Nb_Comp_6 57.45753  -11.9901592  0.0975 -0.9579674 13.42798  6.786392 0.6737336
#> Nb_Comp_7 58.73951  -21.9760455  0.0975 -0.7687270 12.00327  6.640327 0.6807558
#> Nb_Comp_8 60.61227  -43.4713170  0.0975 -0.9355514 12.85269  6.614773 0.6819844
#> Nb_Comp_9 62.25948 -171.1558997  0.0975 -2.8711671 25.60689  6.544432 0.6853661
#>            AIC.std DoF.dof sigmahat.dof AIC.dof BIC.dof GMDL.dof DoF.naive
#> Nb_Comp_0 96.63448      NA           NA      NA      NA       NA         1
#> Nb_Comp_1 77.90810      NA           NA      NA      NA       NA         2
#> Nb_Comp_2 72.56787      NA           NA      NA      NA       NA         3
#> Nb_Comp_3 70.58113      NA           NA      NA      NA       NA         4
#> Nb_Comp_4 68.23976      NA           NA      NA      NA       NA         5
#> Nb_Comp_5 70.02009      NA           NA      NA      NA       NA         6
#> Nb_Comp_6 71.67313      NA           NA      NA      NA       NA         7
#> Nb_Comp_7 72.95511      NA           NA      NA      NA       NA         8
#> Nb_Comp_8 74.82787      NA           NA      NA      NA       NA         9
#> Nb_Comp_9 76.47507      NA           NA      NA      NA       NA        10
#>           sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1      0.5983679 0.3797438 0.4122175 -11.413749
#> Nb_Comp_2      0.5442372 0.3231208 0.3634169 -12.847656
#> Nb_Comp_3      0.5210941 0.3044529 0.3537087 -12.776843
#> Nb_Comp_4      0.4965569 0.2839276 0.3398355 -12.891035
#> Nb_Comp_5      0.5039885 0.3001871 0.3692997 -11.349498
#> Nb_Comp_6      0.5108963 0.3163819 0.3992388  -9.922119
#> Nb_Comp_7      0.5153766 0.3300041 0.4263658  -8.696873
#> Nb_Comp_8      0.5249910 0.3507834 0.4632727  -7.337679
#> Nb_Comp_9      0.5334234 0.3707649 0.4998004  -6.033403
#> attr(,"computed_nt")
#> [1] 9
#> 
#> [[2]]
#>                AIC      Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y      R2_Y
#> Nb_Comp_0 82.41888           NA      NA         NA       NA 20.800152        NA
#> Nb_Comp_1 63.69250   0.31341253  0.0975  0.3134125 14.28112 11.099368 0.4663804
#> Nb_Comp_2 58.35228   0.08234691  0.0975 -0.3365421 14.83477  8.885823 0.5728001
#> Nb_Comp_3 56.36553  -0.45720572  0.0975 -0.5879702 14.11042  7.874634 0.6214146
#> Nb_Comp_4 54.02416  -1.20066552  0.0975 -0.5101955 11.89224  6.903925 0.6680830
#> Nb_Comp_5 55.80450  -2.54778037  0.0975 -0.6121398 11.13009  6.858120 0.6702851
#> Nb_Comp_6 57.45753  -5.55466263  0.0975 -0.8475390 12.67064  6.786392 0.6737336
#> Nb_Comp_7 58.73951 -11.31988857  0.0975 -0.8795610 12.75544  6.640327 0.6807558
#> Nb_Comp_8 60.61227 -19.41015264  0.0975 -0.6566832 11.00092  6.614773 0.6819844
#> Nb_Comp_9 62.25948 -40.36330967  0.0975 -1.0266046 13.40553  6.544432 0.6853661
#>            AIC.std DoF.dof sigmahat.dof AIC.dof BIC.dof GMDL.dof DoF.naive
#> Nb_Comp_0 96.63448      NA           NA      NA      NA       NA         1
#> Nb_Comp_1 77.90810      NA           NA      NA      NA       NA         2
#> Nb_Comp_2 72.56787      NA           NA      NA      NA       NA         3
#> Nb_Comp_3 70.58113      NA           NA      NA      NA       NA         4
#> Nb_Comp_4 68.23976      NA           NA      NA      NA       NA         5
#> Nb_Comp_5 70.02009      NA           NA      NA      NA       NA         6
#> Nb_Comp_6 71.67313      NA           NA      NA      NA       NA         7
#> Nb_Comp_7 72.95511      NA           NA      NA      NA       NA         8
#> Nb_Comp_8 74.82787      NA           NA      NA      NA       NA         9
#> Nb_Comp_9 76.47507      NA           NA      NA      NA       NA        10
#>           sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1      0.5983679 0.3797438 0.4122175 -11.413749
#> Nb_Comp_2      0.5442372 0.3231208 0.3634169 -12.847656
#> Nb_Comp_3      0.5210941 0.3044529 0.3537087 -12.776843
#> Nb_Comp_4      0.4965569 0.2839276 0.3398355 -12.891035
#> Nb_Comp_5      0.5039885 0.3001871 0.3692997 -11.349498
#> Nb_Comp_6      0.5108963 0.3163819 0.3992388  -9.922119
#> Nb_Comp_7      0.5153766 0.3300041 0.4263658  -8.696873
#> Nb_Comp_8      0.5249910 0.3507834 0.4632727  -7.337679
#> Nb_Comp_9      0.5334234 0.3707649 0.4998004  -6.033403
#> attr(,"computed_nt")
#> [1] 9
#> 
#> [[3]]
#>                AIC      Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y      R2_Y
#> Nb_Comp_0 82.41888           NA      NA         NA       NA 20.800152        NA
#> Nb_Comp_1 63.69250    0.3365687  0.0975  0.3365687 13.79947 11.099368 0.4663804
#> Nb_Comp_2 58.35228    0.1455661  0.0975 -0.2879011 14.29489  8.885823 0.5728001
#> Nb_Comp_3 56.36553   -0.4774579  0.0975 -0.7291658 15.36506  7.874634 0.6214146
#> Nb_Comp_4 54.02416   -2.6861938  0.0975 -1.4949568 19.64687  6.903925 0.6680830
#> Nb_Comp_5 55.80450   -8.2679002  0.0975 -1.5142195 17.35798  6.858120 0.6702851
#> Nb_Comp_6 57.45753  -21.8257919  0.0975 -1.4628871 16.89078  6.786392 0.6737336
#> Nb_Comp_7 58.73951  -62.7683695  0.0975 -1.7936980 18.95913  6.640327 0.6807558
#> Nb_Comp_8 60.61227 -180.3072283  0.0975 -1.8432157 18.87988  6.614773 0.6819844
#> Nb_Comp_9 62.25948 -538.8858574  0.0975 -1.9777404 19.69708  6.544432 0.6853661
#>            AIC.std DoF.dof sigmahat.dof AIC.dof BIC.dof GMDL.dof DoF.naive
#> Nb_Comp_0 96.63448      NA           NA      NA      NA       NA         1
#> Nb_Comp_1 77.90810      NA           NA      NA      NA       NA         2
#> Nb_Comp_2 72.56787      NA           NA      NA      NA       NA         3
#> Nb_Comp_3 70.58113      NA           NA      NA      NA       NA         4
#> Nb_Comp_4 68.23976      NA           NA      NA      NA       NA         5
#> Nb_Comp_5 70.02009      NA           NA      NA      NA       NA         6
#> Nb_Comp_6 71.67313      NA           NA      NA      NA       NA         7
#> Nb_Comp_7 72.95511      NA           NA      NA      NA       NA         8
#> Nb_Comp_8 74.82787      NA           NA      NA      NA       NA         9
#> Nb_Comp_9 76.47507      NA           NA      NA      NA       NA        10
#>           sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1      0.5983679 0.3797438 0.4122175 -11.413749
#> Nb_Comp_2      0.5442372 0.3231208 0.3634169 -12.847656
#> Nb_Comp_3      0.5210941 0.3044529 0.3537087 -12.776843
#> Nb_Comp_4      0.4965569 0.2839276 0.3398355 -12.891035
#> Nb_Comp_5      0.5039885 0.3001871 0.3692997 -11.349498
#> Nb_Comp_6      0.5108963 0.3163819 0.3992388  -9.922119
#> Nb_Comp_7      0.5153766 0.3300041 0.4263658  -8.696873
#> Nb_Comp_8      0.5249910 0.3507834 0.4632727  -7.337679
#> Nb_Comp_9      0.5334234 0.3707649 0.4998004  -6.033403
#> attr(,"computed_nt")
#> [1] 9
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"

data(aze_compl)
summary(cv.plsR(y~.,data=aze_compl,nt=10,K=8,NK=3,
verbose=FALSE),MClassed=TRUE,verbose=FALSE)
#> [[1]]
#>                 AIC MissClassed CV_MissClassed      Q2cum_Y LimQ2_Y       Q2_Y
#> Nb_Comp_0  154.6179          49             NA           NA      NA         NA
#> Nb_Comp_1  126.4083          27             42   -0.1061711  0.0975 -0.1061711
#> Nb_Comp_2  119.3375          25             39   -0.5960203  0.0975 -0.4428331
#> Nb_Comp_3  114.2313          27             45   -1.8401711  0.0975 -0.7795332
#> Nb_Comp_4  112.3463          23             43   -4.9481058  0.0975 -1.0942773
#> Nb_Comp_5  113.2362          22             41  -12.0833749  0.0975 -1.1995868
#> Nb_Comp_6  114.7620          21             42  -28.5283073  0.0975 -1.2569335
#> Nb_Comp_7  116.5264          20             41  -66.5458644  0.0975 -1.2874953
#> Nb_Comp_8  118.4601          20             41 -153.7402407  0.0975 -1.2908914
#> Nb_Comp_9  120.4452          19             41 -355.4269515  0.0975 -1.3033889
#> Nb_Comp_10 122.4395          19             41 -821.2862196  0.0975 -1.3070259
#>             PRESS_Y    RSS_Y      R2_Y  AIC.std  DoF.dof sigmahat.dof   AIC.dof
#> Nb_Comp_0        NA 25.91346        NA 298.1344  1.00000    0.5015845 0.2540061
#> Nb_Comp_1  28.66472 19.38086 0.2520929 269.9248 22.55372    0.4848429 0.2883114
#> Nb_Comp_2  27.96335 17.76209 0.3145613 262.8540 27.31542    0.4781670 0.2908950
#> Nb_Comp_3  31.60823 16.58896 0.3598323 257.7478 30.52370    0.4719550 0.2902572
#> Nb_Comp_4  34.74189 15.98071 0.3833049 255.8628 34.00000    0.4744263 0.3008285
#> Nb_Comp_5  35.15095 15.81104 0.3898523 256.7527 34.00000    0.4719012 0.2976347
#> Nb_Comp_6  35.68447 15.73910 0.3926285 258.2785 34.00000    0.4708264 0.2962804
#> Nb_Comp_7  36.00312 15.70350 0.3940024 260.0429 33.71066    0.4693382 0.2937976
#> Nb_Comp_8  35.97500 15.69348 0.3943888 261.9766 34.00000    0.4701436 0.2954217
#> Nb_Comp_9  36.14819 15.69123 0.3944758 263.9617 33.87284    0.4696894 0.2945815
#> Nb_Comp_10 36.20007 15.69037 0.3945088 265.9560 34.00000    0.4700970 0.2953632
#>              BIC.dof  GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive
#> Nb_Comp_0  0.2604032 -67.17645         1      0.5015845 0.2540061 0.2604032
#> Nb_Comp_1  0.4231184 -53.56607         2      0.4358996 0.1936625 0.2033251
#> Nb_Comp_2  0.4496983 -52.42272         3      0.4193593 0.1809352 0.1943501
#> Nb_Comp_3  0.4631316 -51.93343         4      0.4072955 0.1722700 0.1891422
#> Nb_Comp_4  0.4954133 -50.37079         5      0.4017727 0.1691819 0.1897041
#> Nb_Comp_5  0.4901536 -50.65724         6      0.4016679 0.1706451 0.1952588
#> Nb_Comp_6  0.4879234 -50.78005         7      0.4028135 0.1731800 0.2020601
#> Nb_Comp_7  0.4826103 -51.05525         8      0.4044479 0.1761610 0.2094352
#> Nb_Comp_8  0.4865092 -50.85833         9      0.4064413 0.1794902 0.2172936
#> Nb_Comp_9  0.4845867 -50.95616        10      0.4085682 0.1829787 0.2254232
#> Nb_Comp_10 0.4864128 -50.86368        11      0.4107477 0.1865584 0.2337468
#>            GMDL.naive
#> Nb_Comp_0   -67.17645
#> Nb_Comp_1   -79.67755
#> Nb_Comp_2   -81.93501
#> Nb_Comp_3   -83.31503
#> Nb_Comp_4   -83.23369
#> Nb_Comp_5   -81.93513
#> Nb_Comp_6   -80.42345
#> Nb_Comp_7   -78.87607
#> Nb_Comp_8   -77.31942
#> Nb_Comp_9   -75.80069
#> Nb_Comp_10  -74.33325
#> attr(,"computed_nt")
#> [1] 10
#> 
#> [[2]]
#>                 AIC MissClassed CV_MissClassed       Q2cum_Y LimQ2_Y       Q2_Y
#> Nb_Comp_0  154.6179          49             NA            NA      NA         NA
#> Nb_Comp_1  126.4083          27             47    -0.1977984  0.0975 -0.1977984
#> Nb_Comp_2  119.3375          25             50    -0.9531630  0.0975 -0.6306276
#> Nb_Comp_3  114.2313          27             44    -2.6313264  0.0975 -0.8592029
#> Nb_Comp_4  112.3463          23             49    -6.6647559  0.0975 -1.1107318
#> Nb_Comp_5  113.2362          22             50   -16.2882937  0.0975 -1.2555570
#> Nb_Comp_6  114.7620          21             49   -38.3610775  0.0975 -1.2767474
#> Nb_Comp_7  116.5264          20             49   -88.7850764  0.0975 -1.2810625
#> Nb_Comp_8  118.4601          20             49  -205.1585036  0.0975 -1.2961333
#> Nb_Comp_9  120.4452          19             48  -473.9867210  0.0975 -1.3039880
#> Nb_Comp_10 122.4395          19             49 -1098.9905814  0.0975 -1.3158344
#>             PRESS_Y    RSS_Y      R2_Y  AIC.std  DoF.dof sigmahat.dof   AIC.dof
#> Nb_Comp_0        NA 25.91346        NA 298.1344  1.00000    0.5015845 0.2540061
#> Nb_Comp_1  31.03910 19.38086 0.2520929 269.9248 22.55372    0.4848429 0.2883114
#> Nb_Comp_2  31.60297 17.76209 0.3145613 262.8540 27.31542    0.4781670 0.2908950
#> Nb_Comp_3  33.02333 16.58896 0.3598323 257.7478 30.52370    0.4719550 0.2902572
#> Nb_Comp_4  35.01485 15.98071 0.3833049 255.8628 34.00000    0.4744263 0.3008285
#> Nb_Comp_5  36.04539 15.81104 0.3898523 256.7527 34.00000    0.4719012 0.2976347
#> Nb_Comp_6  35.99774 15.73910 0.3926285 258.2785 34.00000    0.4708264 0.2962804
#> Nb_Comp_7  35.90187 15.70350 0.3940024 260.0429 33.71066    0.4693382 0.2937976
#> Nb_Comp_8  36.05732 15.69348 0.3943888 261.9766 34.00000    0.4701436 0.2954217
#> Nb_Comp_9  36.15759 15.69123 0.3944758 263.9617 33.87284    0.4696894 0.2945815
#> Nb_Comp_10 36.33829 15.69037 0.3945088 265.9560 34.00000    0.4700970 0.2953632
#>              BIC.dof  GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive
#> Nb_Comp_0  0.2604032 -67.17645         1      0.5015845 0.2540061 0.2604032
#> Nb_Comp_1  0.4231184 -53.56607         2      0.4358996 0.1936625 0.2033251
#> Nb_Comp_2  0.4496983 -52.42272         3      0.4193593 0.1809352 0.1943501
#> Nb_Comp_3  0.4631316 -51.93343         4      0.4072955 0.1722700 0.1891422
#> Nb_Comp_4  0.4954133 -50.37079         5      0.4017727 0.1691819 0.1897041
#> Nb_Comp_5  0.4901536 -50.65724         6      0.4016679 0.1706451 0.1952588
#> Nb_Comp_6  0.4879234 -50.78005         7      0.4028135 0.1731800 0.2020601
#> Nb_Comp_7  0.4826103 -51.05525         8      0.4044479 0.1761610 0.2094352
#> Nb_Comp_8  0.4865092 -50.85833         9      0.4064413 0.1794902 0.2172936
#> Nb_Comp_9  0.4845867 -50.95616        10      0.4085682 0.1829787 0.2254232
#> Nb_Comp_10 0.4864128 -50.86368        11      0.4107477 0.1865584 0.2337468
#>            GMDL.naive
#> Nb_Comp_0   -67.17645
#> Nb_Comp_1   -79.67755
#> Nb_Comp_2   -81.93501
#> Nb_Comp_3   -83.31503
#> Nb_Comp_4   -83.23369
#> Nb_Comp_5   -81.93513
#> Nb_Comp_6   -80.42345
#> Nb_Comp_7   -78.87607
#> Nb_Comp_8   -77.31942
#> Nb_Comp_9   -75.80069
#> Nb_Comp_10  -74.33325
#> attr(,"computed_nt")
#> [1] 10
#> 
#> [[3]]
#>                 AIC MissClassed CV_MissClassed      Q2cum_Y LimQ2_Y       Q2_Y
#> Nb_Comp_0  154.6179          49             NA           NA      NA         NA
#> Nb_Comp_1  126.4083          27             46   -0.1232931  0.0975 -0.1232931
#> Nb_Comp_2  119.3375          25             48   -0.7269344  0.0975 -0.5373854
#> Nb_Comp_3  114.2313          27             45   -2.1064449  0.0975 -0.7988204
#> Nb_Comp_4  112.3463          23             50   -5.5416837  0.0975 -1.1058425
#> Nb_Comp_5  113.2362          22             51  -13.4470511  0.0975 -1.2084607
#> Nb_Comp_6  114.7620          21             50  -31.7742632  0.0975 -1.2685781
#> Nb_Comp_7  116.5264          20             50  -74.0957259  0.0975 -1.2913017
#> Nb_Comp_8  118.4601          20             50 -172.7995657  0.0975 -1.3143736
#> Nb_Comp_9  120.4452          19             51 -402.7194577  0.0975 -1.3229026
#> Nb_Comp_10 122.4395          19             51 -938.6834170  0.0975 -1.3275653
#>             PRESS_Y    RSS_Y      R2_Y  AIC.std  DoF.dof sigmahat.dof   AIC.dof
#> Nb_Comp_0        NA 25.91346        NA 298.1344  1.00000    0.5015845 0.2540061
#> Nb_Comp_1  29.10841 19.38086 0.2520929 269.9248 22.55372    0.4848429 0.2883114
#> Nb_Comp_2  29.79585 17.76209 0.3145613 262.8540 27.31542    0.4781670 0.2908950
#> Nb_Comp_3  31.95081 16.58896 0.3598323 257.7478 30.52370    0.4719550 0.2902572
#> Nb_Comp_4  34.93374 15.98071 0.3833049 255.8628 34.00000    0.4744263 0.3008285
#> Nb_Comp_5  35.29276 15.81104 0.3898523 256.7527 34.00000    0.4719012 0.2976347
#> Nb_Comp_6  35.86858 15.73910 0.3926285 258.2785 34.00000    0.4708264 0.2962804
#> Nb_Comp_7  36.06303 15.70350 0.3940024 260.0429 33.71066    0.4693382 0.2937976
#> Nb_Comp_8  36.34376 15.69348 0.3943888 261.9766 34.00000    0.4701436 0.2954217
#> Nb_Comp_9  36.45443 15.69123 0.3944758 263.9617 33.87284    0.4696894 0.2945815
#> Nb_Comp_10 36.52236 15.69037 0.3945088 265.9560 34.00000    0.4700970 0.2953632
#>              BIC.dof  GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive
#> Nb_Comp_0  0.2604032 -67.17645         1      0.5015845 0.2540061 0.2604032
#> Nb_Comp_1  0.4231184 -53.56607         2      0.4358996 0.1936625 0.2033251
#> Nb_Comp_2  0.4496983 -52.42272         3      0.4193593 0.1809352 0.1943501
#> Nb_Comp_3  0.4631316 -51.93343         4      0.4072955 0.1722700 0.1891422
#> Nb_Comp_4  0.4954133 -50.37079         5      0.4017727 0.1691819 0.1897041
#> Nb_Comp_5  0.4901536 -50.65724         6      0.4016679 0.1706451 0.1952588
#> Nb_Comp_6  0.4879234 -50.78005         7      0.4028135 0.1731800 0.2020601
#> Nb_Comp_7  0.4826103 -51.05525         8      0.4044479 0.1761610 0.2094352
#> Nb_Comp_8  0.4865092 -50.85833         9      0.4064413 0.1794902 0.2172936
#> Nb_Comp_9  0.4845867 -50.95616        10      0.4085682 0.1829787 0.2254232
#> Nb_Comp_10 0.4864128 -50.86368        11      0.4107477 0.1865584 0.2337468
#>            GMDL.naive
#> Nb_Comp_0   -67.17645
#> Nb_Comp_1   -79.67755
#> Nb_Comp_2   -81.93501
#> Nb_Comp_3   -83.31503
#> Nb_Comp_4   -83.23369
#> Nb_Comp_5   -81.93513
#> Nb_Comp_6   -80.42345
#> Nb_Comp_7   -78.87607
#> Nb_Comp_8   -77.31942
#> Nb_Comp_9   -75.80069
#> Nb_Comp_10  -74.33325
#> attr(,"computed_nt")
#> [1] 10
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"
# }