Extracts and computes information criteria and fits statistics for k-fold cross validated partial least squares models

This function extracts and computes information criteria and fits statistics for k-fold cross validated partial least squares models for both formula or classic specifications of the model.

kfolds2CVinfos_lm(pls_kfolds, MClassed = FALSE, verbose = TRUE)

Arguments

pls_kfolds
an object computed using PLS_lm_kfoldcv

MClassed
should number of miss classed be computed

verbose
should infos be displayed ?

Value

list

table of fit statistics for first group partition

list()

...

list

table of fit statistics for last group partition

Details

The Mclassed option should only set to TRUE if the response is binary.

Note

Use summary and cv.plsR instead.

References

Nicolas Meyer, Myriam Maumy-Bertrand et Frédéric Bertrand (2010). Comparing the linear and the logistic PLS regression with qualitative predictors: application to allelotyping data. Journal de la Societe Francaise de Statistique, 151(2), pages 1-18. http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/47

See also

Author

Frédéric Bertrand
frederic.bertrand@utt.fr
https://fbertran.github.io/homepage/

Examples

# \donttest{
data(Cornell)
summary(cv.plsR(Y~.,data=Cornell,nt=10,K=6,verbose=FALSE))

#> ____************************************************____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> Warning : 1 2 3 4 5 6 7 < 10^{-12}
#> Warning only 6 components could thus be extracted
#> ____Predicting X without NA neither in X nor in Y____
#> ****________________________________________________****
#> 
#> 
#> NK: 1

#> [[1]]
#>                AIC    Q2cum_Y LimQ2_Y        Q2_Y  PRESS_Y      RSS_Y      R2_Y
#> Nb_Comp_0 82.01205         NA      NA          NA       NA 467.796667        NA
#> Nb_Comp_1 53.15173  0.8931139  0.0975  0.89311387 50.00098  35.742486 0.9235940
#> Nb_Comp_2 41.08283  0.8827013  0.0975 -0.09741769 39.22444  11.066606 0.9763431
#> Nb_Comp_3 32.06411  0.7545436  0.0975 -1.09257499 23.15770   4.418081 0.9905556
#> Nb_Comp_4 33.76477 -0.4218576  0.0975 -4.79270956 25.59266   4.309235 0.9907882
#> Nb_Comp_5 33.34373 -9.6400064  0.0975 -6.48317314 32.24675   3.521924 0.9924713
#> Nb_Comp_6 35.25533         NA  0.0975          NA       NA   3.496074 0.9925265
#>              AIC.std  DoF.dof sigmahat.dof    AIC.dof    BIC.dof GMDL.dof
#> Nb_Comp_0  37.010388 1.000000    6.5212706 46.0708838 47.7893514 27.59461
#> Nb_Comp_1   8.150064 2.740749    1.8665281  4.5699686  4.9558156 21.34020
#> Nb_Comp_2  -3.918831 5.085967    1.1825195  2.1075461  2.3949331 27.40202
#> Nb_Comp_3 -12.937550 5.121086    0.7488308  0.8467795  0.9628191 24.40842
#> Nb_Comp_4 -11.236891 5.103312    0.7387162  0.8232505  0.9357846 24.23105
#> Nb_Comp_5 -11.657929 6.006316    0.7096382  0.7976101  0.9198348 28.21184
#> Nb_Comp_6  -9.746328 7.000002    0.7633343  0.9711322  1.1359501 33.18348
#>           DoF.naive sigmahat.naive  AIC.naive  BIC.naive GMDL.naive
#> Nb_Comp_0         1      6.5212706 46.0708838 47.7893514   27.59461
#> Nb_Comp_1         2      1.8905683  4.1699567  4.4588195   18.37545
#> Nb_Comp_2         3      1.1088836  1.5370286  1.6860917   17.71117
#> Nb_Comp_3         4      0.7431421  0.7363469  0.8256118   19.01033
#> Nb_Comp_4         5      0.7846050  0.8721072  0.9964867   24.16510
#> Nb_Comp_5         6      0.7661509  0.8804809  1.0227979   28.64206
#> Nb_Comp_6         7      0.8361907  1.1070902  1.3048716   33.63927
#> attr(,"computed_nt")
#> [1] 6
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"

data(pine)
summary(cv.plsR(x11~.,data=pine,nt=10,NK=3,verbose=FALSE),verbose=FALSE)

#> [[1]]
#>                 AIC       Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0  82.41888            NA      NA         NA       NA 20.800152
#> Nb_Comp_1  63.61896     0.2524957  0.0975  0.2524957 15.54820 11.074659
#> Nb_Comp_2  58.47638    -0.4059117  0.0975 -0.8808077 20.82930  8.919303
#> Nb_Comp_3  56.55421    -2.3701532  0.0975 -1.3971300 21.38073  7.919786
#> Nb_Comp_4  54.35053    -7.1380325  0.0975 -1.4147367 19.12420  6.972542
#> Nb_Comp_5  55.99834   -21.3317065  0.0975 -1.7441162 19.13346  6.898523
#> Nb_Comp_6  57.69592   -61.8227477  0.0975 -1.8131638 19.40668  6.835594
#> Nb_Comp_7  59.37953  -195.6320969  0.0975 -2.1299506 21.39507  6.770369
#> Nb_Comp_8  61.21213  -569.3861000  0.0975 -1.9007782 19.63934  6.736112
#> Nb_Comp_9  63.18426 -1659.4082959  0.0975 -1.9110252 19.60899  6.730426
#> Nb_Comp_10 65.15982 -4903.9188404  0.0975 -1.9540438 19.88197  6.725443
#>                 R2_Y  AIC.std   DoF.dof sigmahat.dof   AIC.dof   BIC.dof
#> Nb_Comp_0         NA 96.63448  1.000000    0.8062287 0.6697018 0.6991787
#> Nb_Comp_1  0.4675684 77.83455  3.176360    0.5994089 0.4047616 0.4565153
#> Nb_Comp_2  0.5711905 72.69198  7.133559    0.5761829 0.4138120 0.5212090
#> Nb_Comp_3  0.6192438 70.76981  8.778329    0.5603634 0.4070516 0.5320535
#> Nb_Comp_4  0.6647841 68.56612  8.427874    0.5221703 0.3505594 0.4547689
#> Nb_Comp_5  0.6683426 70.21393  9.308247    0.5285695 0.3666578 0.4845912
#> Nb_Comp_6  0.6713681 71.91152  9.291931    0.5259794 0.3629363 0.4795121
#> Nb_Comp_7  0.6745039 73.59512  9.756305    0.5284535 0.3702885 0.4938445
#> Nb_Comp_8  0.6761508 75.42772 10.363948    0.5338475 0.3831339 0.5170783
#> Nb_Comp_9  0.6764242 77.39986 10.732136    0.5378275 0.3920954 0.5328742
#> Nb_Comp_10 0.6766638 79.37542 11.000000    0.5407500 0.3987417 0.5446065
#>             GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0  -3.605128         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1  -9.875081         2      0.5977015 0.3788984 0.4112998 -11.451340
#> Nb_Comp_2  -6.985517         3      0.5452615 0.3243383 0.3647862 -12.822703
#> Nb_Comp_3  -6.260610         4      0.5225859 0.3061986 0.3557368 -12.756838
#> Nb_Comp_4  -8.152986         5      0.4990184 0.2867496 0.3432131 -12.811575
#> Nb_Comp_5  -7.111583         6      0.5054709 0.3019556 0.3714754 -11.329638
#> Nb_Comp_6  -7.233043         7      0.5127450 0.3186757 0.4021333  -9.918688
#> Nb_Comp_7  -6.742195         8      0.5203986 0.3364668 0.4347156  -8.592770
#> Nb_Comp_8  -6.038372         9      0.5297842 0.3572181 0.4717708  -7.287834
#> Nb_Comp_9  -5.600249        10      0.5409503 0.3813021 0.5140048  -6.008747
#> Nb_Comp_10 -5.288422        11      0.5529032 0.4076026 0.5600977  -4.799453
#> attr(,"computed_nt")
#> [1] 10
#> 
#> [[2]]
#>                 AIC       Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0  82.41888            NA      NA         NA       NA 20.800152
#> Nb_Comp_1  63.61896    0.30583260  0.0975  0.3058326 14.43879 11.074659
#> Nb_Comp_2  58.47638   -0.04771339  0.0975 -0.5093094 16.71509  8.919303
#> Nb_Comp_3  56.55421   -1.08065939  0.0975 -0.9859051 17.71289  7.919786
#> Nb_Comp_4  54.35053   -3.45410702  0.0975 -1.1407190 16.95404  6.972542
#> Nb_Comp_5  55.99834   -9.58877040  0.0975 -1.3773049 16.57586  6.898523
#> Nb_Comp_6  57.69592  -23.56419460  0.0975 -1.3198345 16.00343  6.835594
#> Nb_Comp_7  59.37953  -57.04052789  0.0975 -1.3628101 16.15121  6.770369
#> Nb_Comp_8  61.21213 -145.81509386  0.0975 -1.5295272 17.12583  6.736112
#> Nb_Comp_9  63.18426 -375.06970076  0.0975 -1.5615193 17.25468  6.730426
#> Nb_Comp_10 65.15982 -968.97005847  0.0975 -1.5792295 17.35931  6.725443
#>                 R2_Y  AIC.std   DoF.dof sigmahat.dof   AIC.dof   BIC.dof
#> Nb_Comp_0         NA 96.63448  1.000000    0.8062287 0.6697018 0.6991787
#> Nb_Comp_1  0.4675684 77.83455  3.176360    0.5994089 0.4047616 0.4565153
#> Nb_Comp_2  0.5711905 72.69198  7.133559    0.5761829 0.4138120 0.5212090
#> Nb_Comp_3  0.6192438 70.76981  8.778329    0.5603634 0.4070516 0.5320535
#> Nb_Comp_4  0.6647841 68.56612  8.427874    0.5221703 0.3505594 0.4547689
#> Nb_Comp_5  0.6683426 70.21393  9.308247    0.5285695 0.3666578 0.4845912
#> Nb_Comp_6  0.6713681 71.91152  9.291931    0.5259794 0.3629363 0.4795121
#> Nb_Comp_7  0.6745039 73.59512  9.756305    0.5284535 0.3702885 0.4938445
#> Nb_Comp_8  0.6761508 75.42772 10.363948    0.5338475 0.3831339 0.5170783
#> Nb_Comp_9  0.6764242 77.39986 10.732136    0.5378275 0.3920954 0.5328742
#> Nb_Comp_10 0.6766638 79.37542 11.000000    0.5407500 0.3987417 0.5446065
#>             GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0  -3.605128         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1  -9.875081         2      0.5977015 0.3788984 0.4112998 -11.451340
#> Nb_Comp_2  -6.985517         3      0.5452615 0.3243383 0.3647862 -12.822703
#> Nb_Comp_3  -6.260610         4      0.5225859 0.3061986 0.3557368 -12.756838
#> Nb_Comp_4  -8.152986         5      0.4990184 0.2867496 0.3432131 -12.811575
#> Nb_Comp_5  -7.111583         6      0.5054709 0.3019556 0.3714754 -11.329638
#> Nb_Comp_6  -7.233043         7      0.5127450 0.3186757 0.4021333  -9.918688
#> Nb_Comp_7  -6.742195         8      0.5203986 0.3364668 0.4347156  -8.592770
#> Nb_Comp_8  -6.038372         9      0.5297842 0.3572181 0.4717708  -7.287834
#> Nb_Comp_9  -5.600249        10      0.5409503 0.3813021 0.5140048  -6.008747
#> Nb_Comp_10 -5.288422        11      0.5529032 0.4076026 0.5600977  -4.799453
#> attr(,"computed_nt")
#> [1] 10
#> 
#> [[3]]
#>                 AIC      Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0  82.41888           NA      NA         NA       NA 20.800152
#> Nb_Comp_1  63.61896    0.2836758  0.0975  0.2836758 14.89965 11.074659
#> Nb_Comp_2  58.47638    0.1687602  0.0975 -0.1604241 12.85130  8.919303
#> Nb_Comp_3  56.55421   -0.1115724  0.0975 -0.3372464 11.92731  7.919786
#> Nb_Comp_4  54.35053   -0.9543054  0.0975 -0.7581450 13.92413  6.972542
#> Nb_Comp_5  55.99834   -2.8509953  0.0975 -0.9705187 13.73952  6.898523
#> Nb_Comp_6  57.69592   -6.5680846  0.0975 -0.9652282 13.55717  6.835594
#> Nb_Comp_7  59.37953  -14.3119269  0.0975 -1.0232235 13.82993  6.770369
#> Nb_Comp_8  61.21213  -30.2195942  0.0975 -1.0389069 13.80415  6.736112
#> Nb_Comp_9  63.18426  -65.5737652  0.0975 -1.1324353 14.36432  6.730426
#> Nb_Comp_10 65.15982 -145.6150317  0.0975 -1.2022944 14.82238  6.725443
#>                 R2_Y  AIC.std   DoF.dof sigmahat.dof   AIC.dof   BIC.dof
#> Nb_Comp_0         NA 96.63448  1.000000    0.8062287 0.6697018 0.6991787
#> Nb_Comp_1  0.4675684 77.83455  3.176360    0.5994089 0.4047616 0.4565153
#> Nb_Comp_2  0.5711905 72.69198  7.133559    0.5761829 0.4138120 0.5212090
#> Nb_Comp_3  0.6192438 70.76981  8.778329    0.5603634 0.4070516 0.5320535
#> Nb_Comp_4  0.6647841 68.56612  8.427874    0.5221703 0.3505594 0.4547689
#> Nb_Comp_5  0.6683426 70.21393  9.308247    0.5285695 0.3666578 0.4845912
#> Nb_Comp_6  0.6713681 71.91152  9.291931    0.5259794 0.3629363 0.4795121
#> Nb_Comp_7  0.6745039 73.59512  9.756305    0.5284535 0.3702885 0.4938445
#> Nb_Comp_8  0.6761508 75.42772 10.363948    0.5338475 0.3831339 0.5170783
#> Nb_Comp_9  0.6764242 77.39986 10.732136    0.5378275 0.3920954 0.5328742
#> Nb_Comp_10 0.6766638 79.37542 11.000000    0.5407500 0.3987417 0.5446065
#>             GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0  -3.605128         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1  -9.875081         2      0.5977015 0.3788984 0.4112998 -11.451340
#> Nb_Comp_2  -6.985517         3      0.5452615 0.3243383 0.3647862 -12.822703
#> Nb_Comp_3  -6.260610         4      0.5225859 0.3061986 0.3557368 -12.756838
#> Nb_Comp_4  -8.152986         5      0.4990184 0.2867496 0.3432131 -12.811575
#> Nb_Comp_5  -7.111583         6      0.5054709 0.3019556 0.3714754 -11.329638
#> Nb_Comp_6  -7.233043         7      0.5127450 0.3186757 0.4021333  -9.918688
#> Nb_Comp_7  -6.742195         8      0.5203986 0.3364668 0.4347156  -8.592770
#> Nb_Comp_8  -6.038372         9      0.5297842 0.3572181 0.4717708  -7.287834
#> Nb_Comp_9  -5.600249        10      0.5409503 0.3813021 0.5140048  -6.008747
#> Nb_Comp_10 -5.288422        11      0.5529032 0.4076026 0.5600977  -4.799453
#> attr(,"computed_nt")
#> [1] 10
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"

data(pineNAX21)
summary(cv.plsR(x11~.,data=pineNAX21,nt=10,NK=3,
verbose=FALSE),verbose=FALSE)

#> [[1]]
#>                AIC       Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0 82.41888            NA      NA         NA       NA 20.800152
#> Nb_Comp_1 63.69250    0.34379557  0.0975  0.3437956 13.64915 11.099368
#> Nb_Comp_2 58.35228    0.24147514  0.0975 -0.1559277 12.83007  8.885823
#> Nb_Comp_3 56.36553   -0.07844383  0.0975 -0.4217647 12.63355  7.874634
#> Nb_Comp_4 54.02416   -0.81709069  0.0975 -0.6849192 13.26812  6.903925
#> Nb_Comp_5 55.80450   -3.84079157  0.0975 -1.6640341 18.39229  6.858120
#> Nb_Comp_6 57.45753  -12.56989896  0.0975 -1.8032397 19.22495  6.786392
#> Nb_Comp_7 58.73951  -54.58483571  0.0975 -3.0961864 27.79832  6.640327
#> Nb_Comp_8 60.61227 -130.32236880  0.0975 -1.3625575 15.68815  6.614773
#> Nb_Comp_9 62.25948 -462.81358782  0.0975 -2.5318704 23.36252  6.544432
#>                R2_Y  AIC.std DoF.dof sigmahat.dof AIC.dof BIC.dof GMDL.dof
#> Nb_Comp_0        NA 96.63448      NA           NA      NA      NA       NA
#> Nb_Comp_1 0.4663804 77.90810      NA           NA      NA      NA       NA
#> Nb_Comp_2 0.5728001 72.56787      NA           NA      NA      NA       NA
#> Nb_Comp_3 0.6214146 70.58113      NA           NA      NA      NA       NA
#> Nb_Comp_4 0.6680830 68.23976      NA           NA      NA      NA       NA
#> Nb_Comp_5 0.6702851 70.02009      NA           NA      NA      NA       NA
#> Nb_Comp_6 0.6737336 71.67313      NA           NA      NA      NA       NA
#> Nb_Comp_7 0.6807558 72.95511      NA           NA      NA      NA       NA
#> Nb_Comp_8 0.6819844 74.82787      NA           NA      NA      NA       NA
#> Nb_Comp_9 0.6853661 76.47507      NA           NA      NA      NA       NA
#>           DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1         2      0.5983679 0.3797438 0.4122175 -11.413749
#> Nb_Comp_2         3      0.5442372 0.3231208 0.3634169 -12.847656
#> Nb_Comp_3         4      0.5210941 0.3044529 0.3537087 -12.776843
#> Nb_Comp_4         5      0.4965569 0.2839276 0.3398355 -12.891035
#> Nb_Comp_5         6      0.5039885 0.3001871 0.3692997 -11.349498
#> Nb_Comp_6         7      0.5108963 0.3163819 0.3992388  -9.922119
#> Nb_Comp_7         8      0.5153766 0.3300041 0.4263658  -8.696873
#> Nb_Comp_8         9      0.5249910 0.3507834 0.4632727  -7.337679
#> Nb_Comp_9        10      0.5334234 0.3707649 0.4998004  -6.033403
#> attr(,"computed_nt")
#> [1] 9
#> 
#> [[2]]
#>                AIC       Q2cum_Y LimQ2_Y         Q2_Y   PRESS_Y     RSS_Y
#> Nb_Comp_0 82.41888            NA      NA           NA        NA 20.800152
#> Nb_Comp_1 63.69250  3.463809e-01  0.0975    0.3463809  13.59538 11.099368
#> Nb_Comp_2 58.35228  8.500069e-02  0.0975   -0.3998968  15.53797  8.885823
#> Nb_Comp_3 56.36553 -9.976388e-01  0.0975   -1.1832134  19.39965  7.874634
#> Nb_Comp_4 54.02416 -1.848667e+00  0.0975   -0.4260169  11.22936  6.903925
#> Nb_Comp_5 55.80450 -4.105698e+00  0.0975   -0.7923116  12.37398  6.858120
#> Nb_Comp_6 57.45753 -8.243631e+00  0.0975   -0.8104539  12.41631  6.786392
#> Nb_Comp_7 58.73951 -3.229306e+01  0.0975   -2.6017292  24.44274  6.640327
#> Nb_Comp_8 60.61227 -3.483847e+03  0.0975 -103.6718915 695.05559  6.614773
#> Nb_Comp_9 62.25948 -3.673683e+05  0.0975 -104.4190450 697.32303  6.544432
#>                R2_Y  AIC.std DoF.dof sigmahat.dof AIC.dof BIC.dof GMDL.dof
#> Nb_Comp_0        NA 96.63448      NA           NA      NA      NA       NA
#> Nb_Comp_1 0.4663804 77.90810      NA           NA      NA      NA       NA
#> Nb_Comp_2 0.5728001 72.56787      NA           NA      NA      NA       NA
#> Nb_Comp_3 0.6214146 70.58113      NA           NA      NA      NA       NA
#> Nb_Comp_4 0.6680830 68.23976      NA           NA      NA      NA       NA
#> Nb_Comp_5 0.6702851 70.02009      NA           NA      NA      NA       NA
#> Nb_Comp_6 0.6737336 71.67313      NA           NA      NA      NA       NA
#> Nb_Comp_7 0.6807558 72.95511      NA           NA      NA      NA       NA
#> Nb_Comp_8 0.6819844 74.82787      NA           NA      NA      NA       NA
#> Nb_Comp_9 0.6853661 76.47507      NA           NA      NA      NA       NA
#>           DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1         2      0.5983679 0.3797438 0.4122175 -11.413749
#> Nb_Comp_2         3      0.5442372 0.3231208 0.3634169 -12.847656
#> Nb_Comp_3         4      0.5210941 0.3044529 0.3537087 -12.776843
#> Nb_Comp_4         5      0.4965569 0.2839276 0.3398355 -12.891035
#> Nb_Comp_5         6      0.5039885 0.3001871 0.3692997 -11.349498
#> Nb_Comp_6         7      0.5108963 0.3163819 0.3992388  -9.922119
#> Nb_Comp_7         8      0.5153766 0.3300041 0.4263658  -8.696873
#> Nb_Comp_8         9      0.5249910 0.3507834 0.4632727  -7.337679
#> Nb_Comp_9        10      0.5334234 0.3707649 0.4998004  -6.033403
#> attr(,"computed_nt")
#> [1] 9
#> 
#> [[3]]
#>                AIC       Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0 82.41888            NA      NA         NA       NA 20.800152
#> Nb_Comp_1 63.69250  3.837847e-01  0.0975  0.3837847 12.81737 11.099368
#> Nb_Comp_2 58.35228  7.359297e-02  0.0975 -0.5033821 16.68659  8.885823
#> Nb_Comp_3 56.36553 -8.262912e-01  0.0975 -0.9713702 17.51725  7.874634
#> Nb_Comp_4 54.02416 -2.952771e+00  0.0975 -1.1643705 17.04362  6.903925
#> Nb_Comp_5 55.80450 -1.333758e+01  0.0975 -2.6272237 25.04208  6.858120
#> Nb_Comp_6 57.45753 -7.056600e+01  0.0975 -3.9914969 34.23229  6.786392
#> Nb_Comp_7 58.73951 -1.662936e+02  0.0975 -1.3376134 15.86396  6.640327
#> Nb_Comp_8 60.61227 -3.809729e+02  0.0975 -1.2832479 15.16151  6.614773
#> Nb_Comp_9 62.25948 -1.038634e+03  0.0975 -1.7217480 18.00374  6.544432
#>                R2_Y  AIC.std DoF.dof sigmahat.dof AIC.dof BIC.dof GMDL.dof
#> Nb_Comp_0        NA 96.63448      NA           NA      NA      NA       NA
#> Nb_Comp_1 0.4663804 77.90810      NA           NA      NA      NA       NA
#> Nb_Comp_2 0.5728001 72.56787      NA           NA      NA      NA       NA
#> Nb_Comp_3 0.6214146 70.58113      NA           NA      NA      NA       NA
#> Nb_Comp_4 0.6680830 68.23976      NA           NA      NA      NA       NA
#> Nb_Comp_5 0.6702851 70.02009      NA           NA      NA      NA       NA
#> Nb_Comp_6 0.6737336 71.67313      NA           NA      NA      NA       NA
#> Nb_Comp_7 0.6807558 72.95511      NA           NA      NA      NA       NA
#> Nb_Comp_8 0.6819844 74.82787      NA           NA      NA      NA       NA
#> Nb_Comp_9 0.6853661 76.47507      NA           NA      NA      NA       NA
#>           DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1         2      0.5983679 0.3797438 0.4122175 -11.413749
#> Nb_Comp_2         3      0.5442372 0.3231208 0.3634169 -12.847656
#> Nb_Comp_3         4      0.5210941 0.3044529 0.3537087 -12.776843
#> Nb_Comp_4         5      0.4965569 0.2839276 0.3398355 -12.891035
#> Nb_Comp_5         6      0.5039885 0.3001871 0.3692997 -11.349498
#> Nb_Comp_6         7      0.5108963 0.3163819 0.3992388  -9.922119
#> Nb_Comp_7         8      0.5153766 0.3300041 0.4263658  -8.696873
#> Nb_Comp_8         9      0.5249910 0.3507834 0.4632727  -7.337679
#> Nb_Comp_9        10      0.5334234 0.3707649 0.4998004  -6.033403
#> attr(,"computed_nt")
#> [1] 9
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"

data(aze_compl)
summary(cv.plsR(y~.,data=aze_compl,nt=10,K=8,NK=3,
verbose=FALSE),MClassed=TRUE,verbose=FALSE)

#> [[1]]
#>                 AIC MissClassed CV_MissClassed       Q2cum_Y LimQ2_Y
#> Nb_Comp_0  154.6179          49             NA            NA      NA
#> Nb_Comp_1  126.4083          27             45   -0.09807511  0.0975
#> Nb_Comp_2  119.3375          25             53   -0.69629366  0.0975
#> Nb_Comp_3  114.2313          27             46   -2.15584036  0.0975
#> Nb_Comp_4  112.3463          23             43   -5.66282796  0.0975
#> Nb_Comp_5  113.2362          22             41  -13.76390693  0.0975
#> Nb_Comp_6  114.7620          21             39  -32.24434771  0.0975
#> Nb_Comp_7  116.5264          20             39  -74.36539708  0.0975
#> Nb_Comp_8  118.4601          20             41 -170.43829999  0.0975
#> Nb_Comp_9  120.4452          19             41 -388.61593231  0.0975
#> Nb_Comp_10 122.4395          19             41 -884.58342556  0.0975
#>                   Q2_Y  PRESS_Y    RSS_Y      R2_Y  AIC.std  DoF.dof
#> Nb_Comp_0           NA       NA 25.91346        NA 298.1344  1.00000
#> Nb_Comp_1  -0.09807511 28.45493 19.38086 0.2520929 269.9248 22.55372
#> Nb_Comp_2  -0.54478837 29.93933 17.76209 0.3145613 262.8540 27.31542
#> Nb_Comp_3  -0.86043280 33.04517 16.58896 0.3598323 257.7478 30.52370
#> Nb_Comp_4  -1.11126901 35.02376 15.98071 0.3833049 255.8628 34.00000
#> Nb_Comp_5  -1.21586194 35.41104 15.81104 0.3898523 256.7527 34.00000
#> Nb_Comp_6  -1.25173105 35.60221 15.73910 0.3926285 258.2785 34.00000
#> Nb_Comp_7  -1.26701386 35.68076 15.70350 0.3940024 260.0429 33.71066
#> Nb_Comp_8  -1.27476145 35.72171 15.69348 0.3943888 261.9766 34.00000
#> Nb_Comp_9  -1.27263063 35.66549 15.69123 0.3944758 263.9617 33.87284
#> Nb_Comp_10 -1.27296512 35.66562 15.69037 0.3945088 265.9560 34.00000
#>            sigmahat.dof   AIC.dof   BIC.dof  GMDL.dof DoF.naive sigmahat.naive
#> Nb_Comp_0     0.5015845 0.2540061 0.2604032 -67.17645         1      0.5015845
#> Nb_Comp_1     0.4848429 0.2883114 0.4231184 -53.56607         2      0.4358996
#> Nb_Comp_2     0.4781670 0.2908950 0.4496983 -52.42272         3      0.4193593
#> Nb_Comp_3     0.4719550 0.2902572 0.4631316 -51.93343         4      0.4072955
#> Nb_Comp_4     0.4744263 0.3008285 0.4954133 -50.37079         5      0.4017727
#> Nb_Comp_5     0.4719012 0.2976347 0.4901536 -50.65724         6      0.4016679
#> Nb_Comp_6     0.4708264 0.2962804 0.4879234 -50.78005         7      0.4028135
#> Nb_Comp_7     0.4693382 0.2937976 0.4826103 -51.05525         8      0.4044479
#> Nb_Comp_8     0.4701436 0.2954217 0.4865092 -50.85833         9      0.4064413
#> Nb_Comp_9     0.4696894 0.2945815 0.4845867 -50.95616        10      0.4085682
#> Nb_Comp_10    0.4700970 0.2953632 0.4864128 -50.86368        11      0.4107477
#>            AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0  0.2540061 0.2604032  -67.17645
#> Nb_Comp_1  0.1936625 0.2033251  -79.67755
#> Nb_Comp_2  0.1809352 0.1943501  -81.93501
#> Nb_Comp_3  0.1722700 0.1891422  -83.31503
#> Nb_Comp_4  0.1691819 0.1897041  -83.23369
#> Nb_Comp_5  0.1706451 0.1952588  -81.93513
#> Nb_Comp_6  0.1731800 0.2020601  -80.42345
#> Nb_Comp_7  0.1761610 0.2094352  -78.87607
#> Nb_Comp_8  0.1794902 0.2172936  -77.31942
#> Nb_Comp_9  0.1829787 0.2254232  -75.80069
#> Nb_Comp_10 0.1865584 0.2337468  -74.33325
#> attr(,"computed_nt")
#> [1] 10
#> 
#> [[2]]
#>                 AIC MissClassed CV_MissClassed       Q2cum_Y LimQ2_Y       Q2_Y
#> Nb_Comp_0  154.6179          49             NA            NA      NA         NA
#> Nb_Comp_1  126.4083          27             46    -0.1665311  0.0975 -0.1665311
#> Nb_Comp_2  119.3375          25             53    -0.8447623  0.0975 -0.5814085
#> Nb_Comp_3  114.2313          27             46    -2.4669664  0.0975 -0.8793567
#> Nb_Comp_4  112.3463          23             48    -6.7118845  0.0975 -1.2243898
#> Nb_Comp_5  113.2362          22             47   -16.9681124  0.0975 -1.3299250
#> Nb_Comp_6  114.7620          21             46   -42.2260383  0.0975 -1.4057084
#> Nb_Comp_7  116.5264          20             46  -104.7676681  0.0975 -1.4468508
#> Nb_Comp_8  118.4601          20             48  -259.3463269  0.0975 -1.4614925
#> Nb_Comp_9  120.4452          19             49  -641.1264847  0.0975 -1.4664319
#> Nb_Comp_10 122.4395          19             49 -1580.3337088  0.0975 -1.4626514
#>             PRESS_Y    RSS_Y      R2_Y  AIC.std  DoF.dof sigmahat.dof   AIC.dof
#> Nb_Comp_0        NA 25.91346        NA 298.1344  1.00000    0.5015845 0.2540061
#> Nb_Comp_1  30.22886 19.38086 0.2520929 269.9248 22.55372    0.4848429 0.2883114
#> Nb_Comp_2  30.64906 17.76209 0.3145613 262.8540 27.31542    0.4781670 0.2908950
#> Nb_Comp_3  33.38130 16.58896 0.3598323 257.7478 30.52370    0.4719550 0.2902572
#> Nb_Comp_4  36.90032 15.98071 0.3833049 255.8628 34.00000    0.4744263 0.3008285
#> Nb_Comp_5  37.23384 15.81104 0.3898523 256.7527 34.00000    0.4719012 0.2976347
#> Nb_Comp_6  38.03675 15.73910 0.3926285 258.2785 34.00000    0.4708264 0.2962804
#> Nb_Comp_7  38.51123 15.70350 0.3940024 260.0429 33.71066    0.4693382 0.2937976
#> Nb_Comp_8  38.65404 15.69348 0.3943888 261.9766 34.00000    0.4701436 0.2954217
#> Nb_Comp_9  38.70690 15.69123 0.3944758 263.9617 33.87284    0.4696894 0.2945815
#> Nb_Comp_10 38.64203 15.69037 0.3945088 265.9560 34.00000    0.4700970 0.2953632
#>              BIC.dof  GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive
#> Nb_Comp_0  0.2604032 -67.17645         1      0.5015845 0.2540061 0.2604032
#> Nb_Comp_1  0.4231184 -53.56607         2      0.4358996 0.1936625 0.2033251
#> Nb_Comp_2  0.4496983 -52.42272         3      0.4193593 0.1809352 0.1943501
#> Nb_Comp_3  0.4631316 -51.93343         4      0.4072955 0.1722700 0.1891422
#> Nb_Comp_4  0.4954133 -50.37079         5      0.4017727 0.1691819 0.1897041
#> Nb_Comp_5  0.4901536 -50.65724         6      0.4016679 0.1706451 0.1952588
#> Nb_Comp_6  0.4879234 -50.78005         7      0.4028135 0.1731800 0.2020601
#> Nb_Comp_7  0.4826103 -51.05525         8      0.4044479 0.1761610 0.2094352
#> Nb_Comp_8  0.4865092 -50.85833         9      0.4064413 0.1794902 0.2172936
#> Nb_Comp_9  0.4845867 -50.95616        10      0.4085682 0.1829787 0.2254232
#> Nb_Comp_10 0.4864128 -50.86368        11      0.4107477 0.1865584 0.2337468
#>            GMDL.naive
#> Nb_Comp_0   -67.17645
#> Nb_Comp_1   -79.67755
#> Nb_Comp_2   -81.93501
#> Nb_Comp_3   -83.31503
#> Nb_Comp_4   -83.23369
#> Nb_Comp_5   -81.93513
#> Nb_Comp_6   -80.42345
#> Nb_Comp_7   -78.87607
#> Nb_Comp_8   -77.31942
#> Nb_Comp_9   -75.80069
#> Nb_Comp_10  -74.33325
#> attr(,"computed_nt")
#> [1] 10
#> 
#> [[3]]
#>                 AIC MissClassed CV_MissClassed       Q2cum_Y LimQ2_Y       Q2_Y
#> Nb_Comp_0  154.6179          49             NA            NA      NA         NA
#> Nb_Comp_1  126.4083          27             51    -0.1958246  0.0975 -0.1958246
#> Nb_Comp_2  119.3375          25             51    -0.8933818  0.0975 -0.5833274
#> Nb_Comp_3  114.2313          27             48    -2.7197247  0.0975 -0.9645930
#> Nb_Comp_4  112.3463          23             49    -7.6463722  0.0975 -1.3244657
#> Nb_Comp_5  113.2362          22             48   -20.8559807  0.0975 -1.5277631
#> Nb_Comp_6  114.7620          21             50   -55.7688361  0.0975 -1.5974051
#> Nb_Comp_7  116.5264          20             49  -147.3906228  0.0975 -1.6139451
#> Nb_Comp_8  118.4601          20             51  -390.2551359  0.0975 -1.6366567
#> Nb_Comp_9  120.4452          19             51 -1033.0680149  0.0975 -1.6429506
#> Nb_Comp_10 122.4395          19             51 -2736.1990844  0.0975 -1.6470204
#>             PRESS_Y    RSS_Y      R2_Y  AIC.std  DoF.dof sigmahat.dof   AIC.dof
#> Nb_Comp_0        NA 25.91346        NA 298.1344  1.00000    0.5015845 0.2540061
#> Nb_Comp_1  30.98795 19.38086 0.2520929 269.9248 22.55372    0.4848429 0.2883114
#> Nb_Comp_2  30.68625 17.76209 0.3145613 262.8540 27.31542    0.4781670 0.2908950
#> Nb_Comp_3  34.89527 16.58896 0.3598323 257.7478 30.52370    0.4719550 0.2902572
#> Nb_Comp_4  38.56047 15.98071 0.3833049 255.8628 34.00000    0.4744263 0.3008285
#> Nb_Comp_5  40.39544 15.81104 0.3898523 256.7527 34.00000    0.4719012 0.2976347
#> Nb_Comp_6  41.06768 15.73910 0.3926285 258.2785 34.00000    0.4708264 0.2962804
#> Nb_Comp_7  41.14114 15.70350 0.3940024 260.0429 33.71066    0.4693382 0.2937976
#> Nb_Comp_8  41.40473 15.69348 0.3943888 261.9766 34.00000    0.4701436 0.2954217
#> Nb_Comp_9  41.47710 15.69123 0.3944758 263.9617 33.87284    0.4696894 0.2945815
#> Nb_Comp_10 41.53500 15.69037 0.3945088 265.9560 34.00000    0.4700970 0.2953632
#>              BIC.dof  GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive
#> Nb_Comp_0  0.2604032 -67.17645         1      0.5015845 0.2540061 0.2604032
#> Nb_Comp_1  0.4231184 -53.56607         2      0.4358996 0.1936625 0.2033251
#> Nb_Comp_2  0.4496983 -52.42272         3      0.4193593 0.1809352 0.1943501
#> Nb_Comp_3  0.4631316 -51.93343         4      0.4072955 0.1722700 0.1891422
#> Nb_Comp_4  0.4954133 -50.37079         5      0.4017727 0.1691819 0.1897041
#> Nb_Comp_5  0.4901536 -50.65724         6      0.4016679 0.1706451 0.1952588
#> Nb_Comp_6  0.4879234 -50.78005         7      0.4028135 0.1731800 0.2020601
#> Nb_Comp_7  0.4826103 -51.05525         8      0.4044479 0.1761610 0.2094352
#> Nb_Comp_8  0.4865092 -50.85833         9      0.4064413 0.1794902 0.2172936
#> Nb_Comp_9  0.4845867 -50.95616        10      0.4085682 0.1829787 0.2254232
#> Nb_Comp_10 0.4864128 -50.86368        11      0.4107477 0.1865584 0.2337468
#>            GMDL.naive
#> Nb_Comp_0   -67.17645
#> Nb_Comp_1   -79.67755
#> Nb_Comp_2   -81.93501
#> Nb_Comp_3   -83.31503
#> Nb_Comp_4   -83.23369
#> Nb_Comp_5   -81.93513
#> Nb_Comp_6   -80.42345
#> Nb_Comp_7   -78.87607
#> Nb_Comp_8   -77.31942
#> Nb_Comp_9   -75.80069
#> Nb_Comp_10  -74.33325
#> attr(,"computed_nt")
#> [1] 10
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"

# }