Extracts and computes information criteria and fits statistics for k-fold cross validated partial least squares models

This function extracts and computes information criteria and fits statistics for k-fold cross validated partial least squares models for both formula or classic specifications of the model.

kfolds2CVinfos_lm(pls_kfolds, MClassed = FALSE,verbose=TRUE)

Arguments

pls_kfolds
an object computed using PLS_lm_kfoldcv

MClassed
should number of miss classed be computed

verbose
should infos be displayed ?

Details

The Mclassed option should only set to TRUE if the response is binary.

Value

list

table of fit statistics for first group partition

…

…

list

table of fit statistics for last group partition

References

Nicolas Meyer, Myriam Maumy-Bertrand et Frederic Bertrand (2010). Comparing the linear and the logistic PLS regression with qualitative predictors: application to allelotyping data. Journal de la Societe Francaise de Statistique, 151(2), pages 1-18. http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/47

Note

Use summary and cv.plsR instead.

See also

Examples

data(Cornell)
summary(cv.plsR(Y~.,data=Cornell,nt=10,K=6,verbose=FALSE))

#> ____************************************************____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> Warning : 1 2 3 4 5 6 7 < 10^{-12}
#> Warning only 6 components could thus be extracted
#> ____Predicting X without NA neither in X nor in Y____
#> ****________________________________________________****
#> 
#> 
#> NK: 1

#> [[1]]
#>                AIC    Q2cum_Y LimQ2_Y        Q2_Y  PRESS_Y      RSS_Y      R2_Y
#> Nb_Comp_0 82.01205         NA      NA          NA       NA 467.796667        NA
#> Nb_Comp_1 53.15173  0.8818336  0.0975  0.88183361 55.27784  35.742486 0.9235940
#> Nb_Comp_2 41.08283  0.8919577  0.0975  0.08567656 32.68019  11.066606 0.9763431
#> Nb_Comp_3 32.06411  0.7578009  0.0975 -1.24170621 24.80808   4.418081 0.9905556
#> Nb_Comp_4 33.76477 -0.2596372  0.0975 -4.20083388 22.97771   4.309235 0.9907882
#> Nb_Comp_5 33.34373 -7.2527267  0.0975 -5.55166934 28.23269   3.521924 0.9924713
#> Nb_Comp_6 35.25533         NA  0.0975          NA       NA   3.496074 0.9925265
#>              AIC.std  DoF.dof sigmahat.dof    AIC.dof    BIC.dof GMDL.dof
#> Nb_Comp_0  37.010388 1.000000    6.5212706 46.0708838 47.7893514 27.59461
#> Nb_Comp_1   8.150064 2.740749    1.8665281  4.5699686  4.9558156 21.34020
#> Nb_Comp_2  -3.918831 5.085967    1.1825195  2.1075461  2.3949331 27.40202
#> Nb_Comp_3 -12.937550 5.121086    0.7488308  0.8467795  0.9628191 24.40842
#> Nb_Comp_4 -11.236891 5.103312    0.7387162  0.8232505  0.9357846 24.23105
#> Nb_Comp_5 -11.657929 6.006316    0.7096382  0.7976101  0.9198348 28.21184
#> Nb_Comp_6  -9.746328 7.000002    0.7633343  0.9711322  1.1359501 33.18348
#>           DoF.naive sigmahat.naive  AIC.naive  BIC.naive GMDL.naive
#> Nb_Comp_0         1      6.5212706 46.0708838 47.7893514   27.59461
#> Nb_Comp_1         2      1.8905683  4.1699567  4.4588195   18.37545
#> Nb_Comp_2         3      1.1088836  1.5370286  1.6860917   17.71117
#> Nb_Comp_3         4      0.7431421  0.7363469  0.8256118   19.01033
#> Nb_Comp_4         5      0.7846050  0.8721072  0.9964867   24.16510
#> Nb_Comp_5         6      0.7661509  0.8804809  1.0227979   28.64206
#> Nb_Comp_6         7      0.8361907  1.1070902  1.3048716   33.63927
#> attr(,"computed_nt")
#> [1] 6
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"

data(pine)
summary(cv.plsR(x11~.,data=pine,nt=10,NK=3,verbose=FALSE),verbose=FALSE)

#> [[1]]
#>                 AIC       Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0  82.41888            NA      NA         NA       NA 20.800152
#> Nb_Comp_1  63.61896     0.3615485  0.0975  0.3615485 13.27989 11.074659
#> Nb_Comp_2  58.47638     0.1841549  0.0975 -0.2778498 14.15175  8.919303
#> Nb_Comp_3  56.55421    -0.3160025  0.0975 -0.6130544 14.38732  7.919786
#> Nb_Comp_4  54.35053    -1.3548061  0.0975 -0.7893630 14.17137  6.972542
#> Nb_Comp_5  55.99834    -4.4766268  0.0975 -1.3257230 16.21620  6.898523
#> Nb_Comp_6  57.69592   -12.5243519  0.0975 -1.4694675 17.03568  6.835594
#> Nb_Comp_7  59.37953   -34.6710608  0.0975 -1.6375431 18.02917  6.770369
#> Nb_Comp_8  61.21213  -102.6774097  0.0975 -1.9064852 19.67798  6.736112
#> Nb_Comp_9  63.18426  -403.9544265  0.0975 -2.9059080 26.31063  6.730426
#> Nb_Comp_10 65.15982 -1566.4076536  0.0975 -2.8705779 26.05064  6.725443
#>                 R2_Y  AIC.std   DoF.dof sigmahat.dof   AIC.dof   BIC.dof
#> Nb_Comp_0         NA 96.63448  1.000000    0.8062287 0.6697018 0.6991787
#> Nb_Comp_1  0.4675684 77.83455  3.176360    0.5994089 0.4047616 0.4565153
#> Nb_Comp_2  0.5711905 72.69198  7.133559    0.5761829 0.4138120 0.5212090
#> Nb_Comp_3  0.6192438 70.76981  8.778329    0.5603634 0.4070516 0.5320535
#> Nb_Comp_4  0.6647841 68.56612  8.427874    0.5221703 0.3505594 0.4547689
#> Nb_Comp_5  0.6683426 70.21393  9.308247    0.5285695 0.3666578 0.4845912
#> Nb_Comp_6  0.6713681 71.91152  9.291931    0.5259794 0.3629363 0.4795121
#> Nb_Comp_7  0.6745039 73.59512  9.756305    0.5284535 0.3702885 0.4938445
#> Nb_Comp_8  0.6761508 75.42772 10.363948    0.5338475 0.3831339 0.5170783
#> Nb_Comp_9  0.6764242 77.39986 10.732136    0.5378275 0.3920954 0.5328742
#> Nb_Comp_10 0.6766638 79.37542 11.000000    0.5407500 0.3987417 0.5446065
#>             GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0  -3.605128         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1  -9.875081         2      0.5977015 0.3788984 0.4112998 -11.451340
#> Nb_Comp_2  -6.985517         3      0.5452615 0.3243383 0.3647862 -12.822703
#> Nb_Comp_3  -6.260610         4      0.5225859 0.3061986 0.3557368 -12.756838
#> Nb_Comp_4  -8.152986         5      0.4990184 0.2867496 0.3432131 -12.811575
#> Nb_Comp_5  -7.111583         6      0.5054709 0.3019556 0.3714754 -11.329638
#> Nb_Comp_6  -7.233043         7      0.5127450 0.3186757 0.4021333  -9.918688
#> Nb_Comp_7  -6.742195         8      0.5203986 0.3364668 0.4347156  -8.592770
#> Nb_Comp_8  -6.038372         9      0.5297842 0.3572181 0.4717708  -7.287834
#> Nb_Comp_9  -5.600249        10      0.5409503 0.3813021 0.5140048  -6.008747
#> Nb_Comp_10 -5.288422        11      0.5529032 0.4076026 0.5600977  -4.799453
#> attr(,"computed_nt")
#> [1] 10
#> 
#> [[2]]
#>                 AIC      Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0  82.41888           NA      NA         NA       NA 20.800152
#> Nb_Comp_1  63.61896    0.3785200  0.0975  0.3785200 12.92688 11.074659
#> Nb_Comp_2  58.47638    0.1220639  0.0975 -0.4126538 15.64466  8.919303
#> Nb_Comp_3  56.55421   -0.7206650  0.0975 -0.9598978 17.48092  7.919786
#> Nb_Comp_4  54.35053   -2.3915322  0.0975 -0.9710589 15.61036  6.972542
#> Nb_Comp_5  55.99834   -6.1233545  0.0975 -1.1003352 14.64468  6.898523
#> Nb_Comp_6  57.69592  -14.6877305  0.0975 -1.2022953 15.19259  6.835594
#> Nb_Comp_7  59.37953  -33.6696044  0.0975 -1.2099821 15.10654  6.770369
#> Nb_Comp_8  61.21213  -87.1141173  0.0975 -1.5415380 17.20715  6.736112
#> Nb_Comp_9  63.18426 -238.6745951  0.0975 -1.7200476 18.32255  6.730426
#> Nb_Comp_10 65.15982 -655.1503006  0.0975 -1.7376715 18.42570  6.725443
#>                 R2_Y  AIC.std   DoF.dof sigmahat.dof   AIC.dof   BIC.dof
#> Nb_Comp_0         NA 96.63448  1.000000    0.8062287 0.6697018 0.6991787
#> Nb_Comp_1  0.4675684 77.83455  3.176360    0.5994089 0.4047616 0.4565153
#> Nb_Comp_2  0.5711905 72.69198  7.133559    0.5761829 0.4138120 0.5212090
#> Nb_Comp_3  0.6192438 70.76981  8.778329    0.5603634 0.4070516 0.5320535
#> Nb_Comp_4  0.6647841 68.56612  8.427874    0.5221703 0.3505594 0.4547689
#> Nb_Comp_5  0.6683426 70.21393  9.308247    0.5285695 0.3666578 0.4845912
#> Nb_Comp_6  0.6713681 71.91152  9.291931    0.5259794 0.3629363 0.4795121
#> Nb_Comp_7  0.6745039 73.59512  9.756305    0.5284535 0.3702885 0.4938445
#> Nb_Comp_8  0.6761508 75.42772 10.363948    0.5338475 0.3831339 0.5170783
#> Nb_Comp_9  0.6764242 77.39986 10.732136    0.5378275 0.3920954 0.5328742
#> Nb_Comp_10 0.6766638 79.37542 11.000000    0.5407500 0.3987417 0.5446065
#>             GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0  -3.605128         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1  -9.875081         2      0.5977015 0.3788984 0.4112998 -11.451340
#> Nb_Comp_2  -6.985517         3      0.5452615 0.3243383 0.3647862 -12.822703
#> Nb_Comp_3  -6.260610         4      0.5225859 0.3061986 0.3557368 -12.756838
#> Nb_Comp_4  -8.152986         5      0.4990184 0.2867496 0.3432131 -12.811575
#> Nb_Comp_5  -7.111583         6      0.5054709 0.3019556 0.3714754 -11.329638
#> Nb_Comp_6  -7.233043         7      0.5127450 0.3186757 0.4021333  -9.918688
#> Nb_Comp_7  -6.742195         8      0.5203986 0.3364668 0.4347156  -8.592770
#> Nb_Comp_8  -6.038372         9      0.5297842 0.3572181 0.4717708  -7.287834
#> Nb_Comp_9  -5.600249        10      0.5409503 0.3813021 0.5140048  -6.008747
#> Nb_Comp_10 -5.288422        11      0.5529032 0.4076026 0.5600977  -4.799453
#> attr(,"computed_nt")
#> [1] 10
#> 
#> [[3]]
#>                 AIC       Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0  82.41888            NA      NA         NA       NA 20.800152
#> Nb_Comp_1  63.61896     0.3334595  0.0975  0.3334595 13.86414 11.074659
#> Nb_Comp_2  58.47638     0.1741522  0.0975 -0.2390062 13.72157  8.919303
#> Nb_Comp_3  56.55421    -0.5819818  0.0975 -0.9155852 17.08569  7.919786
#> Nb_Comp_4  54.35053    -2.4150262  0.0975 -1.1587013 17.09645  6.972542
#> Nb_Comp_5  55.99834    -7.8669928  0.0975 -1.5964640 18.10395  6.898523
#> Nb_Comp_6  57.69592   -22.8627880  0.0975 -1.6911929 18.56526  6.835594
#> Nb_Comp_7  59.37953   -62.4972444  0.0975 -1.6609315 18.18905  6.770369
#> Nb_Comp_8  61.21213  -173.1156818  0.0975 -1.7420982 18.56502  6.736112
#> Nb_Comp_9  63.18426  -474.9278497  0.0975 -1.7334003 18.41249  6.730426
#> Nb_Comp_10 65.15982 -1321.4542586  0.0975 -1.7786864 18.70174  6.725443
#>                 R2_Y  AIC.std   DoF.dof sigmahat.dof   AIC.dof   BIC.dof
#> Nb_Comp_0         NA 96.63448  1.000000    0.8062287 0.6697018 0.6991787
#> Nb_Comp_1  0.4675684 77.83455  3.176360    0.5994089 0.4047616 0.4565153
#> Nb_Comp_2  0.5711905 72.69198  7.133559    0.5761829 0.4138120 0.5212090
#> Nb_Comp_3  0.6192438 70.76981  8.778329    0.5603634 0.4070516 0.5320535
#> Nb_Comp_4  0.6647841 68.56612  8.427874    0.5221703 0.3505594 0.4547689
#> Nb_Comp_5  0.6683426 70.21393  9.308247    0.5285695 0.3666578 0.4845912
#> Nb_Comp_6  0.6713681 71.91152  9.291931    0.5259794 0.3629363 0.4795121
#> Nb_Comp_7  0.6745039 73.59512  9.756305    0.5284535 0.3702885 0.4938445
#> Nb_Comp_8  0.6761508 75.42772 10.363948    0.5338475 0.3831339 0.5170783
#> Nb_Comp_9  0.6764242 77.39986 10.732136    0.5378275 0.3920954 0.5328742
#> Nb_Comp_10 0.6766638 79.37542 11.000000    0.5407500 0.3987417 0.5446065
#>             GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0  -3.605128         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1  -9.875081         2      0.5977015 0.3788984 0.4112998 -11.451340
#> Nb_Comp_2  -6.985517         3      0.5452615 0.3243383 0.3647862 -12.822703
#> Nb_Comp_3  -6.260610         4      0.5225859 0.3061986 0.3557368 -12.756838
#> Nb_Comp_4  -8.152986         5      0.4990184 0.2867496 0.3432131 -12.811575
#> Nb_Comp_5  -7.111583         6      0.5054709 0.3019556 0.3714754 -11.329638
#> Nb_Comp_6  -7.233043         7      0.5127450 0.3186757 0.4021333  -9.918688
#> Nb_Comp_7  -6.742195         8      0.5203986 0.3364668 0.4347156  -8.592770
#> Nb_Comp_8  -6.038372         9      0.5297842 0.3572181 0.4717708  -7.287834
#> Nb_Comp_9  -5.600249        10      0.5409503 0.3813021 0.5140048  -6.008747
#> Nb_Comp_10 -5.288422        11      0.5529032 0.4076026 0.5600977  -4.799453
#> attr(,"computed_nt")
#> [1] 10
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"

data(pineNAX21)
summary(cv.plsR(x11~.,data=pineNAX21,nt=10,NK=3,
verbose=FALSE),verbose=FALSE)

#> [[1]]
#>                AIC      Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y      R2_Y
#> Nb_Comp_0 82.41888           NA      NA         NA       NA 20.800152        NA
#> Nb_Comp_1 63.69250    0.3683553  0.0975  0.3683553 13.13830 11.099368 0.4663804
#> Nb_Comp_2 58.35228    0.1079989  0.0975 -0.4121882 15.67440  8.885823 0.5728001
#> Nb_Comp_3 56.36553   -0.6527359  0.0975 -0.8528406 16.46401  7.874634 0.6214146
#> Nb_Comp_4 54.02416   -2.1772273  0.0975 -0.9224047 15.13823  6.903925 0.6680830
#> Nb_Comp_5 55.80450   -6.9077142  0.0975 -1.4888727 17.18299  6.858120 0.6702851
#> Nb_Comp_6 57.45753  -16.9266752  0.0975 -1.2669857 15.54726  6.786392 0.6737336
#> Nb_Comp_7 58.73951  -37.1511506  0.0975 -1.1281777 14.44265  6.640327 0.6807558
#> Nb_Comp_8 60.61227  -94.8860766  0.0975 -1.5133207 16.68927  6.614773 0.6819844
#> Nb_Comp_9 62.25948 -277.1402136  0.0975 -1.9007362 19.18771  6.544432 0.6853661
#>            AIC.std DoF.dof sigmahat.dof AIC.dof BIC.dof GMDL.dof DoF.naive
#> Nb_Comp_0 96.63448      NA           NA      NA      NA       NA         1
#> Nb_Comp_1 77.90810      NA           NA      NA      NA       NA         2
#> Nb_Comp_2 72.56787      NA           NA      NA      NA       NA         3
#> Nb_Comp_3 70.58113      NA           NA      NA      NA       NA         4
#> Nb_Comp_4 68.23976      NA           NA      NA      NA       NA         5
#> Nb_Comp_5 70.02009      NA           NA      NA      NA       NA         6
#> Nb_Comp_6 71.67313      NA           NA      NA      NA       NA         7
#> Nb_Comp_7 72.95511      NA           NA      NA      NA       NA         8
#> Nb_Comp_8 74.82787      NA           NA      NA      NA       NA         9
#> Nb_Comp_9 76.47507      NA           NA      NA      NA       NA        10
#>           sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1      0.5983679 0.3797438 0.4122175 -11.413749
#> Nb_Comp_2      0.5442372 0.3231208 0.3634169 -12.847656
#> Nb_Comp_3      0.5210941 0.3044529 0.3537087 -12.776843
#> Nb_Comp_4      0.4965569 0.2839276 0.3398355 -12.891035
#> Nb_Comp_5      0.5039885 0.3001871 0.3692997 -11.349498
#> Nb_Comp_6      0.5108963 0.3163819 0.3992388  -9.922119
#> Nb_Comp_7      0.5153766 0.3300041 0.4263658  -8.696873
#> Nb_Comp_8      0.5249910 0.3507834 0.4632727  -7.337679
#> Nb_Comp_9      0.5334234 0.3707649 0.4998004  -6.033403
#> attr(,"computed_nt")
#> [1] 9
#> 
#> [[2]]
#>                AIC       Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0 82.41888            NA      NA         NA       NA 20.800152
#> Nb_Comp_1 63.69250  3.220016e-01  0.0975  0.3220016 14.10247 11.099368
#> Nb_Comp_2 58.35228  6.405947e-02  0.0975 -0.3804465 15.32208  8.885823
#> Nb_Comp_3 56.36553 -1.113752e+00  0.0975 -1.2584252 20.06797  7.874634
#> Nb_Comp_4 54.02416 -3.457788e+00  0.0975 -1.1089460 16.60718  6.903925
#> Nb_Comp_5 55.80450 -1.068425e+01  0.0975 -1.6210875 18.09579  6.858120
#> Nb_Comp_6 57.45753 -3.734563e+01  0.0975 -2.2818210 22.50712  6.786392
#> Nb_Comp_7 58.73951 -1.275108e+02  0.0975 -2.3513818 22.74379  6.640327
#> Nb_Comp_8 60.61227 -4.912283e+02  0.0975 -2.8302473 25.43409  6.614773
#> Nb_Comp_9 62.25948 -2.206764e+03  0.0975 -3.4852436 29.66887  6.544432
#>                R2_Y  AIC.std DoF.dof sigmahat.dof AIC.dof BIC.dof GMDL.dof
#> Nb_Comp_0        NA 96.63448      NA           NA      NA      NA       NA
#> Nb_Comp_1 0.4663804 77.90810      NA           NA      NA      NA       NA
#> Nb_Comp_2 0.5728001 72.56787      NA           NA      NA      NA       NA
#> Nb_Comp_3 0.6214146 70.58113      NA           NA      NA      NA       NA
#> Nb_Comp_4 0.6680830 68.23976      NA           NA      NA      NA       NA
#> Nb_Comp_5 0.6702851 70.02009      NA           NA      NA      NA       NA
#> Nb_Comp_6 0.6737336 71.67313      NA           NA      NA      NA       NA
#> Nb_Comp_7 0.6807558 72.95511      NA           NA      NA      NA       NA
#> Nb_Comp_8 0.6819844 74.82787      NA           NA      NA      NA       NA
#> Nb_Comp_9 0.6853661 76.47507      NA           NA      NA      NA       NA
#>           DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1         2      0.5983679 0.3797438 0.4122175 -11.413749
#> Nb_Comp_2         3      0.5442372 0.3231208 0.3634169 -12.847656
#> Nb_Comp_3         4      0.5210941 0.3044529 0.3537087 -12.776843
#> Nb_Comp_4         5      0.4965569 0.2839276 0.3398355 -12.891035
#> Nb_Comp_5         6      0.5039885 0.3001871 0.3692997 -11.349498
#> Nb_Comp_6         7      0.5108963 0.3163819 0.3992388  -9.922119
#> Nb_Comp_7         8      0.5153766 0.3300041 0.4263658  -8.696873
#> Nb_Comp_8         9      0.5249910 0.3507834 0.4632727  -7.337679
#> Nb_Comp_9        10      0.5334234 0.3707649 0.4998004  -6.033403
#> attr(,"computed_nt")
#> [1] 9
#> 
#> [[3]]
#>                AIC       Q2cum_Y LimQ2_Y       Q2_Y  PRESS_Y     RSS_Y
#> Nb_Comp_0 82.41888            NA      NA         NA       NA 20.800152
#> Nb_Comp_1 63.69250  2.884891e-01  0.0975  0.2884891 14.79953 11.099368
#> Nb_Comp_2 58.35228  4.236248e-03  0.0975 -0.3995061 15.53363  8.885823
#> Nb_Comp_3 56.36553 -8.954496e-01  0.0975 -0.9035134 16.91428  7.874634
#> Nb_Comp_4 54.02416 -3.284001e+00  0.0975 -1.2601503 17.79786  6.903925
#> Nb_Comp_5 55.80450 -1.061425e+01  0.0975 -1.7110748 18.71706  6.858120
#> Nb_Comp_6 57.45753 -3.227259e+01  0.0975 -1.8648080 19.64720  6.786392
#> Nb_Comp_7 58.73951 -8.475017e+01  0.0975 -1.5772018 17.48990  6.640327
#> Nb_Comp_8 60.61227 -2.180818e+02  0.0975 -1.5548843 16.96527  6.614773
#> Nb_Comp_9 62.25948 -5.832137e+02  0.0975 -1.6666468 17.63926  6.544432
#>                R2_Y  AIC.std DoF.dof sigmahat.dof AIC.dof BIC.dof GMDL.dof
#> Nb_Comp_0        NA 96.63448      NA           NA      NA      NA       NA
#> Nb_Comp_1 0.4663804 77.90810      NA           NA      NA      NA       NA
#> Nb_Comp_2 0.5728001 72.56787      NA           NA      NA      NA       NA
#> Nb_Comp_3 0.6214146 70.58113      NA           NA      NA      NA       NA
#> Nb_Comp_4 0.6680830 68.23976      NA           NA      NA      NA       NA
#> Nb_Comp_5 0.6702851 70.02009      NA           NA      NA      NA       NA
#> Nb_Comp_6 0.6737336 71.67313      NA           NA      NA      NA       NA
#> Nb_Comp_7 0.6807558 72.95511      NA           NA      NA      NA       NA
#> Nb_Comp_8 0.6819844 74.82787      NA           NA      NA      NA       NA
#> Nb_Comp_9 0.6853661 76.47507      NA           NA      NA      NA       NA
#>           DoF.naive sigmahat.naive AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0         1      0.8062287 0.6697018 0.6991787  -3.605128
#> Nb_Comp_1         2      0.5983679 0.3797438 0.4122175 -11.413749
#> Nb_Comp_2         3      0.5442372 0.3231208 0.3634169 -12.847656
#> Nb_Comp_3         4      0.5210941 0.3044529 0.3537087 -12.776843
#> Nb_Comp_4         5      0.4965569 0.2839276 0.3398355 -12.891035
#> Nb_Comp_5         6      0.5039885 0.3001871 0.3692997 -11.349498
#> Nb_Comp_6         7      0.5108963 0.3163819 0.3992388  -9.922119
#> Nb_Comp_7         8      0.5153766 0.3300041 0.4263658  -8.696873
#> Nb_Comp_8         9      0.5249910 0.3507834 0.4632727  -7.337679
#> Nb_Comp_9        10      0.5334234 0.3707649 0.4998004  -6.033403
#> attr(,"computed_nt")
#> [1] 9
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"

data(aze_compl)
summary(cv.plsR(y~.,data=aze_compl,nt=10,K=8,NK=3,
verbose=FALSE),MClassed=TRUE,verbose=FALSE)

#> [[1]]
#>                 AIC MissClassed CV_MissClassed      Q2cum_Y LimQ2_Y       Q2_Y
#> Nb_Comp_0  154.6179          49             NA           NA      NA         NA
#> Nb_Comp_1  126.4083          27             45   -0.1285235  0.0975 -0.1285235
#> Nb_Comp_2  119.3375          25             44   -0.7308226  0.0975 -0.5337053
#> Nb_Comp_3  114.2313          27             41   -2.0779348  0.0975 -0.7783075
#> Nb_Comp_4  112.3463          23             46   -5.2967753  0.0975 -1.0457793
#> Nb_Comp_5  113.2362          22             45  -12.5021849  0.0975 -1.1443015
#> Nb_Comp_6  114.7620          21             45  -28.7556843  0.0975 -1.2037681
#> Nb_Comp_7  116.5264          20             44  -64.8964771  0.0975 -1.2145845
#> Nb_Comp_8  118.4601          20             45 -146.4878939  0.0975 -1.2381757
#> Nb_Comp_9  120.4452          19             45 -331.2780670  0.0975 -1.2529176
#> Nb_Comp_10 122.4395          19             46 -751.6321438  0.0975 -1.2650672
#>             PRESS_Y    RSS_Y      R2_Y  AIC.std  DoF.dof sigmahat.dof   AIC.dof
#> Nb_Comp_0        NA 25.91346        NA 298.1344  1.00000    0.5015845 0.2540061
#> Nb_Comp_1  29.24395 19.38086 0.2520929 269.9248 22.55372    0.4848429 0.2883114
#> Nb_Comp_2  29.72453 17.76209 0.3145613 262.8540 27.31542    0.4781670 0.2908950
#> Nb_Comp_3  31.58646 16.58896 0.3598323 257.7478 30.52370    0.4719550 0.2902572
#> Nb_Comp_4  33.93735 15.98071 0.3833049 255.8628 34.00000    0.4744263 0.3008285
#> Nb_Comp_5  34.26745 15.81104 0.3898523 256.7527 34.00000    0.4719012 0.2976347
#> Nb_Comp_6  34.84387 15.73910 0.3926285 258.2785 34.00000    0.4708264 0.2962804
#> Nb_Comp_7  34.85557 15.70350 0.3940024 260.0429 33.71066    0.4693382 0.2937976
#> Nb_Comp_8  35.14718 15.69348 0.3943888 261.9766 34.00000    0.4701436 0.2954217
#> Nb_Comp_9  35.35612 15.69123 0.3944758 263.9617 33.87284    0.4696894 0.2945815
#> Nb_Comp_10 35.54169 15.69037 0.3945088 265.9560 34.00000    0.4700970 0.2953632
#>              BIC.dof  GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive
#> Nb_Comp_0  0.2604032 -67.17645         1      0.5015845 0.2540061 0.2604032
#> Nb_Comp_1  0.4231184 -53.56607         2      0.4358996 0.1936625 0.2033251
#> Nb_Comp_2  0.4496983 -52.42272         3      0.4193593 0.1809352 0.1943501
#> Nb_Comp_3  0.4631316 -51.93343         4      0.4072955 0.1722700 0.1891422
#> Nb_Comp_4  0.4954133 -50.37079         5      0.4017727 0.1691819 0.1897041
#> Nb_Comp_5  0.4901536 -50.65724         6      0.4016679 0.1706451 0.1952588
#> Nb_Comp_6  0.4879234 -50.78005         7      0.4028135 0.1731800 0.2020601
#> Nb_Comp_7  0.4826103 -51.05525         8      0.4044479 0.1761610 0.2094352
#> Nb_Comp_8  0.4865092 -50.85833         9      0.4064413 0.1794902 0.2172936
#> Nb_Comp_9  0.4845867 -50.95616        10      0.4085682 0.1829787 0.2254232
#> Nb_Comp_10 0.4864128 -50.86368        11      0.4107477 0.1865584 0.2337468
#>            GMDL.naive
#> Nb_Comp_0   -67.17645
#> Nb_Comp_1   -79.67755
#> Nb_Comp_2   -81.93501
#> Nb_Comp_3   -83.31503
#> Nb_Comp_4   -83.23369
#> Nb_Comp_5   -81.93513
#> Nb_Comp_6   -80.42345
#> Nb_Comp_7   -78.87607
#> Nb_Comp_8   -77.31942
#> Nb_Comp_9   -75.80069
#> Nb_Comp_10  -74.33325
#> attr(,"computed_nt")
#> [1] 10
#> 
#> [[2]]
#>                 AIC MissClassed CV_MissClassed       Q2cum_Y LimQ2_Y
#> Nb_Comp_0  154.6179          49             NA            NA      NA
#> Nb_Comp_1  126.4083          27             44   -0.08103491  0.0975
#> Nb_Comp_2  119.3375          25             42   -0.54455526  0.0975
#> Nb_Comp_3  114.2313          27             43   -1.62495695  0.0975
#> Nb_Comp_4  112.3463          23             45   -4.19366808  0.0975
#> Nb_Comp_5  113.2362          22             45   -9.89561807  0.0975
#> Nb_Comp_6  114.7620          21             45  -22.31431378  0.0975
#> Nb_Comp_7  116.5264          20             47  -49.46336691  0.0975
#> Nb_Comp_8  118.4601          20             48 -109.38592158  0.0975
#> Nb_Comp_9  120.4452          19             47 -240.91286847  0.0975
#> Nb_Comp_10 122.4395          19             47 -529.51122639  0.0975
#>                   Q2_Y  PRESS_Y    RSS_Y      R2_Y  AIC.std  DoF.dof
#> Nb_Comp_0           NA       NA 25.91346        NA 298.1344  1.00000
#> Nb_Comp_1  -0.08103491 28.01336 19.38086 0.2520929 269.9248 22.55372
#> Nb_Comp_2  -0.42877463 27.69088 17.76209 0.3145613 262.8540 27.31542
#> Nb_Comp_3  -0.69949048 30.18650 16.58896 0.3598323 257.7478 30.52370
#> Nb_Comp_4  -0.97857267 32.82247 15.98071 0.3833049 255.8628 34.00000
#> Nb_Comp_5  -1.09786569 33.52537 15.81104 0.3898523 256.7527 34.00000
#> Nb_Comp_6  -1.13978809 33.83227 15.73910 0.3926285 258.2785 34.00000
#> Nb_Comp_7  -1.16448004 34.06697 15.70350 0.3940024 260.0429 33.71066
#> Nb_Comp_8  -1.18744662 34.35056 15.69348 0.3943888 261.9766 34.00000
#> Nb_Comp_9  -1.19151922 34.39257 15.69123 0.3944758 263.9617 33.87284
#> Nb_Comp_10 -1.19298473 34.41063 15.69037 0.3945088 265.9560 34.00000
#>            sigmahat.dof   AIC.dof   BIC.dof  GMDL.dof DoF.naive sigmahat.naive
#> Nb_Comp_0     0.5015845 0.2540061 0.2604032 -67.17645         1      0.5015845
#> Nb_Comp_1     0.4848429 0.2883114 0.4231184 -53.56607         2      0.4358996
#> Nb_Comp_2     0.4781670 0.2908950 0.4496983 -52.42272         3      0.4193593
#> Nb_Comp_3     0.4719550 0.2902572 0.4631316 -51.93343         4      0.4072955
#> Nb_Comp_4     0.4744263 0.3008285 0.4954133 -50.37079         5      0.4017727
#> Nb_Comp_5     0.4719012 0.2976347 0.4901536 -50.65724         6      0.4016679
#> Nb_Comp_6     0.4708264 0.2962804 0.4879234 -50.78005         7      0.4028135
#> Nb_Comp_7     0.4693382 0.2937976 0.4826103 -51.05525         8      0.4044479
#> Nb_Comp_8     0.4701436 0.2954217 0.4865092 -50.85833         9      0.4064413
#> Nb_Comp_9     0.4696894 0.2945815 0.4845867 -50.95616        10      0.4085682
#> Nb_Comp_10    0.4700970 0.2953632 0.4864128 -50.86368        11      0.4107477
#>            AIC.naive BIC.naive GMDL.naive
#> Nb_Comp_0  0.2540061 0.2604032  -67.17645
#> Nb_Comp_1  0.1936625 0.2033251  -79.67755
#> Nb_Comp_2  0.1809352 0.1943501  -81.93501
#> Nb_Comp_3  0.1722700 0.1891422  -83.31503
#> Nb_Comp_4  0.1691819 0.1897041  -83.23369
#> Nb_Comp_5  0.1706451 0.1952588  -81.93513
#> Nb_Comp_6  0.1731800 0.2020601  -80.42345
#> Nb_Comp_7  0.1761610 0.2094352  -78.87607
#> Nb_Comp_8  0.1794902 0.2172936  -77.31942
#> Nb_Comp_9  0.1829787 0.2254232  -75.80069
#> Nb_Comp_10 0.1865584 0.2337468  -74.33325
#> attr(,"computed_nt")
#> [1] 10
#> 
#> [[3]]
#>                 AIC MissClassed CV_MissClassed       Q2cum_Y LimQ2_Y       Q2_Y
#> Nb_Comp_0  154.6179          49             NA            NA      NA         NA
#> Nb_Comp_1  126.4083          27             42    -0.1441651  0.0975 -0.1441651
#> Nb_Comp_2  119.3375          25             43    -0.7334836  0.0975 -0.5150642
#> Nb_Comp_3  114.2313          27             47    -2.1922101  0.0975 -0.8415000
#> Nb_Comp_4  112.3463          23             49    -5.8233499  0.0975 -1.1375002
#> Nb_Comp_5  113.2362          22             48   -14.7739883  0.0975 -1.3117660
#> Nb_Comp_6  114.7620          21             50   -37.2667116  0.0975 -1.4259376
#> Nb_Comp_7  116.5264          20             48   -92.7653510  0.0975 -1.4503112
#> Nb_Comp_8  118.4601          20             49  -230.0101804  0.0975 -1.4637052
#> Nb_Comp_9  120.4452          19             49  -569.9483592  0.0975 -1.4715290
#> Nb_Comp_10 122.4395          19             49 -1410.4961895  0.0975 -1.4721959
#>             PRESS_Y    RSS_Y      R2_Y  AIC.std  DoF.dof sigmahat.dof   AIC.dof
#> Nb_Comp_0        NA 25.91346        NA 298.1344  1.00000    0.5015845 0.2540061
#> Nb_Comp_1  29.64928 19.38086 0.2520929 269.9248 22.55372    0.4848429 0.2883114
#> Nb_Comp_2  29.36325 17.76209 0.3145613 262.8540 27.31542    0.4781670 0.2908950
#> Nb_Comp_3  32.70889 16.58896 0.3598323 257.7478 30.52370    0.4719550 0.2902572
#> Nb_Comp_4  35.45891 15.98071 0.3833049 255.8628 34.00000    0.4744263 0.3008285
#> Nb_Comp_5  36.94365 15.81104 0.3898523 256.7527 34.00000    0.4719012 0.2976347
#> Nb_Comp_6  38.35660 15.73910 0.3926285 258.2785 34.00000    0.4708264 0.2962804
#> Nb_Comp_7  38.56569 15.70350 0.3940024 260.0429 33.71066    0.4693382 0.2937976
#> Nb_Comp_8  38.68878 15.69348 0.3943888 261.9766 34.00000    0.4701436 0.2954217
#> Nb_Comp_9  38.78690 15.69123 0.3944758 263.9617 33.87284    0.4696894 0.2945815
#> Nb_Comp_10 38.79179 15.69037 0.3945088 265.9560 34.00000    0.4700970 0.2953632
#>              BIC.dof  GMDL.dof DoF.naive sigmahat.naive AIC.naive BIC.naive
#> Nb_Comp_0  0.2604032 -67.17645         1      0.5015845 0.2540061 0.2604032
#> Nb_Comp_1  0.4231184 -53.56607         2      0.4358996 0.1936625 0.2033251
#> Nb_Comp_2  0.4496983 -52.42272         3      0.4193593 0.1809352 0.1943501
#> Nb_Comp_3  0.4631316 -51.93343         4      0.4072955 0.1722700 0.1891422
#> Nb_Comp_4  0.4954133 -50.37079         5      0.4017727 0.1691819 0.1897041
#> Nb_Comp_5  0.4901536 -50.65724         6      0.4016679 0.1706451 0.1952588
#> Nb_Comp_6  0.4879234 -50.78005         7      0.4028135 0.1731800 0.2020601
#> Nb_Comp_7  0.4826103 -51.05525         8      0.4044479 0.1761610 0.2094352
#> Nb_Comp_8  0.4865092 -50.85833         9      0.4064413 0.1794902 0.2172936
#> Nb_Comp_9  0.4845867 -50.95616        10      0.4085682 0.1829787 0.2254232
#> Nb_Comp_10 0.4864128 -50.86368        11      0.4107477 0.1865584 0.2337468
#>            GMDL.naive
#> Nb_Comp_0   -67.17645
#> Nb_Comp_1   -79.67755
#> Nb_Comp_2   -81.93501
#> Nb_Comp_3   -83.31503
#> Nb_Comp_4   -83.23369
#> Nb_Comp_5   -81.93513
#> Nb_Comp_6   -80.42345
#> Nb_Comp_7   -78.87607
#> Nb_Comp_8   -77.31942
#> Nb_Comp_9   -75.80069
#> Nb_Comp_10  -74.33325
#> attr(,"computed_nt")
#> [1] 10
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"