Computes Predicted Chisquare for k-fold cross-validated partial least squares regression models.

This function computes Predicted Chisquare for k-fold cross validated partial least squares regression models.

kfolds2Chisq(pls_kfolds)

Arguments

pls_kfolds
a k-fold cross validated partial least squares regression glm model

Value

list

Total Predicted Chisquare vs number of components for the first group partition

list()

...

list

Total Predicted Chisquare vs number of components for the last group partition

Note

Use cv.plsRglm to create k-fold cross validated partial least squares regression glm models.

References

Nicolas Meyer, Myriam Maumy-Bertrand et Frédéric Bertrand (2010). Comparing the linear and the logistic PLS regression with qualitative predictors: application to allelotyping data. Journal de la Societe Francaise de Statistique, 151(2), pages 1-18. http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/47

See also

Author

Frédéric Bertrand
frederic.bertrand@math.unistra.fr
https://fbertran.github.io/homepage/

Examples

# \donttest{
data(Cornell)
XCornell<-Cornell[,1:7]
yCornell<-Cornell[,8]
bbb <- cv.plsRglm(dataY=yCornell,dataX=XCornell,nt=3,modele="pls-glm-gaussian",K=16,verbose=FALSE)
bbb2 <- cv.plsRglm(dataY=yCornell,dataX=XCornell,nt=3,modele="pls-glm-gaussian",K=5,verbose=FALSE)
kfolds2Chisq(bbb)

#> [[1]]
#> [1] 55.70774 24.52966 20.84377
#> 

kfolds2Chisq(bbb2)

#> [[1]]
#> [1] 63.70823 44.75324 17.98309
#> 

rm(list=c("XCornell","yCornell","bbb","bbb2"))

data(pine)
Xpine<-pine[,1:10]
ypine<-pine[,11]
bbb <- cv.plsRglm(dataY=ypine,dataX=Xpine,nt=4,modele="pls-glm-gaussian",verbose=FALSE)
bbb2 <- cv.plsRglm(dataY=ypine,dataX=Xpine,nt=10,modele="pls-glm-gaussian",K=10,verbose=FALSE)
kfolds2Chisq(bbb)

#> [[1]]
#> [1] 14.67323 15.83399 13.44226 13.87808
#> 

kfolds2Chisq(bbb2)

#> [[1]]
#>  [1] 15.86442 14.20078 12.52512 12.76048 13.24603 13.93794 14.08724 14.38690
#>  [9] 14.53544 14.55343
#> 

                  
XpineNAX21 <- Xpine
XpineNAX21[1,2] <- NA
bbbNA <- cv.plsRglm(dataY=ypine,dataX=XpineNAX21,nt=10,modele="pls",K=10,verbose=FALSE)
kfolds2Press(bbbNA)

#> [[1]]
#> [1]  13.32860  13.34267  13.12193  12.51588  12.28941  12.28860 476.84506
#> [8] 387.24917 217.07164
#> 

kfolds2Chisq(bbbNA)

#> [[1]]
#> [1]  13.32860  13.34267  13.12193  12.51588  12.28941  12.28860 476.84506
#> [8] 387.24917 217.07164
#> 

bbbNA2 <- cv.plsRglm(dataY=ypine,dataX=XpineNAX21,nt=4,modele="pls-glm-gaussian",verbose=FALSE)
bbbNA3 <- cv.plsRglm(dataY=ypine,dataX=XpineNAX21,nt=10,modele="pls-glm-gaussian",K=10,
verbose=FALSE)
kfolds2Chisq(bbbNA2)

#> [[1]]
#> [1] 14.19954 15.20372 12.59614 15.63394
#> 

kfolds2Chisq(bbbNA3)

#> [[1]]
#> [1] 14.53105 15.59662 12.31008 13.13875 13.53567 16.90914 17.28329 16.21366
#> [9] 18.26045
#> 

rm(list=c("Xpine","XpineNAX21","ypine","bbb","bbb2","bbbNA","bbbNA2","bbbNA3"))

data(aze_compl)
Xaze_compl<-aze_compl[,2:34]
yaze_compl<-aze_compl$y
kfolds2Chisq(cv.plsRglm(dataY=yaze_compl,dataX=Xaze_compl,nt=4,modele="pls-glm-family",
family="binomial",verbose=FALSE))

#> [[1]]
#> [1]     265.5705    1052.6715   59043.5881 3377976.8075
#> 

kfolds2Chisq(cv.plsRglm(dataY=yaze_compl,dataX=Xaze_compl,nt=4,modele="pls-glm-logistic",
verbose=FALSE))

#> [[1]]
#> [1]   178.1906   423.0431  3325.0455 11007.2283
#> 

kfolds2Chisq(cv.plsRglm(dataY=yaze_compl,dataX=Xaze_compl,nt=10,modele="pls-glm-family",
family=binomial(),K=10,verbose=FALSE))

#> [[1]]
#>  [1]   199.2679   357.0954  1932.2912  9368.4989 12883.8045 14290.3535
#>  [7] 19095.1710 32588.9462 32956.4707 34580.0804
#> 

kfolds2Chisq(cv.plsRglm(dataY=yaze_compl,dataX=Xaze_compl,nt=10,modele="pls-glm-logistic",
K=10,verbose=FALSE))

#> [[1]]
#>  [1]  210.3097  330.7290  943.3300 1694.3946 2390.2444 3179.1935 4440.1936
#>  [8] 5705.0685 6110.1043 6899.4188
#> 

rm(list=c("Xaze_compl","yaze_compl"))
# }