Computes individual Predicted Chisquare for k-fold cross validated partial least squares regression models.

This function computes individual Predicted Chisquare for k-fold cross validated partial least squares regression models.

kfolds2Chisqind(pls_kfolds)

Arguments

pls_kfolds
a k-fold cross validated partial least squares regression glm model

Value

list

Individual PChisq vs number of components for the first group partition

…

…

list

Individual PChisq vs number of components for the last group partition

References

Nicolas Meyer, Myriam Maumy-Bertrand et Frederic Bertrand (2010). Comparing the linear and the logistic PLS regression with qualitative predictors: application to allelotyping data. Journal de la Societe Francaise de Statistique, 151(2), pages 1-18. http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/47

Note

Use cv.plsRglm to create k-fold cross validated partial least squares regression glm models.

See also

Examples

data(Cornell)
XCornell<-Cornell[,1:7]
yCornell<-Cornell[,8]
bbb <- cv.plsRglm(dataY=yCornell,dataX=XCornell,nt=3,modele="pls-glm-gaussian",K=16,verbose=FALSE)
bbb2 <- cv.plsRglm(dataY=yCornell,dataX=XCornell,nt=3,modele="pls-glm-gaussian",K=5,verbose=FALSE)
kfolds2Chisqind(bbb)

#> [[1]]
#> [[1]][[1]]
#>       [,1]    [,2]     [,3]
#> 1 24.52456 11.3923 6.837075
#> 
#> [[1]][[2]]
#>       [,1]        [,2]        [,3]
#> 2 4.124502 0.007566232 0.005169207
#> 
#> [[1]][[3]]
#>       [,1]     [,2]     [,3]
#> 3 1.551301 1.917305 1.907019
#> 
#> [[1]][[4]]
#>       [,1]     [,2]        [,3]
#> 4 12.26717 1.278425 0.003572679
#> 
#> [[1]][[5]]
#>       [,1]     [,2]     [,3]
#> 5 4.315406 5.393862 4.249662
#> 
#> [[1]][[6]]
#>       [,1]     [,2]     [,3]
#> 6 6.209332 3.528493 1.696534
#> 
#> [[1]][[7]]
#>         [,1]     [,2]      [,3]
#> 7 0.08812214 0.256731 0.1887149
#> 
#> [[1]][[8]]
#>        [,1]     [,2]     [,3]
#> 8 0.9442012 0.240037 0.359011
#> 
#> [[1]][[9]]
#>        [,1]       [,2]       [,3]
#> 9 0.2203272 0.02253865 0.03724274
#> 
#> [[1]][[10]]
#>         [,1]       [,2]       [,3]
#> 10 0.6034613 0.05144812 0.01893424
#> 
#> [[1]][[11]]
#>         [,1]      [,2]      [,3]
#> 11 0.8275623 0.4409475 0.7467284
#> 
#> [[1]][[12]]
#>          [,1]         [,2]     [,3]
#> 12 0.03180524 1.455262e-05 4.794104
#> 
#> 

kfolds2Chisqind(bbb2)

#> [[1]]
#> [[1]][[1]]
#> [1] 22.067161 10.440654  9.138804
#> 
#> [[1]][[2]]
#> [1] 16.115776  1.264465 12.766493
#> 
#> [[1]][[3]]
#> [1] 12.3290316  1.1463315  0.2526569
#> 
#> [[1]][[4]]
#> [1] 1.36983347 0.12981845 0.09180623
#> 
#> [[1]][[5]]
#> [1] 24.934791 10.816612  6.402283
#> 
#> 

rm(list=c("XCornell","yCornell","bbb","bbb2"))

data(pine)
Xpine<-pine[,1:10]
ypine<-pine[,11]
bbb <- cv.plsRglm(dataY=ypine,dataX=Xpine,nt=4,modele="pls-glm-gaussian",verbose=FALSE)
bbb2 <- cv.plsRglm(dataY=ypine,dataX=Xpine,nt=10,modele="pls-glm-gaussian",K=10,verbose=FALSE)
kfolds2Chisqind(bbb)

#> [[1]]
#> [[1]][[1]]
#> [1] 3.725803 4.641892 3.727099 3.833408
#> 
#> [[1]][[2]]
#> [1] 1.780774 1.208238 1.088441 1.618209
#> 
#> [[1]][[3]]
#> [1] 2.194861 2.021653 1.424912 1.452488
#> 
#> [[1]][[4]]
#> [1] 4.025689 2.601652 1.846375 1.850026
#> 
#> [[1]][[5]]
#> [1] 1.048989 2.307589 2.449096 2.985206
#> 
#> 

kfolds2Chisqind(bbb2)

#> [[1]]
#> [[1]][[1]]
#>  [1] 0.3419143 1.2871525 0.9803600 0.7506202 0.7905536 0.8309758 0.8356678
#>  [8] 0.6086570 0.6055412 0.5949694
#> 
#> [[1]][[2]]
#>  [1] 0.662799 1.484359 1.332373 1.414554 1.331225 1.588503 1.612451 1.550377
#>  [9] 1.631001 1.632272
#> 
#> [[1]][[3]]
#>  [1] 1.221701 1.251410 1.178048 1.630324 1.734206 1.846905 1.970479 2.023093
#>  [9] 2.049284 2.049541
#> 
#> [[1]][[4]]
#>  [1] 1.1759274 1.0864142 0.7793264 0.8984418 0.8348334 0.8560066 0.9165210
#>  [8] 0.9249125 0.8886741 0.8794710
#> 
#> [[1]][[5]]
#>  [1] 1.4915206 0.8736928 0.5645673 0.4914656 0.3833316 0.4498878 0.5654009
#>  [8] 0.6819146 0.8342383 0.7914959
#> 
#> [[1]][[6]]
#>  [1] 3.071831 3.438070 3.192888 2.994723 3.053770 3.146367 3.069776 3.011280
#>  [9] 3.009318 3.005890
#> 
#> [[1]][[7]]
#>  [1] 2.642612 2.022987 1.584486 1.818472 1.918153 2.203889 2.271137 2.211318
#>  [9] 2.191416 2.189715
#> 
#> [[1]][[8]]
#>  [1] 1.677329 1.604957 1.219848 1.281519 1.168038 1.191113 1.090612 1.129559
#>  [9] 1.124897 1.143311
#> 
#> [[1]][[9]]
#>  [1] 0.28665167 0.18409591 0.21870328 0.11196187 0.12788922 0.08672610
#>  [7] 0.09312459 0.09746886 0.10491397 0.10561476
#> 
#> [[1]][[10]]
#>  [1] 0.9185294 0.5650289 1.3828531 1.1540607 1.2611508 1.6456344 1.5793486
#>  [8] 1.6113090 1.6027154 1.6018576
#> 
#> 

XpineNAX21 <- Xpine
XpineNAX21[1,2] <- NA
bbbNA <- cv.plsRglm(dataY=ypine,dataX=XpineNAX21,nt=10,modele="pls",K=10,verbose=FALSE)
kfolds2Pressind(bbbNA)

#> [[1]]
#> [[1]][[1]]
#> [1] 1.9877352 1.2400202 0.7051718 0.8867770 0.8441230 1.0924820 1.1256926
#> [8] 1.1410708 1.1322459
#> 
#> [[1]][[2]]
#> [1] 1.4479585 1.4689772 0.7610111 0.4986794 0.6937640 0.5985580 0.6577486
#> [8] 0.6619129 0.5192713
#> 
#> [[1]][[3]]
#> [1] 1.3262781 1.4559167 1.6820091 0.9222805 1.0030430 1.0409797 1.0500127
#> [8] 1.0599875 1.0809069
#> 
#> [[1]][[4]]
#> [1] 2.660431 2.073841 2.281311 1.616194 1.534777 1.573232 1.488643 1.506703
#> [9] 1.567544
#> 
#> [[1]][[5]]
#> [1] 0.4061746 0.1552108 0.5365588 0.4410263 0.3840123 0.4203356 1.1238525
#> [8] 1.1402924 1.6239737
#> 
#> [[1]][[6]]
#> [1] 3.910643 3.922573 3.907449 3.793423 3.654508 3.386096 3.725462 3.994194
#> [9] 3.979033
#> 
#> [[1]][[7]]
#> [1] 0.2632494 0.6960174 0.8235719 0.9461809 0.8853331 0.9396041 0.9063967
#> [8] 0.8457616 1.0364687
#> 
#> [[1]][[8]]
#> [1] 0.07199681 1.18109274 1.04199664 0.73820125 0.59595048 0.54074035 0.52536412
#> [8] 0.40869899 0.46172085
#> 
#> [[1]][[9]]
#> [1] 0.4712493 1.2890847 1.1220104 1.5396032 1.9899198 1.9565588 1.9248824
#> [8] 1.9056878 1.9416277
#> 
#> [[1]][[10]]
#> [1] 2.9406589 2.2171850 2.7499497 0.6436153 0.9737924 2.4309491 2.8802695
#> [8] 5.2223547 4.1539794
#> 
#> 

kfolds2Chisqind(bbbNA)

#> [[1]]
#> [[1]][[1]]
#> [1] 1.9877352 1.2400202 0.7051718 0.8867770 0.8441230 1.0924820 1.1256926
#> [8] 1.1410708 1.1322459
#> 
#> [[1]][[2]]
#> [1] 1.4479585 1.4689772 0.7610111 0.4986794 0.6937640 0.5985580 0.6577486
#> [8] 0.6619129 0.5192713
#> 
#> [[1]][[3]]
#> [1] 1.3262781 1.4559167 1.6820091 0.9222805 1.0030430 1.0409797 1.0500127
#> [8] 1.0599875 1.0809069
#> 
#> [[1]][[4]]
#> [1] 2.660431 2.073841 2.281311 1.616194 1.534777 1.573232 1.488643 1.506703
#> [9] 1.567544
#> 
#> [[1]][[5]]
#> [1] 0.4061746 0.1552108 0.5365588 0.4410263 0.3840123 0.4203356 1.1238525
#> [8] 1.1402924 1.6239737
#> 
#> [[1]][[6]]
#> [1] 3.910643 3.922573 3.907449 3.793423 3.654508 3.386096 3.725462 3.994194
#> [9] 3.979033
#> 
#> [[1]][[7]]
#> [1] 0.2632494 0.6960174 0.8235719 0.9461809 0.8853331 0.9396041 0.9063967
#> [8] 0.8457616 1.0364687
#> 
#> [[1]][[8]]
#> [1] 0.07199681 1.18109274 1.04199664 0.73820125 0.59595048 0.54074035 0.52536412
#> [8] 0.40869899 0.46172085
#> 
#> [[1]][[9]]
#> [1] 0.4712493 1.2890847 1.1220104 1.5396032 1.9899198 1.9565588 1.9248824
#> [8] 1.9056878 1.9416277
#> 
#> [[1]][[10]]
#> [1] 2.9406589 2.2171850 2.7499497 0.6436153 0.9737924 2.4309491 2.8802695
#> [8] 5.2223547 4.1539794
#> 
#> 

bbbNA2 <- cv.plsRglm(dataY=ypine,dataX=XpineNAX21,nt=4,modele="pls-glm-gaussian",verbose=FALSE)
bbbNA3 <- cv.plsRglm(dataY=ypine,dataX=XpineNAX21,nt=10,modele="pls-glm-gaussian",
K=10,verbose=FALSE)
kfolds2Chisqind(bbbNA2)

#> [[1]]
#> [[1]][[1]]
#> [1] 1.604104 1.305470 1.490864 1.603281
#> 
#> [[1]][[2]]
#> [1] 2.821231 2.891760 2.367716 2.638871
#> 
#> [[1]][[3]]
#> [1] 1.8473845 2.2130315 0.6013368 1.0218579
#> 
#> [[1]][[4]]
#> [1] 1.3615135 0.8546937 1.4215356 1.6234697
#> 
#> [[1]][[5]]
#> [1] 4.906348 4.622124 4.068846 4.048347
#> 
#> 

kfolds2Chisqind(bbbNA3)

#> [[1]]
#> [[1]][[1]]
#> [1] 0.7106165 1.5732375 1.2593267 1.4657069 1.1089610 0.9435593 0.8352889
#> [8] 0.7899435 0.8036632
#> 
#> [[1]][[2]]
#> [1]  2.1807850  0.5655184  0.4423157  0.5164543  1.9422958  4.5314055  7.8685753
#> [8]  7.8032168 13.8306226
#> 
#> [[1]][[3]]
#> [1] 3.225404 3.930058 3.655811 3.522632 3.632707 3.623796 3.698964 3.900180
#> [9] 3.890371
#> 
#> [[1]][[4]]
#> [1] 2.0698698 1.9545453 0.9748627 0.7793856 0.7250085 0.8368241 0.8803358
#> [8] 0.8858411 0.9103985
#> 
#> [[1]][[5]]
#> [1] 0.7891970 0.3511947 0.4190787 0.7638380 0.7307571 0.8912552 0.8046593
#> [8] 1.0567351 4.1251038
#> 
#> [[1]][[6]]
#> [1] 0.8862628 0.8078544 0.2244958 0.1368507 0.1489448 0.1069866 0.1678020
#> [8] 0.1438793 0.2780674
#> 
#> [[1]][[7]]
#> [1] 0.3009420 1.2239979 0.7422919 0.6427654 0.4981736 0.6986124 0.5457891
#> [8] 1.0292046 0.6573797
#> 
#> [[1]][[8]]
#> [1] 2.772124 2.644073 2.549261 2.800353 2.604072 2.690133 2.282196 2.661910
#> [9] 7.247917
#> 
#> [[1]][[9]]
#> [1] 0.4379035 1.1482476 1.1421259 1.5923654 1.7423105 1.7075681 1.9254434
#> [8] 1.8774651 2.0897768
#> 
#> [[1]][[10]]
#> [1] 1.1619103 0.3863560 0.1956144 0.4075736 0.9364455 0.8268879 0.8156700
#> [8] 0.7841018 0.6122844
#> 
#> 

rm(list=c("Xpine","XpineNAX21","ypine","bbb","bbb2","bbbNA","bbbNA2","bbbNA3"))

data(aze_compl)
Xaze_compl<-aze_compl[,2:34]
yaze_compl<-aze_compl$y
kfolds2Chisqind(cv.plsRglm(dataY=yaze_compl,dataX=Xaze_compl,nt=4,modele="pls-glm-family",
family=binomial(),verbose=FALSE))

#> [[1]]
#> [[1]][[1]]
#> [1]  25.71637  38.63916  74.63584 374.27350
#> 
#> [[1]][[2]]
#> [1] 20.65464 19.69346 27.66413 41.22474
#> 
#> [[1]][[3]]
#> [1]   44.71689   94.22743  594.56748 2384.88970
#> 
#> [[1]][[4]]
#> [1]  63.73964 118.00707 282.93099 458.26215
#> 
#> [[1]][[5]]
#> [1]  33.56827  44.77075  54.57548 298.04498
#> 
#> 

kfolds2Chisqind(cv.plsRglm(dataY=yaze_compl,dataX=Xaze_compl,nt=4,modele="pls-glm-logistic",
verbose=FALSE))

#> [[1]]
#> [[1]][[1]]
#> [1]  57.10107  61.29625 164.01760 511.84712
#> 
#> [[1]][[2]]
#> [1]     182.8760     858.1587   32478.3224 1069891.1202
#> 
#> [[1]][[3]]
#> [1]   38.18775  124.71319 1466.22715 3102.84293
#> 
#> [[1]][[4]]
#> [1]  22.80330  30.50470  88.29049 316.20268
#> 
#> [[1]][[5]]
#> [1]  36.85715  59.15028 111.00632 476.49618
#> 
#> 

kfolds2Chisqind(cv.plsRglm(dataY=yaze_compl,dataX=Xaze_compl,nt=10,modele="pls-glm-family",
family=binomial(),K=10,verbose=FALSE))

#> [[1]]
#> [[1]][[1]]
#>  [1]  25.03881  61.15117 116.96963 326.32259 403.44765 538.39049 645.64491
#>  [8] 575.70117 549.17718 514.68384
#> 
#> [[1]][[2]]
#>  [1]  17.65843  23.67390  36.81069  90.98192  93.10981 128.47980 200.60186
#>  [8] 252.74403 308.06547 319.82667
#> 
#> [[1]][[3]]
#>  [1]  8.526926  8.076974 11.106051 15.860676 15.004941 17.290974 19.757907
#>  [8] 18.251522 19.879533 20.435341
#> 
#> [[1]][[4]]
#>  [1]  11.78145  14.06469  19.05486  43.10176  45.58889  90.98788  98.04530
#>  [8] 140.01973 141.30290 125.97587
#> 
#> [[1]][[5]]
#>  [1] 3.639156e+01 8.736942e+01 7.270198e+02 4.700947e+03 1.192978e+04
#>  [6] 1.894180e+05 1.015788e+06 3.511915e+06 8.845400e+06 2.010484e+07
#> 
#> [[1]][[6]]
#>  [1]   42.56071  120.19779  484.16973 1598.17187 2348.21872 2944.96148
#>  [7] 3904.13470 5212.63956 5477.85020 5711.21293
#> 
#> [[1]][[7]]
#>  [1] 11.80989 11.77089 14.79555 24.43992 34.59222 50.31828 59.87192 70.03518
#>  [9] 76.06148 76.06860
#> 
#> [[1]][[8]]
#>  [1]  9.254977 12.326114 18.237317 20.176162 26.492609 27.169289 25.801275
#>  [8] 24.927435 23.869865 23.539210
#> 
#> [[1]][[9]]
#>  [1]  11.10385  13.49195  12.63281  33.81317  44.92917  80.57929 116.02671
#>  [8] 154.45650 252.50944 359.22135
#> 
#> [[1]][[10]]
#>  [1]  19.60646  19.94476  56.81703 198.90182 213.40298 338.58119 310.21947
#>  [8] 272.40363 247.20657 257.02387
#> 
#> 

kfolds2Chisqind(cv.plsRglm(dataY=yaze_compl,dataX=Xaze_compl,nt=10,
modele="pls-glm-logistic",K=10,verbose=FALSE))

#> [[1]]
#> [[1]][[1]]
#>  [1]  15.82372  23.76066  39.21934 129.83669 177.56668 239.38790 327.49421
#>  [8] 409.21113 422.35622 441.76642
#> 
#> [[1]][[2]]
#>  [1]  13.60109  26.32456  27.38987 127.96344 208.07980 367.54309 399.44914
#>  [8] 363.71844 365.64184 357.50253
#> 
#> [[1]][[3]]
#>  [1]   17.36662   53.73285  294.32163  624.36132  670.57948 1119.79705
#>  [7] 2055.26603 2107.49226 1655.56018 1638.30953
#> 
#> [[1]][[4]]
#>  [1]     36.7037    193.7240  25107.1037 123808.8461 269037.8828 524035.3492
#>  [7] 546111.3374 558621.3681 679124.0093 545228.8818
#> 
#> [[1]][[5]]
#>  [1] 12.118411  9.434291  8.658750 14.359197 12.542803 11.332804 12.060522
#>  [8] 11.717814 12.135901 12.243959
#> 
#> [[1]][[6]]
#>  [1] 19.792164 20.321925 19.158528 16.505363 13.657078 12.277128 10.519013
#>  [8]  9.460846  9.820052  9.652815
#> 
#> [[1]][[7]]
#>  [1]  9.003066 12.173491 15.098615 15.194809 20.440263 28.854874 45.714044
#>  [8] 51.764982 49.676690 42.869772
#> 
#> [[1]][[8]]
#>  [1]    26.21149    62.12644   292.17163  1679.66283  2722.80302  4462.55212
#>  [7]  6022.93573 10821.82879 11251.05271  9985.56587
#> 
#> [[1]][[9]]
#>  [1]  45.73993  59.62235  48.49922  72.23083  83.21256 136.73702 164.57404
#>  [8] 189.73442 197.52295 202.01123
#> 
#> [[1]][[10]]
#>  [1]  7.448592  7.562965 18.342046 24.859778 30.400829 34.927138 41.911143
#>  [8] 45.994860 49.679275 50.390185
#> 
#> 

rm(list=c("Xaze_compl","yaze_compl"))