Computes individual Predicted Chisquare for k-fold cross validated partial least squares regression models.

This function computes individual Predicted Chisquare for k-fold cross validated partial least squares regression models.

Usage

kfolds2Chisqind(pls_kfolds)

Arguments

pls_kfolds

a k-fold cross validated partial least squares regression glm model

Value

list

Individual PChisq vs number of components for the first group partition

list()

...

list

Individual PChisq vs number of components for the last group partition

Note

Use cv.plsRglm to create k-fold cross validated partial least squares regression glm models.

References

Nicolas Meyer, Myriam Maumy-Bertrand et Frédéric Bertrand (2010). Comparing the linear and the logistic PLS regression with qualitative predictors: application to allelotyping data. Journal de la Societe Francaise de Statistique, 151(2), pages 1-18. https://www.numdam.org/item/JSFS_2010__151_2_1_0/

See also

Author

Frédéric Bertrand
frederic.bertrand@lecnam.net
https://fbertran.github.io/homepage/

Examples

# \donttest{
data(Cornell)
XCornell<-Cornell[,1:7]
yCornell<-Cornell[,8]
bbb <- cv.plsRglm(object=yCornell,dataX=XCornell,nt=3,modele="pls-glm-gaussian",K=16,verbose=FALSE)
bbb2 <- cv.plsRglm(object=yCornell,dataX=XCornell,nt=3,modele="pls-glm-gaussian",K=5,verbose=FALSE)
kfolds2Chisqind(bbb)
#> [[1]]
#> [[1]][[1]]
#>       [,1]    [,2]     [,3]
#> 1 24.52456 11.3923 6.837075
#> 
#> [[1]][[2]]
#>       [,1]        [,2]        [,3]
#> 2 4.124502 0.007566232 0.005169207
#> 
#> [[1]][[3]]
#>       [,1]     [,2]     [,3]
#> 3 1.551301 1.917305 1.907019
#> 
#> [[1]][[4]]
#>       [,1]     [,2]        [,3]
#> 4 12.26717 1.278425 0.003572679
#> 
#> [[1]][[5]]
#>       [,1]     [,2]     [,3]
#> 5 4.315406 5.393862 4.249662
#> 
#> [[1]][[6]]
#>       [,1]     [,2]     [,3]
#> 6 6.209332 3.528493 1.696534
#> 
#> [[1]][[7]]
#>         [,1]     [,2]      [,3]
#> 7 0.08812214 0.256731 0.1887149
#> 
#> [[1]][[8]]
#>        [,1]     [,2]     [,3]
#> 8 0.9442012 0.240037 0.359011
#> 
#> [[1]][[9]]
#>        [,1]       [,2]       [,3]
#> 9 0.2203272 0.02253865 0.03724274
#> 
#> [[1]][[10]]
#>         [,1]       [,2]       [,3]
#> 10 0.6034613 0.05144812 0.01893424
#> 
#> [[1]][[11]]
#>         [,1]      [,2]      [,3]
#> 11 0.8275623 0.4409475 0.7467284
#> 
#> [[1]][[12]]
#>          [,1]         [,2]     [,3]
#> 12 0.03180524 1.455262e-05 4.794104
#> 
#> 
kfolds2Chisqind(bbb2)
#> [[1]]
#> [[1]][[1]]
#> [1] 5.270827 1.511911 1.430687
#> 
#> [[1]][[2]]
#> [1] 0.6394877 0.4095096 1.0584582
#> 
#> [[1]][[3]]
#> [1] 26.502867 13.361154  8.250662
#> 
#> [[1]][[4]]
#> [1] 1.985502 2.149674 2.040927
#> 
#> [[1]][[5]]
#> [1] 11.849695  2.460984  6.748015
#> 
#> 
rm(list=c("XCornell","yCornell","bbb","bbb2"))

data(pine)
Xpine<-pine[,1:10]
ypine<-pine[,11]
bbb <- cv.plsRglm(object=ypine,dataX=Xpine,nt=4,modele="pls-glm-gaussian",verbose=FALSE)
bbb2 <- cv.plsRglm(object=ypine,dataX=Xpine,nt=10,modele="pls-glm-gaussian",K=10,verbose=FALSE)
kfolds2Chisqind(bbb)
#> [[1]]
#> [[1]][[1]]
#> [1] 3.403086 4.775572 4.519409 5.002448
#> 
#> [[1]][[2]]
#> [1] 1.824134 2.162432 1.596622 1.359406
#> 
#> [[1]][[3]]
#> [1] 2.587970 1.691074 1.391184 1.443778
#> 
#> [[1]][[4]]
#> [1] 1.266668 1.572788 1.260169 1.344553
#> 
#> [[1]][[5]]
#> [1] 3.749667 3.031469 1.618427 1.615485
#> 
#> 
kfolds2Chisqind(bbb2)
#> [[1]]
#> [[1]][[1]]
#>  [1] 0.8869803 0.8201235 0.7967343 0.7834618 0.7443253 0.7626616 0.6772168
#>  [8] 0.6074659 0.6187567 0.6415269
#> 
#> [[1]][[2]]
#>  [1] 4.727405 4.636497 4.341603 4.546572 4.759693 5.077514 5.229127 5.471423
#>  [9] 5.487885 5.488157
#> 
#> [[1]][[3]]
#>  [1] 1.077921 1.151543 1.204886 1.550415 1.729665 1.970757 1.940021 2.001377
#>  [9] 2.000931 1.989826
#> 
#> [[1]][[4]]
#>  [1] 1.5252320 1.1816437 0.3847527 0.3050577 0.2951873 0.2509198 0.3295281
#>  [8] 0.3587540 0.3784665 0.3785523
#> 
#> [[1]][[5]]
#>  [1] 0.650217 1.536378 1.535037 1.563574 1.454718 1.198908 1.244873 1.296059
#>  [9] 1.271407 1.272678
#> 
#> [[1]][[6]]
#>  [1] 0.5208920 0.1635685 0.2768095 0.3350192 0.4934330 0.4361214 0.4658536
#>  [8] 0.4806951 0.4824459 0.4827584
#> 
#> [[1]][[7]]
#>  [1] 0.006366112 1.201509729 0.769069823 0.770990708 0.805811006 0.908296996
#>  [7] 0.883070758 0.858064121 0.854869289 0.854843256
#> 
#> [[1]][[8]]
#>  [1] 0.05687795 0.16551094 0.37338030 0.32373596 0.30780311 0.17014380
#>  [7] 0.16251227 0.17074300 0.17904736 0.17889468
#> 
#> [[1]][[9]]
#>  [1] 2.514685 1.264220 1.034154 1.011181 1.009631 1.368947 1.651633 1.731609
#>  [9] 1.770509 1.760526
#> 
#> [[1]][[10]]
#>  [1] 1.0707442 0.3930249 0.2207084 0.5900025 0.9509311 0.7454288 0.7667870
#>  [8] 0.7697079 0.7585861 0.7579902
#> 
#> 
                  
XpineNAX21 <- Xpine
XpineNAX21[1,2] <- NA
bbbNA <- cv.plsRglm(object=ypine,dataX=XpineNAX21,nt=10,modele="pls",K=10,verbose=FALSE)
kfolds2Pressind(bbbNA)
#> [[1]]
#> [[1]][[1]]
#> [1] 3.641936 2.715888 2.815467 1.642285 1.704489 1.733388 1.742250 1.746528
#> [9] 1.703715
#> 
#> [[1]][[2]]
#> [1] 0.2565236 0.5619357 0.6086918 0.5428673 0.5498746 0.6248340 0.8229439
#> [8] 0.8311556 0.9891515
#> 
#> [[1]][[3]]
#> [1] 1.9831732 2.8387181 2.8462283 1.4337590 1.2091250 0.9997404 1.2264522
#> [8] 1.2905409 1.4198860
#> 
#> [[1]][[4]]
#> [1] 2.791337 2.883241 2.413335 2.479880 2.666350 2.654463 2.730224 2.694283
#> [9] 2.493838
#> 
#> [[1]][[5]]
#> [1]  1.392256  1.558197  1.293175  0.928126  1.770689  3.311278  4.164474
#> [8] 18.322802 19.378881
#> 
#> [[1]][[6]]
#> [1] 0.6020566 0.2924178 0.2213609 0.2406571 0.4891932 0.4846383 0.5742382
#> [8] 0.5750435 0.7232366
#> 
#> [[1]][[7]]
#> [1] 0.6683048 1.3658249 1.8962920 0.8775053 0.8153458 0.7861862 0.3650173
#> [8] 0.3374038 1.0031900
#> 
#> [[1]][[8]]
#> [1] 0.9329137 0.6598429 0.3628448 0.4212905 0.4961495 0.7362372 0.8537199
#> [8] 0.8657835 0.8664770
#> 
#> [[1]][[9]]
#> [1] 0.5409445 0.8358791 0.5884406 0.5631689 0.5332073 0.4849964 0.3831118
#> [8] 0.3429699 0.4277211
#> 
#> [[1]][[10]]
#> [1] 0.6720491 0.9014411 1.2582868 1.4982960 1.3938938 1.4452417 1.4827895
#> [8] 1.4351993 1.5402511
#> 
#> 
kfolds2Chisqind(bbbNA)
#> [[1]]
#> [[1]][[1]]
#> [1] 3.641936 2.715888 2.815467 1.642285 1.704489 1.733388 1.742250 1.746528
#> [9] 1.703715
#> 
#> [[1]][[2]]
#> [1] 0.2565236 0.5619357 0.6086918 0.5428673 0.5498746 0.6248340 0.8229439
#> [8] 0.8311556 0.9891515
#> 
#> [[1]][[3]]
#> [1] 1.9831732 2.8387181 2.8462283 1.4337590 1.2091250 0.9997404 1.2264522
#> [8] 1.2905409 1.4198860
#> 
#> [[1]][[4]]
#> [1] 2.791337 2.883241 2.413335 2.479880 2.666350 2.654463 2.730224 2.694283
#> [9] 2.493838
#> 
#> [[1]][[5]]
#> [1]  1.392256  1.558197  1.293175  0.928126  1.770689  3.311278  4.164474
#> [8] 18.322802 19.378881
#> 
#> [[1]][[6]]
#> [1] 0.6020566 0.2924178 0.2213609 0.2406571 0.4891932 0.4846383 0.5742382
#> [8] 0.5750435 0.7232366
#> 
#> [[1]][[7]]
#> [1] 0.6683048 1.3658249 1.8962920 0.8775053 0.8153458 0.7861862 0.3650173
#> [8] 0.3374038 1.0031900
#> 
#> [[1]][[8]]
#> [1] 0.9329137 0.6598429 0.3628448 0.4212905 0.4961495 0.7362372 0.8537199
#> [8] 0.8657835 0.8664770
#> 
#> [[1]][[9]]
#> [1] 0.5409445 0.8358791 0.5884406 0.5631689 0.5332073 0.4849964 0.3831118
#> [8] 0.3429699 0.4277211
#> 
#> [[1]][[10]]
#> [1] 0.6720491 0.9014411 1.2582868 1.4982960 1.3938938 1.4452417 1.4827895
#> [8] 1.4351993 1.5402511
#> 
#> 
bbbNA2 <- cv.plsRglm(object=ypine,dataX=XpineNAX21,nt=4,modele="pls-glm-gaussian",verbose=FALSE)
bbbNA3 <- cv.plsRglm(object=ypine,dataX=XpineNAX21,nt=10,modele="pls-glm-gaussian",
K=10,verbose=FALSE)
kfolds2Chisqind(bbbNA2)
#> [[1]]
#> [[1]][[1]]
#> [1] 5.091976 5.668531 7.992906 8.948263
#> 
#> [[1]][[2]]
#> [1] 2.312984 1.441495 1.488056 1.738745
#> 
#> [[1]][[3]]
#> [1] 1.553196 2.551305 2.206062 2.006054
#> 
#> [[1]][[4]]
#> [1] 3.749187 3.766897 3.202139 3.041355
#> 
#> [[1]][[5]]
#> [1] 0.7240605 4.1309348 3.9388575 4.4043879
#> 
#> 
kfolds2Chisqind(bbbNA3)
#> [[1]]
#> [[1]][[1]]
#> [1] 1.1794345 1.1577211 0.5312897 0.7023919 0.9409709 1.1335568 1.5944448
#> [8] 1.0753626 1.0515777
#> 
#> [[1]][[2]]
#> [1] 1.0068477 0.5015760 0.2035914 0.6133291 0.9741998 1.2965126 1.4742455
#> [8] 1.6903695 0.8815662
#> 
#> [[1]][[3]]
#> [1] 1.872407 3.523068 2.687198 4.599433 5.535495 6.715902 6.318444 3.070310
#> [9] 4.747822
#> 
#> [[1]][[4]]
#> [1] 2.680572 1.515843 1.609055 1.487503 1.489059 1.568317 1.673913 1.541438
#> [9] 1.695382
#> 
#> [[1]][[5]]
#> [1] 0.9831575 1.0232133 0.9547098 0.8725859 0.9718035 0.9406500 0.9357887
#> [8] 0.9024237 1.0029078
#> 
#> [[1]][[6]]
#> [1] 0.7022996 1.1960171 0.9894453 0.9033905 0.9318882 1.0201916 1.1004181
#> [8] 1.1643403 1.2396577
#> 
#> [[1]][[7]]
#> [1] 2.893779 2.984157 2.908782 2.823959 2.736246 2.848856 2.692257 2.665015
#> [9] 2.664210
#> 
#> [[1]][[8]]
#> [1] 0.1939123 0.3389540 0.2573845 0.3813121 0.4585366 0.3817952 0.5314726
#> [8] 0.6209207 0.7590313
#> 
#> [[1]][[9]]
#> [1] 0.3042696 0.1721206 0.3637156 0.3045094 0.3176336 0.2976240 0.2469196
#> [8] 0.2111337 0.1470695
#> 
#> [[1]][[10]]
#> [1] 0.9630187 0.8365556 0.5605143 0.5959843 0.4754921 0.4567590 0.5664468
#> [8] 0.6122550 0.6728351
#> 
#> 
rm(list=c("Xpine","XpineNAX21","ypine","bbb","bbb2","bbbNA","bbbNA2","bbbNA3"))

data(aze_compl)
Xaze_compl<-aze_compl[,2:34]
yaze_compl<-aze_compl$y
kfolds2Chisqind(cv.plsRglm(object=yaze_compl,dataX=Xaze_compl,nt=4,modele="pls-glm-family",
family=binomial(),verbose=FALSE))
#> [[1]]
#> [[1]][[1]]
#> [1] 26.90170 36.40484 38.51725 38.87166
#> 
#> [[1]][[2]]
#> [1]   32.61727   62.73811  148.57031 2226.98288
#> 
#> [[1]][[3]]
#> [1]  26.01683  95.75682 614.09120 759.75406
#> 
#> [[1]][[4]]
#> [1]   52.94904  114.88423  989.88793 4112.44160
#> 
#> [[1]][[5]]
#> [1]  29.93034  34.08810  53.25737 113.07159
#> 
#> 
kfolds2Chisqind(cv.plsRglm(object=yaze_compl,dataX=Xaze_compl,nt=4,modele="pls-glm-logistic",
verbose=FALSE))
#> [[1]]
#> [[1]][[1]]
#> [1]  47.59640  80.94215 113.33396 183.61306
#> 
#> [[1]][[2]]
#> [1]  74.79705 207.64307 335.03829 752.79300
#> 
#> [[1]][[3]]
#> [1]   47.09273   74.93908  266.56083 1377.30320
#> 
#> [[1]][[4]]
#> [1]  31.73475  54.16956  93.65354 133.03732
#> 
#> [[1]][[5]]
#> [1]  26.46269  47.19208  98.07683 167.21246
#> 
#> 
kfolds2Chisqind(cv.plsRglm(object=yaze_compl,dataX=Xaze_compl,nt=10,modele="pls-glm-family",
family=binomial(),K=10,verbose=FALSE))
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> [[1]]
#> [[1]][[1]]
#>  [1]  30.37905 131.61933 295.07586 449.43316 465.95404 622.34039 594.66732
#>  [8] 552.40620 541.41116 461.34886
#> 
#> [[1]][[2]]
#>  [1] 12.54716 14.91610 24.01796 31.44286 28.58675 31.12905 30.44918 30.91007
#>  [9] 30.29058 32.20295
#> 
#> [[1]][[3]]
#>  [1]  19.25542  30.38544  37.77202 128.59382 150.09659 243.47865 413.22428
#>  [8] 642.23408 700.38212 741.01251
#> 
#> [[1]][[4]]
#>  [1] 4.490421e+01 1.732139e+02 1.051098e+04 6.309520e+05 9.187609e+06
#>  [6] 1.949896e+09 4.503771e+15 9.010711e+15 1.801494e+16 2.251800e+16
#> 
#> [[1]][[5]]
#>  [1] 16.966171  7.672902  9.012394 11.454132 11.474949 14.098260 14.840711
#>  [8] 14.254955 13.197875 12.721494
#> 
#> [[1]][[6]]
#>  [1]  7.444570  6.333452  8.840397 10.865892 15.210250 16.900480 21.882404
#>  [8] 27.511089 27.772383 28.208200
#> 
#> [[1]][[7]]
#>  [1]  12.57589  15.83949  31.31110 123.82153 135.82582 122.53225  96.60512
#>  [8]  75.62408  88.89131  99.92953
#> 
#> [[1]][[8]]
#>  [1]  36.78350  58.18266 124.57593 241.31382 404.07471 501.45468 515.44248
#>  [8] 504.49675 437.32656 388.22272
#> 
#> [[1]][[9]]
#>  [1]   9.679721  13.718786  15.340246  27.822643  49.031872  78.163101
#>  [7] 152.973442 264.490718 300.856504 253.825429
#> 
#> [[1]][[10]]
#>  [1]  23.52748  35.34624  53.95292 112.67853 182.33070 201.99742 217.28400
#>  [8] 287.72665 298.27349 288.38705
#> 
#> 
kfolds2Chisqind(cv.plsRglm(object=yaze_compl,dataX=Xaze_compl,nt=10,
modele="pls-glm-logistic",K=10,verbose=FALSE))
#> [[1]]
#> [[1]][[1]]
#>  [1] 17.16111 20.54780 42.96324 50.41764 67.83271 72.45856 87.38555 86.40900
#>  [9] 93.93810 92.43602
#> 
#> [[1]][[2]]
#>  [1] 17.488320 10.097871  8.616709 15.236710 22.035719 26.949970 25.709833
#>  [8] 27.980402 27.834312 27.664695
#> 
#> [[1]][[3]]
#>  [1]   9.314635  10.629045  24.802254  47.837962  66.721316 107.068433
#>  [7] 124.279462 134.636647 138.991686 137.468731
#> 
#> [[1]][[4]]
#>  [1]    81.06815   210.21352  3665.07542 19566.78091 33527.51517 32555.33438
#>  [7] 53366.06399 47213.13269 64700.19582 68795.91094
#> 
#> [[1]][[5]]
#>  [1]  23.37379  50.96192 367.69449 344.31163 401.72190 504.27506 762.19446
#>  [8] 810.07412 668.38914 617.19564
#> 
#> [[1]][[6]]
#>  [1]  22.49615  33.89744  60.19610 149.92402 187.79812 217.12727 224.88363
#>  [8] 214.55649 211.12614 227.56614
#> 
#> [[1]][[7]]
#>  [1] 16.26003 19.50735 21.72989 20.71220 20.25186 17.73315 20.04467 27.39599
#>  [9] 26.88810 26.90085
#> 
#> [[1]][[8]]
#>  [1]  12.09134  13.00128  21.37000  29.14232  30.79322  54.11291  79.85591
#>  [8] 159.90995 162.14637 169.92531
#> 
#> [[1]][[9]]
#>  [1] 25.68953 40.57266 39.28939 91.36514 87.23446 80.54601 80.31430 88.24899
#>  [9] 90.94341 89.85898
#> 
#> [[1]][[10]]
#>  [1]   16.42260   36.44474   73.14231  290.03249  575.97750 1119.96831
#>  [7] 1669.19671 1947.92863 1990.44154 1704.00586
#> 
#> 
rm(list=c("Xaze_compl","yaze_compl"))
# }