Computes individual Predicted Chisquare for k-fold cross validated partial least squares regression models.

This function computes individual Predicted Chisquare for k-fold cross validated partial least squares regression models.

kfolds2Chisqind(pls_kfolds)

Arguments

pls_kfolds
a k-fold cross validated partial least squares regression glm model

Value

list

Individual PChisq vs number of components for the first group partition

list()

...

list

Individual PChisq vs number of components for the last group partition

Note

Use cv.plsRglm to create k-fold cross validated partial least squares regression glm models.

References

Nicolas Meyer, Myriam Maumy-Bertrand et Frédéric Bertrand (2010). Comparing the linear and the logistic PLS regression with qualitative predictors: application to allelotyping data. Journal de la Societe Francaise de Statistique, 151(2), pages 1-18. http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/47

See also

Author

Frédéric Bertrand
frederic.bertrand@math.unistra.fr
https://fbertran.github.io/homepage/

Examples

# \donttest{
data(Cornell)
XCornell<-Cornell[,1:7]
yCornell<-Cornell[,8]
bbb <- cv.plsRglm(dataY=yCornell,dataX=XCornell,nt=3,modele="pls-glm-gaussian",K=16,verbose=FALSE)
bbb2 <- cv.plsRglm(dataY=yCornell,dataX=XCornell,nt=3,modele="pls-glm-gaussian",K=5,verbose=FALSE)
kfolds2Chisqind(bbb)

#> [[1]]
#> [[1]][[1]]
#>       [,1]    [,2]     [,3]
#> 1 24.52456 11.3923 6.837075
#> 
#> [[1]][[2]]
#>       [,1]        [,2]        [,3]
#> 2 4.124502 0.007566232 0.005169207
#> 
#> [[1]][[3]]
#>       [,1]     [,2]     [,3]
#> 3 1.551301 1.917305 1.907019
#> 
#> [[1]][[4]]
#>       [,1]     [,2]        [,3]
#> 4 12.26717 1.278425 0.003572679
#> 
#> [[1]][[5]]
#>       [,1]     [,2]     [,3]
#> 5 4.315406 5.393862 4.249662
#> 
#> [[1]][[6]]
#>       [,1]     [,2]     [,3]
#> 6 6.209332 3.528493 1.696534
#> 
#> [[1]][[7]]
#>         [,1]     [,2]      [,3]
#> 7 0.08812214 0.256731 0.1887149
#> 
#> [[1]][[8]]
#>        [,1]     [,2]     [,3]
#> 8 0.9442012 0.240037 0.359011
#> 
#> [[1]][[9]]
#>        [,1]       [,2]       [,3]
#> 9 0.2203272 0.02253865 0.03724274
#> 
#> [[1]][[10]]
#>         [,1]       [,2]       [,3]
#> 10 0.6034613 0.05144812 0.01893424
#> 
#> [[1]][[11]]
#>         [,1]      [,2]      [,3]
#> 11 0.8275623 0.4409475 0.7467284
#> 
#> [[1]][[12]]
#>          [,1]         [,2]     [,3]
#> 12 0.03180524 1.455262e-05 4.794104
#> 
#> 

kfolds2Chisqind(bbb2)

#> [[1]]
#> [[1]][[1]]
#> [1] 25.402740  5.820043  6.644776
#> 
#> [[1]][[2]]
#> [1] 17.1925045  0.9356285  2.2803489
#> 
#> [[1]][[3]]
#> [1] 5.447630 3.918006 7.854290
#> 
#> [[1]][[4]]
#> [1] 0.21442751 0.02857876 4.61573978
#> 
#> [[1]][[5]]
#> [1] 2.5664845 0.2299003 0.2902721
#> 
#> 

rm(list=c("XCornell","yCornell","bbb","bbb2"))

data(pine)
Xpine<-pine[,1:10]
ypine<-pine[,11]
bbb <- cv.plsRglm(dataY=ypine,dataX=Xpine,nt=4,modele="pls-glm-gaussian",verbose=FALSE)
bbb2 <- cv.plsRglm(dataY=ypine,dataX=Xpine,nt=10,modele="pls-glm-gaussian",K=10,verbose=FALSE)
kfolds2Chisqind(bbb)

#> [[1]]
#> [[1]][[1]]
#> [1] 9.076566 7.246153 6.714280 6.591296
#> 
#> [[1]][[2]]
#> [1] 1.776915 2.961976 1.911164 1.965448
#> 
#> [[1]][[3]]
#> [1] 1.008445 1.311562 1.368349 1.412529
#> 
#> [[1]][[4]]
#> [1] 1.712765 2.017014 5.474255 5.793125
#> 
#> [[1]][[5]]
#> [1] 4.636246 2.763710 2.735360 2.499107
#> 
#> 

kfolds2Chisqind(bbb2)

#> [[1]]
#> [[1]][[1]]
#>  [1] 4.645345 4.201920 3.816200 3.709182 3.541944 3.597604 3.664854 3.731275
#>  [9] 3.741750 3.742149
#> 
#> [[1]][[2]]
#>  [1] 1.2565105 0.7710923 0.4896430 0.6556930 0.6454540 0.9444134 0.9728650
#>  [8] 0.9642343 0.9723018 0.9695092
#> 
#> [[1]][[3]]
#>  [1] 0.8904213 1.1285529 1.1088820 1.0149235 1.0876787 1.0985064 1.1282128
#>  [8] 1.1514029 1.1814023 1.1813702
#> 
#> [[1]][[4]]
#>  [1] 1.267388 1.575034 1.329759 1.306605 1.406189 1.495547 1.421957 1.462479
#>  [9] 1.479878 1.483052
#> 
#> [[1]][[5]]
#>  [1] 1.355488 1.295406 1.156780 1.606341 1.736037 1.724438 1.904159 1.936650
#>  [9] 1.943547 1.943207
#> 
#> [[1]][[6]]
#>  [1] 1.0612597 1.0382611 1.1029085 1.0634485 1.0580194 1.1175511 1.0523920
#>  [8] 0.9463867 0.9270626 0.9301250
#> 
#> [[1]][[7]]
#>  [1] 0.54280592 0.25754355 0.06860201 0.05940245 0.31971926 0.32324978
#>  [7] 0.26009635 0.27489032 0.27287222 0.27135352
#> 
#> [[1]][[8]]
#>  [1] 0.6469443 0.5445691 0.3817880 0.4267373 0.4517251 0.5006288 0.3946705
#>  [8] 0.4114368 0.3860094 0.3860218
#> 
#> [[1]][[9]]
#>  [1] 0.3773657 1.6797105 0.7778450 0.7129019 0.8835625 0.7723416 0.6181708
#>  [8] 0.6120069 0.5225405 0.6042908
#> 
#> [[1]][[10]]
#>  [1] 1.1400873 0.5782617 0.4790428 0.5246598 0.4502027 0.9944149 1.2949425
#>  [8] 1.3074861 1.2859273 1.2832650
#> 
#> 

                  
XpineNAX21 <- Xpine
XpineNAX21[1,2] <- NA
bbbNA <- cv.plsRglm(dataY=ypine,dataX=XpineNAX21,nt=10,modele="pls",K=10,verbose=FALSE)
kfolds2Pressind(bbbNA)

#> [[1]]
#> [[1]][[1]]
#> [1] 1.1797769 0.5858419 0.7146959 0.2823875 0.2113418 0.2462732 0.2128635
#> [8] 0.2237419 0.2245836
#> 
#> [[1]][[2]]
#> [1] 1.7714765 1.1916695 1.2908022 0.9277364 0.7952702 5.1897070 9.2858480
#> [8] 2.9019373 0.6866543
#> 
#> [[1]][[3]]
#> [1] 1.2615115 1.5098867 1.3821292 0.7498655 0.7123033 0.6912226 0.6248337
#> [8] 0.8118411 2.3275452
#> 
#> [[1]][[4]]
#> [1] 4.250364 3.703461 3.660875 3.344024 3.400661 3.210164 3.411373 3.371408
#> [9] 3.607807
#> 
#> [[1]][[5]]
#> [1] 0.3982423 0.7944822 0.6634077 0.7371331 0.7285453 0.7173490 0.7310973
#> [8] 0.7071575 0.7143203
#> 
#> [[1]][[6]]
#> [1] 0.51284570 0.51287606 0.12125609 0.10928749 0.08975948 0.10882349 0.14498431
#> [8] 0.08883830 0.10995633
#> 
#> [[1]][[7]]
#> [1] 1.5613272 0.9164189 0.9277530 1.1503787 1.2133997 1.0775861 1.3403639
#> [8] 1.2964642 1.3117802
#> 
#> [[1]][[8]]
#> [1] 0.4820281 1.8779051 1.9209835 1.7895404 2.0025308 1.9319570 1.9114104
#> [8] 2.0338892 2.1002124
#> 
#> [[1]][[9]]
#> [1] 0.2629415 0.5453534 0.6879429 0.7995751 0.7201941 0.8201471 1.0512128
#> [8] 1.1159823 1.1847008
#> 
#> [[1]][[10]]
#> [1] 1.1153310 1.0859986 1.1172589 1.0155093 1.0257271 0.9239704 1.0288646
#> [8] 1.0395270 1.2891789
#> 
#> 

kfolds2Chisqind(bbbNA)

#> [[1]]
#> [[1]][[1]]
#> [1] 1.1797769 0.5858419 0.7146959 0.2823875 0.2113418 0.2462732 0.2128635
#> [8] 0.2237419 0.2245836
#> 
#> [[1]][[2]]
#> [1] 1.7714765 1.1916695 1.2908022 0.9277364 0.7952702 5.1897070 9.2858480
#> [8] 2.9019373 0.6866543
#> 
#> [[1]][[3]]
#> [1] 1.2615115 1.5098867 1.3821292 0.7498655 0.7123033 0.6912226 0.6248337
#> [8] 0.8118411 2.3275452
#> 
#> [[1]][[4]]
#> [1] 4.250364 3.703461 3.660875 3.344024 3.400661 3.210164 3.411373 3.371408
#> [9] 3.607807
#> 
#> [[1]][[5]]
#> [1] 0.3982423 0.7944822 0.6634077 0.7371331 0.7285453 0.7173490 0.7310973
#> [8] 0.7071575 0.7143203
#> 
#> [[1]][[6]]
#> [1] 0.51284570 0.51287606 0.12125609 0.10928749 0.08975948 0.10882349 0.14498431
#> [8] 0.08883830 0.10995633
#> 
#> [[1]][[7]]
#> [1] 1.5613272 0.9164189 0.9277530 1.1503787 1.2133997 1.0775861 1.3403639
#> [8] 1.2964642 1.3117802
#> 
#> [[1]][[8]]
#> [1] 0.4820281 1.8779051 1.9209835 1.7895404 2.0025308 1.9319570 1.9114104
#> [8] 2.0338892 2.1002124
#> 
#> [[1]][[9]]
#> [1] 0.2629415 0.5453534 0.6879429 0.7995751 0.7201941 0.8201471 1.0512128
#> [8] 1.1159823 1.1847008
#> 
#> [[1]][[10]]
#> [1] 1.1153310 1.0859986 1.1172589 1.0155093 1.0257271 0.9239704 1.0288646
#> [8] 1.0395270 1.2891789
#> 
#> 

bbbNA2 <- cv.plsRglm(dataY=ypine,dataX=XpineNAX21,nt=4,modele="pls-glm-gaussian",verbose=FALSE)
bbbNA3 <- cv.plsRglm(dataY=ypine,dataX=XpineNAX21,nt=10,modele="pls-glm-gaussian",
K=10,verbose=FALSE)
kfolds2Chisqind(bbbNA2)

#> [[1]]
#> [[1]][[1]]
#> [1] 2.037958 1.723634 1.346364 1.365178
#> 
#> [[1]][[2]]
#> [1] 3.862256 5.530433 4.822656 4.067779
#> 
#> [[1]][[3]]
#> [1] 4.265301 2.863907 2.704673 2.538554
#> 
#> [[1]][[4]]
#> [1] 3.496627 4.761206 4.794644 4.773269
#> 
#> [[1]][[5]]
#> [1] 0.7557832 0.4917096 0.5170601 0.3566428
#> 
#> 

kfolds2Chisqind(bbbNA3)

#> [[1]]
#> [[1]][[1]]
#> [1] 2.636308 2.353663 2.214089 2.169882 1.915847 1.957653 2.034536 2.105969
#> [9] 6.877071
#> 
#> [[1]][[2]]
#> [1] 2.0281540 1.9176917 0.5571222 0.4958649 0.3789524 0.4262533 0.4049143
#> [8] 0.5024596 0.9286938
#> 
#> [[1]][[3]]
#> [1] 3.010927 3.563619 4.087904 3.674775 3.472211 4.141086 3.939035 3.845466
#> [9] 8.649060
#> 
#> [[1]][[4]]
#> [1] 0.5273587 0.2465437 0.3209524 0.3115761 0.4711832 0.4749724 0.4749858
#> [8] 0.4831043 0.5562632
#> 
#> [[1]][[5]]
#> [1] 0.3710171 0.3728147 0.2859800 0.2427988 0.2297118 0.2567593 0.3977688
#> [8] 0.2630301 0.5135947
#> 
#> [[1]][[6]]
#> [1] 0.4282734 0.5078277 0.5753935 0.5360069 0.7438233 1.1578475 1.1723711
#> [8] 0.8572551 0.9247138
#> 
#> [[1]][[7]]
#> [1] 0.6695759 1.3841260 1.5775366 2.2501972 2.4848670 2.3460409 2.4245118
#> [8] 2.3664219 2.3933889
#> 
#> [[1]][[8]]
#> [1] 1.937844 1.908356 1.482378 1.425287 1.245614 1.377930 1.308169 1.192801
#> [9] 1.045180
#> 
#> [[1]][[9]]
#> [1] 0.2766676 0.7283884 0.4910580 0.2809118 0.2326252 0.1993401 0.1947887
#> [8] 0.2143691 0.1759972
#> 
#> [[1]][[10]]
#> [1]  1.882691  3.839022 10.947253 11.729232  6.059289  4.263496  5.206835
#> [8]  5.000930  4.833011
#> 
#> 

rm(list=c("Xpine","XpineNAX21","ypine","bbb","bbb2","bbbNA","bbbNA2","bbbNA3"))

data(aze_compl)
Xaze_compl<-aze_compl[,2:34]
yaze_compl<-aze_compl$y
kfolds2Chisqind(cv.plsRglm(dataY=yaze_compl,dataX=Xaze_compl,nt=4,modele="pls-glm-family",
family=binomial(),verbose=FALSE))

#> [[1]]
#> [[1]][[1]]
#> [1]    85.87972   415.61815  9552.73771 50146.28867
#> 
#> [[1]][[2]]
#> [1]  30.43847  58.93485 135.11505 697.47163
#> 
#> [[1]][[3]]
#> [1]  84.70229  85.19330 167.77184 407.30071
#> 
#> [[1]][[4]]
#> [1]     76.08916    320.26726  10845.67028 194032.60797
#> 
#> [[1]][[5]]
#> [1] 32.42490 33.98568 32.16911 71.47877
#> 
#> 

kfolds2Chisqind(cv.plsRglm(dataY=yaze_compl,dataX=Xaze_compl,nt=4,modele="pls-glm-logistic",
verbose=FALSE))

#> [[1]]
#> [[1]][[1]]
#> [1]  29.90166  35.35482  61.71210 254.29400
#> 
#> [[1]][[2]]
#> [1]    38.73265    98.94456  2791.98078 48233.47095
#> 
#> [[1]][[3]]
#> [1] 15.89236 13.82798 15.16932 17.65274
#> 
#> [[1]][[4]]
#> [1] 1.069566e+02 1.561990e+04 2.397108e+06 3.591359e+09
#> 
#> [[1]][[5]]
#> [1] 25.76820 26.19004 48.08101 55.80449
#> 
#> 

kfolds2Chisqind(cv.plsRglm(dataY=yaze_compl,dataX=Xaze_compl,nt=10,modele="pls-glm-family",
family=binomial(),K=10,verbose=FALSE))

#> [[1]]
#> [[1]][[1]]
#>  [1] 21.28649 26.78417 28.73733 26.83481 23.90075 21.68535 19.90691 18.78040
#>  [9] 18.29077 19.23925
#> 
#> [[1]][[2]]
#>  [1]  9.996457  9.927373 12.801550 15.388611 19.344144 23.470476 34.211336
#>  [8] 44.979514 58.584141 71.313792
#> 
#> [[1]][[3]]
#>  [1]  21.51415  33.41371 145.58429 361.10028 532.26056 726.73285 783.16952
#>  [8] 785.51028 768.23774 788.03680
#> 
#> [[1]][[4]]
#>  [1] 13.674041  8.917831 11.978234 12.957416 14.704546 18.606187 18.244042
#>  [8] 18.495280 20.292461 21.225835
#> 
#> [[1]][[5]]
#>  [1] 14.02352 17.38584 17.81248 27.78589 33.11865 31.49402 42.89125 41.49750
#>  [9] 41.73666 42.64957
#> 
#> [[1]][[6]]
#>  [1]     41.24855    123.54772   1751.34162   4691.79878   9126.89319
#>  [6]  24564.08017  50530.16650  89644.79558 155678.27450 170044.42087
#> 
#> [[1]][[7]]
#>  [1]  8.443823  9.349425 10.125910 10.906037 10.935151 12.187017 11.610283
#>  [8] 11.842116 12.348539 12.479056
#> 
#> [[1]][[8]]
#>  [1] 3.025017e+01 8.784145e+01 2.580545e+03 1.680360e+04 3.518843e+04
#>  [6] 7.356849e+04 9.710369e+05 1.235870e+07 4.741217e+07 2.571523e+08
#> 
#> [[1]][[9]]
#>  [1] 21.70293 30.28251 38.46878 59.90450 37.31227 31.69412 29.88543 27.72571
#>  [9] 27.40573 28.16437
#> 
#> [[1]][[10]]
#>  [1]  16.37463  29.21695  23.21907  40.23902  47.94648  63.44231 142.12963
#>  [8] 227.62636 265.24474 289.74260
#> 
#> 

kfolds2Chisqind(cv.plsRglm(dataY=yaze_compl,dataX=Xaze_compl,nt=10,
modele="pls-glm-logistic",K=10,verbose=FALSE))

#> [[1]]
#> [[1]][[1]]
#>  [1]  15.80978  44.05306  82.36316 151.17129 265.18977 297.38013 396.12726
#>  [8] 421.15538 444.57173 479.31396
#> 
#> [[1]][[2]]
#>  [1]  26.72293  33.98994  39.78568  71.41097 103.56428 203.20043 422.83932
#>  [8] 502.50340 420.96670 366.37898
#> 
#> [[1]][[3]]
#>  [1] 15.77152 22.72072 14.26971 14.47100 15.25765 18.33798 20.82254 22.50139
#>  [9] 23.86808 23.29184
#> 
#> [[1]][[4]]
#>  [1] 30.99683 35.90686 58.37319 66.24709 46.23476 74.42932 95.70309 94.14539
#>  [9] 91.88929 87.30195
#> 
#> [[1]][[5]]
#>  [1]  10.32360  19.91479  66.90954  65.77721 112.34611 113.68440  86.81271
#>  [8]  75.54479  75.90657  74.58275
#> 
#> [[1]][[6]]
#>  [1]  9.508019  8.632920 14.858834 20.681227 20.345860 37.478821 40.425672
#>  [8] 50.455355 58.890467 58.900185
#> 
#> [[1]][[7]]
#>  [1] 13.56636 13.03348 15.50174 22.37470 32.34973 42.18472 50.69748 61.11245
#>  [9] 60.64369 59.41205
#> 
#> [[1]][[8]]
#>  [1]   43.57178   93.66793 1047.10021 2391.81588 3269.51445 2844.76774
#>  [7] 2775.94551 3477.23768 4685.45692 4949.90941
#> 
#> [[1]][[9]]
#>  [1]  9.407435 10.370350 11.100518 10.720953 11.055086  6.628029  5.362692
#>  [8]  5.634016  5.047590  4.744585
#> 
#> [[1]][[10]]
#>  [1]     53.72128    179.93889   3018.80492  16979.48469  38974.65165
#>  [6]  89023.50572 173627.23970 288675.34724 351847.54392 398071.50416
#> 
#> 

rm(list=c("Xaze_compl","yaze_compl"))
# }