Computes PRESS for k-fold cross validated partial least squares regression models.

This function computes PRESS for k-fold cross validated partial least squares regression models.

kfolds2Press(pls_kfolds)

Arguments

pls_kfolds
a k-fold cross validated partial least squares regression model

Value

list

Press vs number of components for the first group partition

…

…

list

Press vs number of components for the last group partition

References

Nicolas Meyer, Myriam Maumy-Bertrand et Frederic Bertrand (2010). Comparing the linear and the logistic PLS regression with qualitative predictors: application to allelotyping data. Journal de la Societe Francaise de Statistique, 151(2), pages 1-18. http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/47

Note

Use cv.plsR to create k-fold cross validated partial least squares regression models.

See also

Examples

data(Cornell)
XCornell<-Cornell[,1:7]
yCornell<-Cornell[,8]
kfolds2Press(cv.plsR(dataY=yCornell,dataX=data.frame(scale(as.matrix(XCornell))[,]),
nt=6,K=12,NK=1,verbose=FALSE))

#> [[1]]
#> [1] 55.70774 41.43274 20.27397 21.24240 24.51801
#> 

kfolds2Press(cv.plsR(dataY=yCornell,dataX=data.frame(scale(as.matrix(XCornell))[,]),
nt=6,K=6,NK=1,verbose=FALSE))

#> [[1]]
#> [1] 59.91409 46.31581 25.53166 13.58373 13.27998
#> 

rm(list=c("XCornell","yCornell"))

data(pine)
Xpine<-pine[,1:10]
ypine<-pine[,11]
kfolds2Press(cv.plsR(dataY=ypine,dataX=Xpine,nt=10,NK=1,verbose=FALSE))

#> [[1]]
#>  [1] 14.19968 12.97876 11.09295 10.40390 11.18671 11.87286 11.91133 12.07779
#>  [9] 12.42402 12.83051
#> 

kfolds2Press(cv.plsR(dataY=ypine,dataX=Xpine,nt=10,NK=2,verbose=FALSE))

#> [[1]]
#>  [1] 13.57909 13.33459 12.91976 11.09299 10.32684 10.44413 10.74837 11.50118
#>  [9] 11.95366 12.28187
#> 
#> [[2]]
#>  [1] 14.12791 14.89530 15.25155 11.41853 11.62845 12.15636 12.11770 11.81713
#>  [9] 11.87533 12.09169
#> 

XpineNAX21 <- Xpine
XpineNAX21[1,2] <- NA
kfolds2Press(cv.plsR(dataY=ypine,dataX=XpineNAX21,nt=10,NK=1,verbose=FALSE))

#> [[1]]
#> [1] 13.39977 16.61331 16.91174 14.00340 13.03811 13.33037 14.41686 16.91070
#> [9] 18.07653
#> 

kfolds2Press(cv.plsR(dataY=ypine,dataX=XpineNAX21,nt=10,NK=2,verbose=FALSE))

#> [[1]]
#> [1] 12.933603 11.527878 14.258282 10.697099 12.212251  9.837673 12.883893
#> [8] 15.197311 13.506709
#> 
#> [[2]]
#> [1] 13.98605 16.38513 18.12343 12.72618 15.62679 16.13776 16.57595 16.66593
#> [9] 18.84447
#> 

rm(list=c("Xpine","XpineNAX21","ypine"))