Computes PRESS for k-fold cross validated partial least squares regression models.

This function computes PRESS for k-fold cross validated partial least squares regression models.

kfolds2Press(pls_kfolds)

Arguments

pls_kfolds
a k-fold cross validated partial least squares regression model

Value

list

Press vs number of components for the first group partition

list()

...

list

Press vs number of components for the last group partition

Note

Use cv.plsR to create k-fold cross validated partial least squares regression models.

References

Nicolas Meyer, Myriam Maumy-Bertrand et Frédéric Bertrand (2010). Comparing the linear and the logistic PLS regression with qualitative predictors: application to allelotyping data. Journal de la Societe Francaise de Statistique, 151(2), pages 1-18. http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/47

See also

Author

Frédéric Bertrand
frederic.bertrand@math.unistra.fr
https://fbertran.github.io/homepage/

Examples

data(Cornell)
XCornell<-Cornell[,1:7]
yCornell<-Cornell[,8]
kfolds2Press(cv.plsR(dataY=yCornell,dataX=data.frame(scale(as.matrix(XCornell))[,]),
nt=6,K=12,NK=1,verbose=FALSE))

#> [[1]]
#> [1] 55.70774 41.43274 20.27397 21.24240 24.51801
#> 

kfolds2Press(cv.plsR(dataY=yCornell,dataX=data.frame(scale(as.matrix(XCornell))[,]),
nt=6,K=6,NK=1,verbose=FALSE))

#> [[1]]
#> [1] 59.34058 53.49685 27.05261 18.27445 15.81791
#> 

rm(list=c("XCornell","yCornell"))

# \donttest{
data(pine)
Xpine<-pine[,1:10]
ypine<-pine[,11]
kfolds2Press(cv.plsR(dataY=ypine,dataX=Xpine,nt=10,NK=1,verbose=FALSE))

#> [[1]]
#>  [1] 13.19448 13.42081 12.56074 12.71720 12.67167 13.45020 16.27185 16.20320
#>  [9] 17.29195 18.04110
#> 

kfolds2Press(cv.plsR(dataY=ypine,dataX=Xpine,nt=10,NK=2,verbose=FALSE))

#> [[1]]
#>  [1] 14.22690 13.47760 12.18352 11.45640 12.64730 12.89181 13.30853 13.27108
#>  [9] 13.27575 13.37079
#> 
#> [[2]]
#>  [1] 14.04650 15.97984 16.66492 15.24050 14.44297 18.71912 19.86770 20.99399
#>  [9] 25.47755 27.35883
#> 

XpineNAX21 <- Xpine
XpineNAX21[1,2] <- NA
kfolds2Press(cv.plsR(dataY=ypine,dataX=XpineNAX21,nt=10,NK=1,verbose=FALSE))

#> [[1]]
#> [1] 13.39542 14.99605 21.21602 18.13472 19.59230 61.00299 31.08261 29.21998
#> [9] 20.47509
#> 

kfolds2Press(cv.plsR(dataY=ypine,dataX=XpineNAX21,nt=10,NK=2,verbose=FALSE))

#> [[1]]
#> [1] 16.78181 16.37744 19.22876 15.35242 16.75700 13.70953 13.93246 16.11035
#> [9] 16.25827
#> 
#> [[2]]
#> [1] 13.68856 16.33892 15.91760 14.24768 16.34976 14.37112 15.87366 14.89092
#> [9] 44.76294
#> 

rm(list=c("Xpine","XpineNAX21","ypine"))
# }