Computes individual PRESS for k-fold cross validated partial least squares regression models.

This function computes individual PRESS for k-fold cross validated partial least squares regression models.

kfolds2Pressind(pls_kfolds)

Arguments

pls_kfolds
a k-fold cross validated partial least squares regression model

Value

list

Individual Press vs number of components for the first group partition

…

…

list

Individual Press vs number of components for the last group partition

References

Nicolas Meyer, Myriam Maumy-Bertrand et Frederic Bertrand (2010). Comparing the linear and the logistic PLS regression with qualitative predictors: application to allelotyping data. Journal de la Societe Francaise de Statistique, 151(2), pages 1-18. http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/47

Note

Use cv.plsR to create k-fold cross validated partial least squares regression models.

See also

Examples

data(Cornell)
XCornell<-Cornell[,1:7]
yCornell<-Cornell[,8]
kfolds2Pressind(cv.plsR(dataY=yCornell,dataX=data.frame(scale(as.matrix(XCornell))[,]),
nt=6,K=12,NK=1))

#> NK: 1 
#> Leave One Out
#> Number of groups : 12 
#> 1 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 2 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 3 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 4 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 5 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 6 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 7 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 8 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 9 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 10 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 11 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> Warning : 1 2 3 4 5 6 7 < 10^{-12}
#> Warning only 5 components could thus be extracted
#> ****________________________________________________****
#> 
#> 12 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 

#> [[1]]
#> [[1]][[1]]
#>       [,1]     [,2]     [,3]     [,4]     [,5]     [,6]
#> 1 24.52456 19.46751 9.132702 6.212796 6.260403 6.294375
#> 
#> [[1]][[2]]
#>       [,1]      [,2]      [,3]      [,4]     [,5]      [,6]
#> 2 4.124502 0.6664187 0.4453306 0.0288125 0.468846 0.6001117
#> 
#> [[1]][[3]]
#>       [,1]      [,2]     [,3]     [,4]     [,5]     [,6]
#> 3 1.551301 0.9608429 1.224441 2.534867 1.991568 5.504755
#> 
#> [[1]][[4]]
#>       [,1]     [,2]        [,3]     [,4]       [,5]       [,6]
#> 4 12.26717 0.247437 0.001516247 0.111019 0.06429773 0.05142279
#> 
#> [[1]][[5]]
#>       [,1]     [,2]     [,3]     [,4]     [,5]    [,6]
#> 5 4.315406 9.872769 2.420394 4.187705 8.869722 8.95432
#> 
#> [[1]][[6]]
#>       [,1]     [,2]     [,3]     [,4]     [,5]     [,6]
#> 6 6.209332 5.820357 3.366519 1.798485 0.950924 1.959007
#> 
#> [[1]][[7]]
#>         [,1]      [,2]      [,3]      [,4]       [,5]        [,6]
#> 7 0.08812214 0.4818875 0.2496785 0.2425807 0.00951357 0.008071516
#> 
#> [[1]][[8]]
#>        [,1]      [,2]      [,3]     [,4]     [,5]      [,6]
#> 8 0.9442012 0.5163166 0.4084427 0.351376 0.248757 0.2347731
#> 
#> [[1]][[9]]
#>        [,1]       [,2]       [,3]       [,4]      [,5]      [,6]
#> 9 0.2203272 0.01149677 0.02486984 0.02827224 0.1233742 0.1658223
#> 
#> [[1]][[10]]
#>         [,1]         [,2]        [,3]       [,4]        [,5]       [,6]
#> 10 0.6034613 0.0005521844 0.006299065 0.01764824 0.006014414 0.02841849
#> 
#> [[1]][[11]]
#>         [,1]       [,2]      [,3]      [,4]       [,5]
#> 11 0.8275623 0.01285346 0.6440788 0.7149959 0.06577128
#> 
#> [[1]][[12]]
#>          [,1]     [,2]     [,3]     [,4]     [,5]    [,6]
#> 12 0.03180524 3.374303 2.349701 5.013846 5.458814 8.13772
#> 
#> 

kfolds2Pressind(cv.plsR(dataY=yCornell,dataX=data.frame(scale(as.matrix(XCornell))[,]),
nt=6,K=6,NK=1))

#> NK: 1 
#> Number of groups : 6 
#> 1 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> Warning : 1 2 3 4 5 6 7 < 10^{-12}
#> Warning only 5 components could thus be extracted
#> ****________________________________________________****
#> 
#> 2 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 3 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 4 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 5 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 6 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 

#> [[1]]
#> [[1]][[1]]
#> [1] 0.882114006 0.001484329 0.914517981 1.016580217 0.146304708
#> 
#> [[1]][[2]]
#> [1] 7.111137 7.668673 4.541319 2.662954 1.596324 2.289553
#> 
#> [[1]][[3]]
#> [1] 31.564554 19.452709  6.403846  5.518035  5.920875  5.616452
#> 
#> [[1]][[4]]
#> [1] 4.464202 9.455269 2.973374 4.220288 9.280723 9.513030
#> 
#> [[1]][[5]]
#> [1] 12.3290316  0.4061359  0.2343948  0.3601167  0.0596658  0.0477965
#> 
#> [[1]][[6]]
#> [1]  4.597799  5.142740  4.371768  6.750691  5.412892 11.637487
#> 
#> 

rm(list=c("XCornell","yCornell"))

data(pine)
Xpine<-pine[,1:10]
ypine<-pine[,11]
kfolds2Pressind(cv.plsR(dataY=ypine,dataX=Xpine,nt=10,NK=1,verbose=FALSE))

#> [[1]]
#> [[1]][[1]]
#>  [1] 4.805816 4.690943 4.993705 4.011261 3.951879 3.916437 3.986039 3.990145
#>  [9] 3.950266 4.045699
#> 
#> [[1]][[2]]
#>  [1] 1.8084439 1.9412554 1.2111849 1.0489093 1.0260334 1.0071353 0.9781246
#>  [8] 0.9124108 1.0014601 1.0628614
#> 
#> [[1]][[3]]
#>  [1] 2.493163 2.858602 3.223785 3.416047 4.435912 4.798739 5.417191 5.254533
#>  [9] 5.497323 5.595991
#> 
#> [[1]][[4]]
#>  [1] 3.596135 3.905395 3.005953 3.346551 4.001416 4.133644 3.870351 4.039486
#>  [9] 4.033945 4.202232
#> 
#> [[1]][[5]]
#>  [1] 1.6467807 1.7542957 1.1455846 0.6079961 0.7111074 0.7697873 0.8437279
#>  [8] 0.8476967 0.8284894 0.8130661
#> 
#> 

kfolds2Pressind(cv.plsR(dataY=ypine,dataX=Xpine,nt=10,NK=2,verbose=FALSE))

#> [[1]]
#> [[1]][[1]]
#>  [1] 0.9811567 2.3159980 2.9733802 2.6344193 2.6608810 2.5898501 2.6993531
#>  [8] 2.6738005 2.7752440 2.7677219
#> 
#> [[1]][[2]]
#>  [1] 5.066121 4.125184 3.700941 4.245749 4.159462 3.987031 4.133685 4.403150
#>  [9] 4.306571 4.468146
#> 
#> [[1]][[3]]
#>  [1] 1.658590 2.501918 2.703593 1.729325 1.280784 1.260604 1.133493 1.179030
#>  [9] 1.169278 1.258264
#> 
#> [[1]][[4]]
#>  [1] 2.055124 4.071144 5.817431 5.566881 5.020133 5.005039 4.979671 4.800009
#>  [9] 4.788242 4.775184
#> 
#> [[1]][[5]]
#>  [1] 2.513996 2.241073 1.875390 1.400466 1.166504 1.390077 1.552621 1.541425
#>  [9] 1.713488 1.777663
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#>  [1] 4.727509 4.466977 4.027641 3.541287 3.800219 3.618500 3.584943 3.505449
#>  [9] 3.506168 3.505443
#> 
#> [[2]][[2]]
#>  [1] 5.978367 3.561949 3.634328 2.797638 2.781064 2.626177 2.524380 2.510415
#>  [9] 2.498540 2.517815
#> 
#> [[2]][[3]]
#>  [1] 1.440365 2.191388 2.221539 1.867491 1.419634 1.339136 1.321539 1.266315
#>  [9] 1.252382 1.300260
#> 
#> [[2]][[4]]
#>  [1] 2.618379 2.119877 1.959013 4.517709 5.880311 6.012706 6.513663 6.213229
#>  [9] 6.149792 6.145210
#> 
#> [[2]][[5]]
#>  [1] 0.9237837 1.8486418 1.7231464 1.9553428 1.9208527 1.9601307 2.0398647
#>  [8] 2.0341859 1.9540770 2.0274995
#> 
#> 

XpineNAX21 <- Xpine
XpineNAX21[1,2] <- NA
kfolds2Pressind(cv.plsR(dataY=ypine,dataX=XpineNAX21,nt=10,NK=1,verbose=FALSE))

#> [[1]]
#> [[1]][[1]]
#> [1] 1.825409 2.183560 1.813888 1.858482 1.682541 1.721630 1.936074 1.909898
#> [9] 1.937642
#> 
#> [[1]][[2]]
#> [1] 5.238317 5.720575 4.365564 3.929038 4.079202 4.162020 4.087464 4.160192
#> [9] 5.843943
#> 
#> [[1]][[3]]
#> [1] 3.830618 3.161029 3.008763 3.412751 2.776611 2.569760 3.025062 3.089974
#> [9] 4.492908
#> 
#> [[1]][[4]]
#> [1] 1.492786 1.255128 1.980452 1.721162 1.376743 1.206704 1.184542 1.164871
#> [9] 1.028909
#> 
#> [[1]][[5]]
#> [1]  2.142286  2.309175  3.241802  1.639115  2.992033  5.559584  6.087083
#> [8] 11.394478 11.345456
#> 
#> 

kfolds2Pressind(cv.plsR(dataY=ypine,dataX=XpineNAX21,nt=10,NK=2,verbose=FALSE))

#> [[1]]
#> [[1]][[1]]
#> [1] 1.6450105 1.9474777 1.6710299 1.1816905 0.7976980 0.6669618 0.8316049
#> [8] 1.4862295 6.1417577
#> 
#> [[1]][[2]]
#> [1] 3.993344 3.879889 3.797228 3.241105 3.383010 3.447232 4.091492 4.419485
#> [9] 4.483931
#> 
#> [[1]][[3]]
#> [1]  4.112515  4.992535  5.158821  2.873014  3.487017  9.488345  9.561490
#> [8]  7.489757 14.433294
#> 
#> [[1]][[4]]
#> [1] 0.693896 4.151322 3.767016 4.087873 4.487252 4.366913 4.553327 5.056588
#> [9] 4.971096
#> 
#> [[1]][[5]]
#> [1] 2.2653722 2.4945902 1.8992621 0.9052759 0.8748049 0.9798295 0.7800984
#> [8] 0.8085901 0.7018205
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> [1]  2.9716696  3.2779408  2.7714934  4.5684210  1.0175520  5.0862274  0.8564276
#> [8]  1.5783679 30.6224601
#> 
#> [[2]][[2]]
#> [1] 2.131316 1.841545 1.259944 1.685330 1.575113 1.677672 1.947676 1.876553
#> [9] 2.231767
#> 
#> [[2]][[3]]
#> [1] 0.3947604 0.9360777 1.0832259 0.8888252 0.4922144 0.3847514 0.3219615
#> [8] 0.7465462 0.7449231
#> 
#> [[2]][[4]]
#> [1] 4.838824 4.370849 4.201728 3.610481 3.745642 4.006851 3.976486 4.046207
#> [9] 4.457591
#> 
#> [[2]][[5]]
#> [1] 4.039507 4.562471 4.124907 3.461058 3.934664 4.053847 4.442942 4.288277
#> [9] 4.429484
#> 
#> 

rm(list=c("Xpine","XpineNAX21","ypine"))