Computes individual PRESS for k-fold cross validated partial least squares regression models.

This function computes individual PRESS for k-fold cross validated partial least squares regression models.

kfolds2Pressind(pls_kfolds)

Arguments

pls_kfolds
a k-fold cross validated partial least squares regression model

Value

list

Individual Press vs number of components for the first group partition

list()

...

list

Individual Press vs number of components for the last group partition

Note

Use cv.plsR to create k-fold cross validated partial least squares regression models.

References

Nicolas Meyer, Myriam Maumy-Bertrand et Frédéric Bertrand (2010). Comparing the linear and the logistic PLS regression with qualitative predictors: application to allelotyping data. Journal de la Societe Francaise de Statistique, 151(2), pages 1-18. http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/47

See also

Author

Frédéric Bertrand
frederic.bertrand@math.unistra.fr
https://fbertran.github.io/homepage/

Examples

data(Cornell)
XCornell<-Cornell[,1:7]
yCornell<-Cornell[,8]
kfolds2Pressind(cv.plsR(dataY=yCornell,dataX=data.frame(scale(as.matrix(XCornell))[,]),
nt=6,K=12,NK=1))

#> NK: 1 
#> Leave One Out
#> Number of groups : 12 
#> 1 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 2 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 3 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 4 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 5 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 6 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 7 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 8 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 9 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 10 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 11 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> Warning : 1 2 3 4 5 6 7 < 10^{-12}
#> Warning only 5 components could thus be extracted
#> ****________________________________________________****
#> 
#> 12 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 

#> [[1]]
#> [[1]][[1]]
#>       [,1]     [,2]     [,3]     [,4]     [,5]     [,6]
#> 1 24.52456 19.46751 9.132702 6.212796 6.260403 6.294375
#> 
#> [[1]][[2]]
#>       [,1]      [,2]      [,3]      [,4]     [,5]      [,6]
#> 2 4.124502 0.6664187 0.4453306 0.0288125 0.468846 0.6001117
#> 
#> [[1]][[3]]
#>       [,1]      [,2]     [,3]     [,4]     [,5]     [,6]
#> 3 1.551301 0.9608429 1.224441 2.534867 1.991568 5.504755
#> 
#> [[1]][[4]]
#>       [,1]     [,2]        [,3]     [,4]       [,5]       [,6]
#> 4 12.26717 0.247437 0.001516247 0.111019 0.06429773 0.05142279
#> 
#> [[1]][[5]]
#>       [,1]     [,2]     [,3]     [,4]     [,5]    [,6]
#> 5 4.315406 9.872769 2.420394 4.187705 8.869722 8.95432
#> 
#> [[1]][[6]]
#>       [,1]     [,2]     [,3]     [,4]     [,5]     [,6]
#> 6 6.209332 5.820357 3.366519 1.798485 0.950924 1.959007
#> 
#> [[1]][[7]]
#>         [,1]      [,2]      [,3]      [,4]       [,5]        [,6]
#> 7 0.08812214 0.4818875 0.2496785 0.2425807 0.00951357 0.008071516
#> 
#> [[1]][[8]]
#>        [,1]      [,2]      [,3]     [,4]     [,5]      [,6]
#> 8 0.9442012 0.5163166 0.4084427 0.351376 0.248757 0.2347731
#> 
#> [[1]][[9]]
#>        [,1]       [,2]       [,3]       [,4]      [,5]      [,6]
#> 9 0.2203272 0.01149677 0.02486984 0.02827224 0.1233742 0.1658223
#> 
#> [[1]][[10]]
#>         [,1]         [,2]        [,3]       [,4]        [,5]       [,6]
#> 10 0.6034613 0.0005521844 0.006299065 0.01764824 0.006014414 0.02841849
#> 
#> [[1]][[11]]
#>         [,1]       [,2]      [,3]      [,4]       [,5]
#> 11 0.8275623 0.01285346 0.6440788 0.7149959 0.06577128
#> 
#> [[1]][[12]]
#>          [,1]     [,2]     [,3]     [,4]     [,5]    [,6]
#> 12 0.03180524 3.374303 2.349701 5.013846 5.458814 8.13772
#> 
#> 

kfolds2Pressind(cv.plsR(dataY=yCornell,dataX=data.frame(scale(as.matrix(XCornell))[,]),
nt=6,K=6,NK=1))

#> NK: 1 
#> Number of groups : 6 
#> 1 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 2 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 3 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> Warning : 1 2 3 4 5 6 7 < 10^{-12}
#> Warning only 5 components could thus be extracted
#> ****________________________________________________****
#> 
#> 4 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 5 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 
#> 6 
#> ____************************************************____
#> ____Predicting X without NA neither in X nor in Y____
#> ____Component____ 1 ____
#> ____Component____ 2 ____
#> ____Component____ 3 ____
#> ____Component____ 4 ____
#> ____Component____ 5 ____
#> ____Component____ 6 ____
#> ****________________________________________________****
#> 

#> [[1]]
#> [[1]][[1]]
#> [1] 24.451307 18.523423  9.180100  6.193589  6.229356  6.282449
#> 
#> [[1]][[2]]
#> [1]  1.985502  1.086092  1.251931  2.691045  2.145680 11.668350
#> 
#> [[1]][[3]]
#> [1] 5.447630 8.686841 4.958866 7.695888 9.073500
#> 
#> [[1]][[4]]
#> [1]  4.597799  5.142740  4.371768  6.750691  5.412892 11.637487
#> 
#> [[1]][[5]]
#> [1] 12.7082450  1.0186833  0.2740066  0.4714313  0.2948841  0.2824644
#> 
#> [[1]][[6]]
#> [1] 7.7218433 4.8379862 2.2036361 1.9488054 0.9726818 1.9711785
#> 
#> 

rm(list=c("XCornell","yCornell"))

# \donttest{
data(pine)
Xpine<-pine[,1:10]
ypine<-pine[,11]
kfolds2Pressind(cv.plsR(dataY=ypine,dataX=Xpine,nt=10,NK=1,verbose=FALSE))

#> [[1]]
#> [[1]][[1]]
#>  [1] 0.8883961 4.8591387 4.9988689 5.1918102 4.8617199 4.7855573 4.4809842
#>  [8] 5.0117341 5.4889825 5.6206026
#> 
#> [[1]][[2]]
#>  [1] 2.365995 2.347366 1.650375 1.353222 1.617524 1.745726 1.911746 1.984874
#>  [9] 2.016569 2.008852
#> 
#> [[1]][[3]]
#>  [1]  6.820733  9.108225 10.823430 10.358195  8.314969  8.076445  8.113869
#>  [8]  7.942539  7.745235  7.745303
#> 
#> [[1]][[4]]
#>  [1] 3.754125 3.430876 4.033878 4.524629 4.967467 5.014482 5.481731 7.513506
#>  [9] 7.552379 7.525122
#> 
#> [[1]][[5]]
#>  [1] 1.8350435 2.2161201 2.4919795 1.1405288 0.8246359 0.7909550 0.6955416
#>  [8] 0.9916653 1.0886978 1.1385369
#> 
#> 

kfolds2Pressind(cv.plsR(dataY=ypine,dataX=Xpine,nt=10,NK=2,verbose=FALSE))

#> [[1]]
#> [[1]][[1]]
#>  [1] 1.788244 1.332780 1.098828 1.186483 1.160098 1.343736 1.246863 1.275641
#>  [9] 1.292925 1.300878
#> 
#> [[1]][[2]]
#>  [1] 3.131351 3.894390 3.950035 4.208874 3.943091 3.739850 4.186724 4.365240
#>  [9] 4.445756 4.441714
#> 
#> [[1]][[3]]
#>  [1] 4.112767 3.327255 4.115770 3.455209 3.065263 3.090385 2.963401 2.821507
#>  [9] 2.771189 2.837435
#> 
#> [[1]][[4]]
#>  [1] 1.4680092 1.9132824 1.5556197 0.9806226 0.8331973 0.7394623 0.6705268
#>  [8] 0.6757110 0.7041166 0.7180339
#> 
#> [[1]][[5]]
#>  [1] 2.785427 2.610470 1.721480 4.292415 6.060435 6.606396 6.232753 6.685529
#>  [9] 6.653602 6.539093
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#>  [1] 5.059471 4.715052 4.119547 4.647800 4.305667 4.209123 4.227596 4.496081
#>  [9] 4.478542 4.479241
#> 
#> [[2]][[2]]
#>  [1] 0.9359804 3.2277898 3.7710491 3.1979426 3.4698565 3.5108730 3.3465489
#>  [8] 3.5111109 3.8270929 3.9603972
#> 
#> [[2]][[3]]
#>  [1] 1.733402 1.580033 1.611175 1.890218 2.024283 2.060701 4.230665 6.502740
#>  [9] 6.538242 8.695176
#> 
#> [[2]][[4]]
#>  [1] 3.651936 2.412445 2.849109 2.479367 2.183804 1.922960 2.269423 2.249847
#>  [9] 2.178460 2.174420
#> 
#> [[2]][[5]]
#>  [1] 2.479865 2.426439 2.745278 2.746920 2.237868 2.125045 2.090085 2.029537
#>  [9] 1.989560 2.010448
#> 
#> 

XpineNAX21 <- Xpine
XpineNAX21[1,2] <- NA
kfolds2Pressind(cv.plsR(dataY=ypine,dataX=XpineNAX21,nt=10,NK=1,verbose=FALSE))

#> [[1]]
#> [[1]][[1]]
#> [1] 4.265068 4.132190 3.481657 2.681060 2.891796 3.069272 3.075046 3.073725
#> [9] 3.322228
#> 
#> [[1]][[2]]
#> [1] 5.541811 4.582825 4.246572 3.464987 3.687461 3.276194 3.449367 6.283283
#> [9] 8.771922
#> 
#> [[1]][[3]]
#> [1] 2.042533 2.801822 3.008664 3.176387 3.590430 3.942503 3.623385 3.530426
#> [9] 3.585195
#> 
#> [[1]][[4]]
#> [1] 3.391693 3.374958 3.869096 3.116810 2.687136 2.673019 2.671724 2.843307
#> [9] 2.540961
#> 
#> [[1]][[5]]
#> [1]  2.462699  1.417479  3.483799  2.222543 49.820372  1.650904  2.766453
#> [8]  1.093433  2.597486
#> 
#> 

kfolds2Pressind(cv.plsR(dataY=ypine,dataX=XpineNAX21,nt=10,NK=2,verbose=FALSE))

#> [[1]]
#> [[1]][[1]]
#> [1] 2.984719 2.439020 2.463254 5.041728 6.950408 6.919561 7.586048 7.251818
#> [9] 7.239121
#> 
#> [[1]][[2]]
#> [1] 1.325304 3.242733 4.699757 4.492495 3.344519 3.233189 3.188770 3.630489
#> [9] 5.184970
#> 
#> [[1]][[3]]
#> [1] 1.181561 2.298805 2.069457 2.334230 2.462473 2.353147 2.312817 2.486078
#> [9] 2.828056
#> 
#> [[1]][[4]]
#> [1]  5.656907  4.773517  4.371704  4.711435  4.631337  5.195068 10.307834
#> [8]  8.043123 83.859646
#> 
#> [[1]][[5]]
#> [1] 1.872654 1.793273 1.485903 1.433583 1.714767 2.197102 2.547036 2.526357
#> [9] 2.917841
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> [1] 0.9625401 2.9556566 4.6782982 4.8384801 4.5644209 3.9823354 3.5711189
#> [8] 4.5653789 4.8712847
#> 
#> [[2]][[2]]
#> [1] 2.489185 2.819143 2.938148 3.594480 4.727433 4.921556 5.712592 5.491418
#> [9] 3.949971
#> 
#> [[2]][[3]]
#> [1] 1.794697 2.129860 1.503509 1.389300 1.338485 1.231822 1.126318 1.093476
#> [9] 1.077735
#> 
#> [[2]][[4]]
#> [1]  4.589280  3.917830  3.564259  2.275041  5.377915  2.867677  4.752001
#> [8] 66.091534  2.517110
#> 
#> [[2]][[5]]
#> [1] 5.738700 5.360488 4.475283 4.605362 4.553123 5.010755 4.886292 4.927584
#> [9] 4.583930
#> 
#> 

rm(list=c("Xpine","XpineNAX21","ypine"))
# }