Data generating function for multivariate plsR models

This function generates a single multivariate response value \(\boldsymbol{Y}\) and a vector of explinatory variables \((X_1,\ldots,X_{totdim})\) drawn from a model with a given number of latent components.

simul_data_YX(totdim, ncomp)

Arguments

totdim
Number of column of the X vector (from ncomp to hardware limits)

ncomp
Number of latent components in the model (from 2 to 6)

Details

This function should be combined with the replicate function to give rise to a larger dataset. The algorithm used is a R port of the one described in the article of Li which is a multivariate generalization of the algorithm of Naes and Martens.

Value

vector

\((Y_1,\ldots,Y_r,X_1,\ldots,X_{totdim})\)

References

T. Naes, H. Martens, Comparison of prediction methods for multicollinear data, Commun. Stat., Simul. 14 (1985) 545-576.
Baibing Li, Julian Morris, Elaine B. Martin, Model selection for partial least squares regression, Chemometrics and Intelligent Laboratory Systems 64 (2002) 79-89.
http://dx.doi.org/10.1016/S0169-7439(02)00051-5

Note

The value of \(r\) depends on the value of ncomp :

ncomp

\(r\)

2	3

3	3

4	4

ncomp

See also

Examples

simul_data_YX(20,6)

#>         Y 1         Y 2         Y 3         Y 4          X1          X2 
#>  8.21209590  8.21445036  8.21577281  8.22736726  4.01324727  1.38568594 
#>          X3          X4          X5          X6          X7          X8 
#>  0.06505756 -0.46383510  3.33902955  0.42505828 -0.58114336 -1.44627523 
#>          X9         X10         X11         X12         X13         X14 
#> -0.64584263  2.07134264  2.63878433 -1.43669864  1.62261735  0.78481810 
#>         X15         X16         X17         X18         X19         X20 
#> -0.63016117  2.05820761  2.63747459 -1.46768420  1.61442723  0.78660344 

if(require(plsdepot)){
dimX <- 6
Astar <- 2
(dataAstar2 <- t(replicate(50,simul_data_YX(dimX,Astar))))
library(plsdepot)
resAstar2 <- plsreg2(dataAstar2[,4:9],dataAstar2[,1:3],comps=5)
resAstar2$Q2
resAstar2$Q2[,4]>0.0975

dimX <- 6
Astar <- 3
(dataAstar3 <- t(replicate(50,simul_data_YX(dimX,Astar))))
library(plsdepot)
resAstar3 <- plsreg2(dataAstar3[,4:9],dataAstar3[,1:3],comps=5)
resAstar3$Q2
resAstar3$Q2[,4]>0.0975

dimX <- 6
Astar <- 4
(dataAstar4 <- t(replicate(50,simul_data_YX(dimX,Astar))))
library(plsdepot)
resAstar4 <- plsreg2(dataAstar4[,5:10],dataAstar4[,1:4],comps=5)
resAstar4$Q2
resAstar4$Q2[,5]>0.0975

rm(list=c("dimX","Astar"))
}

#> Loading required package: plsdepot