Data generating detailed process for multivariate plsR models

This function generates a single multivariate response value \(\boldsymbol{Y}\) and a vector of explinatory variables \((X_1,\ldots,X_{totdim})\) drawn from a model with a given number of latent components.

simul_data_complete(totdim, ncomp)

Arguments

totdim
Number of columns of the X vector (from ncomp to hardware limits)

ncomp
Number of latent components in the model (from 2 to 6)

Details

This function should be combined with the replicate function to give rise to a larger dataset. The algorithm used is a R port of the one described in the article of Li which is a multivariate generalization of the algorithm of Naes and Martens.

Value

simX

Vector of explanatory variables

HH

Dimension of the response \(\boldsymbol{Y}\)

eta

See Li et al.

r

See Li et al.

epsilon

See Li et al.

ksi

See Li et al.

f

See Li et al.

z

See Li et al.

Y

See Li et al.

References

T. Naes, H. Martens, Comparison of prediction methods for multicollinear data, Commun. Stat., Simul. 14 (1985) 545-576.
Baibing Li, Julian Morris, Elaine B. Martin, Model selection for partial least squares regression, Chemometrics and Intelligent Laboratory Systems 64 (2002) 79-89.
http://dx.doi.org/10.1016/S0169-7439(02)00051-5

Note

The value of \(r\) depends on the value of ncomp :

ncomp

\(r\)

2

3

3	3

4

4

See also

Examples

simul_data_complete(20,6)

#> $simX
#>          [,1]       [,2]     [,3]     [,4]      [,5]      [,6]     [,7]
#> [1,] -2.93587 -0.3950476 4.846287 1.277165 -2.287429 0.5385424 5.446381
#>          [,8]     [,9]     [,10]      [,11]    [,12]      [,13]    [,14]
#> [1,] 2.211102 1.928477 -2.610418 -0.6244556 4.747093 -0.1235033 3.166976
#>         [,15]     [,16]      [,17]    [,18]     [,19]    [,20]
#> [1,] 1.942805 -2.589161 -0.6416808 4.754974 -0.127067 3.159448
#> 
#> $HH
#> [1] 4
#> 
#> $eta
#>      eta61 eta62 eta63 eta64 eta65 eta66
#> [1,]   0.5   0.5   0.5   0.5   0.5   0.5
#> [2,]   0.5   0.5   0.5   0.5   0.5   0.5
#> [3,]   0.5   0.5   0.5   0.5   0.5   0.5
#> [4,]   0.5   0.5   0.5   0.5   0.5   0.5
#> 
#> $r
#> [1]  4.8410080  0.6885465 -9.4116857 -6.7910702 -1.4552003  0.3076811
#> 
#> $epsilon
#>  [1] -0.0189252930  0.0067228071  0.0203302873  0.0101516453  0.0055198584
#>  [6]  0.0086348888 -0.0035718511  0.0124101765 -0.0139251513 -0.0008181468
#> [11]  0.0164883859  0.0027096706 -0.0082031234 -0.0069764585  0.0004029137
#> [16]  0.0204390573 -0.0007367950  0.0105905240 -0.0117668119 -0.0145038788
#> 
#> $ksi
#>            ksi1       ksi2  ksi3       ksi4       ksi5  ksi6
#>  [1,] 0.2236068  0.2672612  0.25  0.2236068  0.2672612  0.25
#>  [2,] 0.2236068 -0.2672612  0.25 -0.2236068  0.2672612 -0.25
#>  [3,] 0.2236068  0.2672612 -0.25 -0.2236068  0.2672612  0.25
#>  [4,] 0.2236068 -0.2672612 -0.25  0.2236068  0.2672612 -0.25
#>  [5,] 0.2236068  0.2672612  0.25  0.2236068 -0.2672612 -0.25
#>  [6,] 0.2236068 -0.2672612  0.25 -0.2236068 -0.2672612  0.25
#>  [7,] 0.2236068  0.2672612 -0.25 -0.2236068 -0.2672612 -0.25
#>  [8,] 0.2236068 -0.2672612 -0.25  0.2236068 -0.2672612  0.25
#>  [9,] 0.2236068 -0.1336306 -0.25  0.2236068 -0.1336306 -0.25
#> [10,] 0.2236068 -0.1336306  0.25  0.2236068 -0.1336306  0.25
#> [11,] 0.2236068  0.2672612  0.00  0.2236068  0.2672612  0.00
#> [12,] 0.2236068 -0.1336306 -0.25 -0.2236068  0.1336306  0.25
#> [13,] 0.2236068 -0.1336306  0.25 -0.2236068  0.1336306 -0.25
#> [14,] 0.2236068  0.2672612  0.00 -0.2236068 -0.2672612  0.00
#> [15,] 0.2236068 -0.1336306 -0.25  0.2236068 -0.1336306 -0.25
#> [16,] 0.2236068 -0.1336306  0.25  0.2236068 -0.1336306  0.25
#> [17,] 0.2236068  0.2672612  0.00  0.2236068  0.2672612  0.00
#> [18,] 0.2236068 -0.1336306 -0.25 -0.2236068  0.1336306  0.25
#> [19,] 0.2236068 -0.1336306  0.25 -0.2236068  0.1336306 -0.25
#> [20,] 0.2236068  0.2672612  0.00 -0.2236068 -0.2672612  0.00
#> 
#> $crossksi
#>              ksi1          ksi2 ksi3         ksi4          ksi5 ksi6
#> ksi1 1.000000e+00  0.000000e+00    0 1.387779e-17  1.387779e-17    0
#> ksi2 0.000000e+00  1.000000e+00    0 0.000000e+00 -1.387779e-17    0
#> ksi3 0.000000e+00  0.000000e+00    1 0.000000e+00  0.000000e+00    0
#> ksi4 1.387779e-17  0.000000e+00    0 1.000000e+00  0.000000e+00    0
#> ksi5 1.387779e-17 -1.387779e-17    0 0.000000e+00  1.000000e+00    0
#> ksi6 0.000000e+00  0.000000e+00    0 0.000000e+00  0.000000e+00    1
#> 
#> $f
#> [1] -0.0010046574 -0.0230956344  0.0338426276 -0.0007509010 -0.0038862354
#> [6]  0.0004088719
#> 
#> $z
#> [1]  4.8400033  0.6654509 -9.3778431 -6.7918211 -1.4590865  0.3080900
#> 
#> $Y
#>          [,1]      [,2]      [,3]      [,4]
#> [1,] -5.91276 -5.918672 -5.875326 -5.867707
#> 

dimX <- 6
Astar <- 2
simul_data_complete(dimX,Astar)

#> $simX
#>          [,1]     [,2]      [,3]     [,4]     [,5]      [,6]
#> [1,] 3.761495 3.751326 -5.616969 3.757672 3.759877 -5.611648
#> 
#> $HH
#> [1] 3
#> 
#> $eta
#>          eta21      eta22
#> [1,] 0.4082483  0.0000000
#> [2,] 0.8164966  0.4472136
#> [3,] 0.4082483 -0.8944272
#> 
#> $r
#> [1]   1.547228 -10.821040
#> 
#> $epsilon
#> [1]  0.006076077 -0.004092920 -0.001091793  0.002253425  0.004458179
#> [6]  0.004229153
#> 
#> $ksi
#>           ksi1       ksi2
#> [1,] 0.4082483 -0.2886751
#> [2,] 0.4082483 -0.2886751
#> [3,] 0.4082483  0.5773503
#> [4,] 0.4082483 -0.2886751
#> [5,] 0.4082483 -0.2886751
#> [6,] 0.4082483  0.5773503
#> 
#> $crossksi
#>      ksi1 ksi2
#> ksi1    1    0
#> ksi2    0    1
#> 
#> $f
#> [1] -0.20258653 -0.08250133
#> 
#> $z
#> [1]   1.344642 -10.903542
#> 
#> $Y
#>           [,1]      [,2]     [,3]
#> [1,] 0.5414763 -3.775636 10.27239
#> 

dimX <- 6
Astar <- 3
simul_data_complete(dimX,Astar)

#> $simX
#>           [,1]      [,2]     [,3]      [,4]      [,5]     [,6]
#> [1,] -2.932286 -1.254333 8.559674 -2.914986 -1.268121 8.559997
#> 
#> $HH
#> [1] 3
#> 
#> $eta
#>          eta31      eta32      eta33
#> [1,] 0.4082483  0.0000000 -0.9128709
#> [2,] 0.8164966  0.4472136  0.3651484
#> [3,] 0.4082483 -0.8944272  0.1825742
#> 
#> $r
#> [1]  3.574811 12.315434  1.659134
#> 
#> $epsilon
#> [1] -0.006970472  0.011848708 -0.010056217  0.010330178 -0.001938485
#> [6] -0.009733224
#> 
#> $ksi
#>           ksi1       ksi2 ksi3
#> [1,] 0.4082483 -0.2886751 -0.5
#> [2,] 0.4082483 -0.2886751  0.5
#> [3,] 0.4082483  0.5773503  0.0
#> [4,] 0.4082483 -0.2886751 -0.5
#> [5,] 0.4082483 -0.2886751  0.5
#> [6,] 0.4082483  0.5773503  0.0
#> 
#> $crossksi
#>      ksi1 ksi2 ksi3
#> ksi1    1    0    0
#> ksi2    0    1    0
#> ksi3    0    0    1
#> 
#> $f
#> [1]  0.23930957  0.02507851 -0.02656191
#> 
#> $z
#> [1]  3.814121 12.340513  1.632572
#> 
#> $Y
#>           [,1]     [,2]     [,3]
#> [1,] 0.0193473 9.154803 -9.24169
#> 

dimX <- 6
Astar <- 4
simul_data_complete(dimX,Astar)

#> $simX
#>          [,1]     [,2]       [,3]     [,4]     [,5]     [,6]
#> [1,] 4.507553 5.889362 -0.2584758 7.516652 8.896185 2.740097
#> 
#> $HH
#> [1] 4
#> 
#> $eta
#>          eta41      eta42      eta43      eta44
#> [1,] 0.4082483  0.0000000  0.0000000 -0.9128709
#> [2,] 0.8164966  0.4472136  0.1825742  0.3651484
#> [3,] 0.4082483 -0.8944272 -0.3651484  0.1825742
#> [4,] 0.0000000  0.4472136 -0.9128709  0.0000000
#> 
#> $r
#> [1] 11.951553 -6.301035  1.381960 -3.671391
#> 
#> $epsilon
#> [1] -7.810426e-04 -9.320445e-04 -9.330947e-04  1.063999e-02  8.212340e-03
#> [6] -3.836779e-05
#> 
#> $ksi
#>           ksi1       ksi2 ksi3       ksi4
#> [1,] 0.4082483 -0.2886751 -0.5  0.4082483
#> [2,] 0.4082483 -0.2886751  0.5  0.4082483
#> [3,] 0.4082483  0.5773503  0.0  0.4082483
#> [4,] 0.4082483 -0.2886751 -0.5 -0.4082483
#> [5,] 0.4082483 -0.2886751  0.5 -0.4082483
#> [6,] 0.4082483  0.5773503  0.0 -0.4082483
#> 
#> $crossksi
#>               ksi1 ksi2 ksi3          ksi4
#> ksi1  1.000000e+00    0    0 -5.551115e-17
#> ksi2  0.000000e+00    1    0  0.000000e+00
#> ksi3  0.000000e+00    0    1  0.000000e+00
#> ksi4 -5.551115e-17    0    0  1.000000e+00
#> 
#> $f
#> [1] -0.071727439 -0.067722099 -0.006022362 -0.006084087
#> 
#> $z
#> [1] 11.879825 -6.368757  1.375938 -3.677475
#> 
#> $Y
#>          [,1]     [,2]     [,3]      [,4]
#> [1,] 8.166173 5.788549 9.328733 -4.125546
#> 

rm(list=c("dimX","Astar"))