Summary method for plsR models

This function provides a summary method for the class "cv.plsRmodel"

# S3 method for cv.plsRmodel
summary(object, ...)

Arguments

object
an object of the class "cv.plsRmodel"

...
further arguments to be passed to or from methods.

Value

An object of class "summary.cv.plsRglmmodel".

References

Nicolas Meyer, Myriam Maumy-Bertrand et Frédéric Bertrand (2010). Comparing the linear and the logistic PLS regression with qualitative predictors: application to allelotyping data. Journal de la Societe Francaise de Statistique, 151(2), pages 1-18. http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/47

See also

Author

Frédéric Bertrand
frederic.bertrand@math.unistra.fr
https://fbertran.github.io/homepage/

Examples

data(Cornell)
summary(cv.plsR(Y~.,data=Cornell,nt=10,K=6, verbose=FALSE), verbose=FALSE)

#> [[1]]
#>                AIC    Q2cum_Y LimQ2_Y          Q2_Y  PRESS_Y      RSS_Y
#> Nb_Comp_0 82.01205         NA      NA            NA       NA 467.796667
#> Nb_Comp_1 53.15173  0.8976415  0.0975  0.8976414594 47.88298  35.742486
#> Nb_Comp_2 41.08283  0.8976794  0.0975  0.0003703326 35.72925  11.066606
#> Nb_Comp_3 32.06411  0.7384839  0.0975 -1.5558493459 28.28458   4.418081
#> Nb_Comp_4 33.76477 -0.3616302  0.0975 -4.2066778020 23.00353   4.309235
#> Nb_Comp_5 33.34373 -7.1795621  0.0975 -5.0071832437 25.88637   3.521924
#> Nb_Comp_6 35.25533         NA  0.0975            NA       NA   3.496074
#>                R2_Y    AIC.std  DoF.dof sigmahat.dof    AIC.dof    BIC.dof
#> Nb_Comp_0        NA  37.010388 1.000000    6.5212706 46.0708838 47.7893514
#> Nb_Comp_1 0.9235940   8.150064 2.740749    1.8665281  4.5699686  4.9558156
#> Nb_Comp_2 0.9763431  -3.918831 5.085967    1.1825195  2.1075461  2.3949331
#> Nb_Comp_3 0.9905556 -12.937550 5.121086    0.7488308  0.8467795  0.9628191
#> Nb_Comp_4 0.9907882 -11.236891 5.103312    0.7387162  0.8232505  0.9357846
#> Nb_Comp_5 0.9924713 -11.657929 6.006316    0.7096382  0.7976101  0.9198348
#> Nb_Comp_6 0.9925265  -9.746328 7.000002    0.7633343  0.9711322  1.1359501
#>           GMDL.dof DoF.naive sigmahat.naive  AIC.naive  BIC.naive GMDL.naive
#> Nb_Comp_0 27.59461         1      6.5212706 46.0708838 47.7893514   27.59461
#> Nb_Comp_1 21.34020         2      1.8905683  4.1699567  4.4588195   18.37545
#> Nb_Comp_2 27.40202         3      1.1088836  1.5370286  1.6860917   17.71117
#> Nb_Comp_3 24.40842         4      0.7431421  0.7363469  0.8256118   19.01033
#> Nb_Comp_4 24.23105         5      0.7846050  0.8721072  0.9964867   24.16510
#> Nb_Comp_5 28.21184         6      0.7661509  0.8804809  1.0227979   28.64206
#> Nb_Comp_6 33.18348         7      0.8361907  1.1070902  1.3048716   33.63927
#> attr(,"computed_nt")
#> [1] 6
#> 
#> attr(,"class")
#> [1] "summary.cv.plsRmodel"